@@ -1,8 +1,15 @@
+import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter/foundation.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:shared_preferences/shared_preferences.dart';
+import 'package:firebase_messaging/firebase_messaging.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'firebase_messaging_background.dart';
+import 'services/push_notification_service.dart';
+import 'services/hive_cache_service.dart';
 import 'screens/splash_screen.dart';
 import 'screens/home_screen.dart';
 import 'screens/notifications_screen.dart';
@@ -14,28 +21,72 @@ import 'screens/my_requests_screen.dart';
 import 'screens/admin_payments_screen.dart';
 import 'services/admin_auth_service.dart';
 import 'l10n/app_localizations.dart';
-import 'services/permission_service.dart';
 import 'services/notification_service.dart';
 import 'services/app_state_service.dart';
 import 'services/audio_service.dart';
 import 'services/auto_login_service.dart';
 import 'services/tiktok_auth_service.dart';
 import 'services/monthly_requests_tracker.dart';
+import 'services/request_reminder_service.dart';
+import 'services/background_location_service.dart';
+import 'services/location_service.dart';
+import 'package:geolocator/geolocator.dart';
+// Guest trial expiry check moved to Cloud Functions
 import 'package:flutter/services.dart';
 import 'package:app_links/app_links.dart';
+import 'firebase_options.dart';
+import 'widgets/background_icons_widget.dart';
 
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   
-  // הגדרת כיוון אנכי בלבד לכל המסכים
-  await SystemChrome.setPreferredOrientations([
-    DeviceOrientation.portraitUp,
-    DeviceOrientation.portraitDown,
-  ]);
+  // אתחול Firebase
+  await Firebase.initializeApp(
+    options: DefaultFirebaseOptions.currentPlatform,
+  );
+  
+  // Firestore offline persistence is enabled by default on mobile
+  // No need to explicitly enable it - it's automatic
+  debugPrint('✅ Firestore offline persistence is enabled by default');
+  
+  // Initialize Hive cache for offline support
+  if (!kIsWeb) {
+    try {
+      await HiveCacheService.init();
+      debugPrint('✅ Hive cache initialized');
+    } catch (e) {
+      debugPrint('⚠️ Could not initialize Hive cache: $e');
+    }
+  }
+  
+  // רישום background message handler (לא עובד ב-web)
+  if (!kIsWeb) {
+    FirebaseMessaging.onBackgroundMessage(firebaseMessagingBackgroundHandler);
+  }
+  
+  // הגדרת כיוון אנכי בלבד לכל המסכים (לא רלוונטי ב-web)
+  if (!kIsWeb) {
+    await SystemChrome.setPreferredOrientations([
+      DeviceOrientation.portraitUp,
+      DeviceOrientation.portraitDown,
+    ]);
+  }
   
   // הצגת מסך Splash מיד - האתחולים יקרו במסך ה-Splash
-  runApp(const CommunityApp());
+  runApp(
+    const ProviderScope(
+      child: CommunityApp(),
+    ),
+  );
+}
+
+// Enum מותאם אישית לערכות
+enum AppTheme {
+  system,
+  light,
+  dark,
+  gold,
 }
 
 class CommunityApp extends StatefulWidget {
@@ -46,33 +97,226 @@ class CommunityApp extends StatefulWidget {
 }
 
 class _CommunityAppState extends State<CommunityApp> {
-  Locale _selectedLocale = const Locale('he'); // ברירת מחדל עברית
   final ValueNotifier<Locale> _localeNotifier = ValueNotifier(const Locale('he'));
-  ThemeMode _themeMode = ThemeMode.dark; // ברירת מחדל כהה
+  AppTheme _appTheme = AppTheme.dark; // ברירת מחדל כהה
+  bool _localeLoaded = false; // דגל שמציין אם השפה נטענה
 
   @override
   void initState() {
     super.initState();
-    _localeNotifier.value = _selectedLocale;
-    _loadThemeMode();
-    _requestNotificationPermission();
+    _initializeApp();
+  }
+  
+  Future<void> _initializeApp() async {
+    // טעינת השפה והערכת הנושא לפני שב-build רץ
+    await _loadLocale();
+    await _loadThemeMode();
+    // השפה נטענה, אפשר להמשיך
+    if (mounted) {
+      setState(() {
+        _localeLoaded = true;
+      });
+    }
+    // שאר האתחולים (לא חוסמים)
+    // הערה: בקשות הרשאות התראות ומיקום מועברות למסך התחברות
     _setupDeepLinkHandling();
+  }
+  
+  Future<void> _loadLocale() async {
+    try {
+      final prefs = await SharedPreferences.getInstance();
+      final languageCode = prefs.getString('selected_language');
+      
+      // אם אין שפה שמורה, זו הפעם הראשונה - נשתמש בעברית כברירת מחדל
+      // הדיאלוג בחירת שפה יוצג ב-Splash Screen
+      final finalLanguageCode = languageCode ?? 'he';
+      final locale = Locale(finalLanguageCode);
+      debugPrint('🌐 Loading locale from SharedPreferences: $finalLanguageCode');
+      debugPrint('🔍 All SharedPreferences keys: ${prefs.getKeys()}');
+      debugPrint('🔍 Current selected_language value: $languageCode');
+      if (mounted) {
+        _localeNotifier.value = locale;
+        debugPrint('✅ Locale loaded and set to: $finalLanguageCode');
+      }
+    } catch (e) {
+      debugPrint('❌ Error loading locale: $e');
+      // במקרה של שגיאה, נשתמש בברירת המחדל
+      if (mounted) {
+        final defaultLocale = const Locale('he');
+        _localeNotifier.value = defaultLocale;
+      }
+    }
+  }
+  
+  Future<void> _saveLocale(Locale locale) async {
+    try {
+      final prefs = await SharedPreferences.getInstance();
+      final saved = await prefs.setString('selected_language', locale.languageCode);
+      debugPrint('💾 Saved locale to SharedPreferences: ${locale.languageCode} (result: $saved)');
+      // וידוא שהשמירה הצליחה - קריאה נוספת
+      final savedLanguage = prefs.getString('selected_language');
+      debugPrint('✅ Verified saved locale: $savedLanguage');
+      if (savedLanguage != locale.languageCode) {
+        debugPrint('⚠️ WARNING: Saved language ($savedLanguage) does not match requested (${locale.languageCode})');
+      }
+    } catch (e) {
+      debugPrint('❌ Error saving locale: $e');
+    }
   }
   
   Future<void> _loadThemeMode() async {
     final prefs = await SharedPreferences.getInstance();
-    final themeModeIndex = prefs.getInt('theme_mode') ?? 2; // 0 = system, 1 = light, 2 = dark (ברירת מחדל)
+    final themeIndex = prefs.getInt('app_theme') ?? 2; // 0 = system, 1 = light, 2 = dark, 3 = gold (ברירת מחדל: dark)
     setState(() {
-      _themeMode = ThemeMode.values[themeModeIndex];
+      _appTheme = AppTheme.values[themeIndex.clamp(0, AppTheme.values.length - 1)];
     });
   }
   
-  Future<void> _saveThemeMode(ThemeMode themeMode) async {
+  Future<void> _saveThemeMode(AppTheme appTheme) async {
     final prefs = await SharedPreferences.getInstance();
-    await prefs.setInt('theme_mode', themeMode.index);
+    await prefs.setInt('app_theme', appTheme.index);
     setState(() {
-      _themeMode = themeMode;
+      _appTheme = appTheme;
     });
+  }
+  
+  // פונקציה שמחזירה ThemeMode בהתאם ל-AppTheme (למקרה של system)
+  ThemeMode _getThemeMode() {
+    if (_appTheme == AppTheme.system) {
+      return ThemeMode.system;
+    } else if (_appTheme == AppTheme.light) {
+      return ThemeMode.light;
+    } else {
+      // dark או gold - נשתמש ב-darkTheme
+      return ThemeMode.dark;
+    }
+  }
+  
+  // פונקציה שמחזירה ThemeData לערכה GOLD
+  ThemeData _getGoldTheme() {
+    // עיצוב יוקרתי עם צבעי זהב/שחור
+    const colorScheme = ColorScheme.dark(
+      primary: Color(0xFFFFD700), // זהב
+      onPrimary: Color(0xFF000000), // שחור על זהב
+      primaryContainer: Color(0xFFFFA000), // זהב כהה (רקע לכפתורים)
+      onPrimaryContainer: Color(0xFF000000), // טקסט על רקע זהב כהה
+      secondary: Color(0xFFFFC107), // זהב בהיר
+      onSecondary: Color(0xFF000000), // שחור על זהב
+      secondaryContainer: Color(0xFFFFA000), // זהב כהה
+      onSecondaryContainer: Color(0xFF000000), // טקסט על רקע זהב כהה
+      tertiary: Color(0xFFFFC107), // זהב בהיר (לשימוש כללי)
+      onTertiary: Color(0xFF000000), // שחור על זהב
+      tertiaryContainer: Color(0xFF2A1A00), // רקע למידע חיובי (זהב כהה עם שקיפות)
+      onTertiaryContainer: Color(0xFFFFC107), // טקסט על רקע מידע חיובי
+      error: Color(0xFFE57373), // אדום רך
+      onError: Colors.white,
+      errorContainer: Color(0xFF3A1A1A), // רקע לשגיאות (אדום כהה עם שקיפות)
+      onErrorContainer: Colors.white, // טקסט על רקע שגיאות
+      surface: Color(0xFF000000), // שחור
+      onSurface: Color(0xFFFFD700), // זהב על שחור
+      onSurfaceVariant: Color(0xFFFFC107), // זהב בהיר
+      surfaceContainer: Color(0xFF1A1A1A), // שחור בהיר
+      surfaceContainerHigh: Color(0xFF2A2A2A), // אפור כהה
+      surfaceContainerHighest: Color(0xFF3A3A3A), // אפור בינוני
+      outline: Color(0xFFFFD700), // גבול זהב
+      outlineVariant: Color(0xFFFFC107), // גבול זהב בהיר
+      shadow: Color(0xFF000000), // צל
+      scrim: Color(0xFF000000), // רקע מעל
+    );
+    
+    return ThemeData(
+      useMaterial3: true,
+      fontFamily: 'Arial',
+      brightness: Brightness.dark,
+      colorScheme: colorScheme,
+      // עיצוב משופר לכפתורים
+      elevatedButtonTheme: ElevatedButtonThemeData(
+        style: ElevatedButton.styleFrom(
+          backgroundColor: const Color(0xFFFFD700), // זהב
+          foregroundColor: const Color(0xFF000000), // שחור
+          elevation: 4,
+          shadowColor: const Color(0xFFFFD700).withValues(alpha: 0.5),
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(12),
+          ),
+          padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 12),
+        ),
+      ),
+      // עיצוב לכפתורי טקסט
+      textButtonTheme: TextButtonThemeData(
+        style: TextButton.styleFrom(
+          foregroundColor: const Color(0xFFFFD700),
+          textStyle: const TextStyle(
+            fontWeight: FontWeight.w600,
+            fontSize: 16,
+          ),
+        ),
+      ),
+      // עיצוב לכפתורי אייקון
+      iconButtonTheme: IconButtonThemeData(
+        style: IconButton.styleFrom(
+          foregroundColor: const Color(0xFFFFD700),
+          backgroundColor: Colors.transparent,
+        ),
+      ),
+      // עיצוב לכפתורי פעולה צפים
+      floatingActionButtonTheme: const FloatingActionButtonThemeData(
+        backgroundColor: Color(0xFFFFD700),
+        foregroundColor: Color(0xFF000000),
+        elevation: 6,
+      ),
+      // עיצוב משופר לשדות טקסט
+      inputDecorationTheme: InputDecorationTheme(
+        filled: true,
+        fillColor: const Color(0xFF1A1A1A),
+        border: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(12),
+          borderSide: const BorderSide(color: Color(0xFFFFD700)),
+        ),
+        enabledBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(12),
+          borderSide: const BorderSide(color: Color(0xFFFFD700)),
+        ),
+        focusedBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(12),
+          borderSide: const BorderSide(color: Color(0xFFFFC107), width: 2),
+        ),
+        errorBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(12),
+          borderSide: const BorderSide(color: Color(0xFFE57373)),
+        ),
+        contentPadding: const EdgeInsets.symmetric(horizontal: 16, vertical: 16),
+      ),
+      // עיצוב משופר לכרטיסים
+      cardTheme: CardThemeData(
+        elevation: 2,
+        shadowColor: const Color(0xFFFFD700).withValues(alpha: 0.3),
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(16),
+        ),
+        color: const Color(0xFF1A1A1A),
+      ),
+      // עיצוב משופר ל-AppBar
+      appBarTheme: AppBarTheme(
+        backgroundColor: colorScheme.surface, // שימוש בצבע מהערכה
+        foregroundColor: colorScheme.onSurface, // שימוש בצבע מהערכה
+        elevation: 0,
+        centerTitle: true,
+        titleTextStyle: TextStyle(
+          fontSize: 20,
+          fontWeight: FontWeight.w600,
+          color: colorScheme.onSurface,
+        ),
+      ),
+      // עיצוב משופר ל-BottomNavigationBar
+      bottomNavigationBarTheme: const BottomNavigationBarThemeData(
+        backgroundColor: Color(0xFF000000),
+        selectedItemColor: Color(0xFFFFD700), // זהב
+        unselectedItemColor: Color(0xFF666666), // אפור
+        elevation: 8,
+        type: BottomNavigationBarType.fixed,
+      ),
+    );
   }
   
   
@@ -113,18 +357,43 @@ class _CommunityAppState extends State<CommunityApp> {
           debugPrint('TikTok authentication failed');
         }
       });
+    } else if (uri.scheme == 'com.example.flutter1' && uri.host == 'email-verified') {
+      // טיפול ב-email verification callback
+      debugPrint('Email verification callback received');
+      // האימייל כבר מאומת - המשתמש יכול להתחבר
+    } else if (uri.path.contains('/payment/success') || 
+               uri.host == 'payment' ||
+               (uri.scheme == 'https' && uri.host.contains('nearme-970f3') && uri.path.contains('/payment/success')) ||
+               (uri.scheme == 'shchunati' && uri.host == 'payment')) {
+      // טיפול ב-payment success callback מ-PayMe
+      debugPrint('✅ Payment success callback received from PayMe: $uri');
+      _navigateToProfileAfterPayment();
     }
   }
   
-  
-  Future<void> _requestNotificationPermission() async {
-    // המתן קצת לפני בקשת ההרשאה
-    await Future.delayed(const Duration(seconds: 2));
+  /// ניווט למסך פרופיל אחרי תשלום מוצלח
+  void _navigateToProfileAfterPayment() {
+    debugPrint('✅ Payment success - setting flag to open profile');
     
+    // בדוק אם המשתמש מחובר
+    final user = FirebaseAuth.instance.currentUser;
+    if (user != null) {
+      // שמור סמן שצריך לפתוח פרופיל אחרי תשלום
+      AppStateService.setShouldOpenProfileAfterPayment(true);
+      
+      // ניווט למסך הראשי (אם לא כבר שם)
     if (mounted) {
-      await PermissionService.requestNotificationPermission(context);
+        Navigator.of(context).pushNamedAndRemoveUntil(
+          '/main',
+          (route) => false,
+        );
+    }
+    } else {
+      debugPrint('⚠️ User not logged in, cannot navigate to profile');
     }
   }
+  
+  
 
   @override
   void dispose() {
@@ -132,10 +401,19 @@ class _CommunityAppState extends State<CommunityApp> {
     super.dispose();
   }
 
-  void _changeLocale(Locale locale) {
-    _localeNotifier.value = locale;
-    setState(() {
-      _selectedLocale = locale;
+  void _changeLocale(Locale locale) async {
+    debugPrint('🔄 Changing locale to: ${locale.languageCode}');
+    // עדכון ה-state לפני השמירה כדי שהמשתמש יראה את השינוי מיד
+    if (mounted) {
+      debugPrint('🔄 Setting _localeNotifier.value to: ${locale.languageCode}');
+      _localeNotifier.value = locale;
+      debugPrint('✅ _localeNotifier.value updated to: ${locale.languageCode}');
+    }
+    // שמירה ב-SharedPreferences - ללא חסימה של UI
+    _saveLocale(locale).then((_) {
+      debugPrint('✅ Locale changed and saved successfully to: ${locale.languageCode}');
+    }).catchError((e) {
+      debugPrint('❌ Error saving locale: $e');
     });
   }
   
@@ -156,21 +434,56 @@ class _CommunityAppState extends State<CommunityApp> {
 
   @override
   Widget build(BuildContext context) {
+    // אם השפה עדיין לא נטענה, הצג loading
+    if (!_localeLoaded) {
+      return MaterialApp(
+        home: Scaffold(
+          body: Center(
+            child: CircularProgressIndicator(),
+          ),
+        ),
+      );
+    }
+    
+    // שימוש ב-ValueListenableBuilder כדי להגיב לשינויים בשפה
+    return ValueListenableBuilder<Locale>(
+      valueListenable: _localeNotifier,
+      builder: (context, locale, child) {
     return Localizations(
-      locale: _selectedLocale,
+          locale: locale,
       delegates: AppLocalizations.localizationsDelegates,
       child: Builder(
         builder: (context) {
-          if (!mounted) return const SizedBox.shrink();
+              final l10n = AppLocalizations.of(context);
           return MaterialApp(
-        title: 'שכונתי',
+                title: l10n.appTitle,
         debugShowCheckedModeBanner: false,
-        locale: _selectedLocale,
+                locale: locale,
         supportedLocales: AppLocalizations.supportedLocales,
         localizationsDelegates: AppLocalizations.localizationsDelegates,
         onGenerateRoute: _generateRoute,
-        themeMode: _themeMode,
-        theme: ThemeData(
+        themeMode: _getThemeMode(),
+        builder: (context, child) {
+          // הגבלת רוחב ל-80% ב-web בלבד עם רקע אייקונים
+          if (kIsWeb) {
+            return Stack(
+              children: [
+                // רקע עם אייקונים מפוזרים
+                BackgroundIconsWidget(child: const SizedBox.shrink()),
+                // התוכן המרכזי
+                Center(
+                  child: Container(
+                    width: MediaQuery.of(context).size.width * 0.8,
+                    constraints: const BoxConstraints(maxWidth: 1200), // מקסימום 1200px
+                    child: child,
+                  ),
+                ),
+              ],
+            );
+          }
+          return child!;
+        },
+        theme: _appTheme == AppTheme.gold ? _getGoldTheme() : ThemeData(
           useMaterial3: true,
           fontFamily: 'Arial',
           brightness: Brightness.light,
@@ -200,7 +513,7 @@ class _CommunityAppState extends State<CommunityApp> {
               backgroundColor: const Color(0xFF03A9F4), // כחול מהלוגו
               foregroundColor: Colors.white,
               elevation: 4,
-              shadowColor: const Color(0xFF03A9F4).withOpacity(0.3),
+                        shadowColor: const Color(0xFF03A9F4).withValues(alpha: 0.3),
               shape: RoundedRectangleBorder(
                 borderRadius: BorderRadius.circular(12),
               ),
@@ -255,7 +568,7 @@ class _CommunityAppState extends State<CommunityApp> {
           // עיצוב משופר לכרטיסים
           cardTheme: CardThemeData(
             elevation: 2,
-            shadowColor: Colors.black.withOpacity(0.1),
+                      shadowColor: Colors.black.withValues(alpha: 0.1),
             shape: RoundedRectangleBorder(
               borderRadius: BorderRadius.circular(16),
             ),
@@ -282,7 +595,7 @@ class _CommunityAppState extends State<CommunityApp> {
             type: BottomNavigationBarType.fixed,
           ),
         ),
-        darkTheme: ThemeData(
+        darkTheme: _appTheme == AppTheme.gold ? _getGoldTheme() : ThemeData(
           useMaterial3: true,
           fontFamily: 'Arial',
           brightness: Brightness.dark,
@@ -312,7 +625,7 @@ class _CommunityAppState extends State<CommunityApp> {
               backgroundColor: const Color(0xFF03A9F4), // כחול מהלוגו
               foregroundColor: Colors.white,
               elevation: 4,
-              shadowColor: const Color(0xFF03A9F4).withOpacity(0.3),
+                        shadowColor: const Color(0xFF03A9F4).withValues(alpha: 0.3),
               shape: RoundedRectangleBorder(
                 borderRadius: BorderRadius.circular(12),
               ),
@@ -367,7 +680,7 @@ class _CommunityAppState extends State<CommunityApp> {
           // עיצוב משופר לכרטיסים
           cardTheme: CardThemeData(
             elevation: 2,
-            shadowColor: Colors.black.withOpacity(0.3),
+                      shadowColor: Colors.black.withValues(alpha: 0.3),
             shape: RoundedRectangleBorder(
               borderRadius: BorderRadius.circular(16),
             ),
@@ -375,7 +688,7 @@ class _CommunityAppState extends State<CommunityApp> {
           ),
           // עיצוב משופר ל-AppBar
           appBarTheme: const AppBarTheme(
-            backgroundColor: Color(0xFFFF9800), // כתום ענתיק למצב כהה
+            backgroundColor: Color(0xFF9C27B0), // סגול יפה למצב כהה
             foregroundColor: Colors.white,
             elevation: 0,
             centerTitle: true,
@@ -394,13 +707,17 @@ class _CommunityAppState extends State<CommunityApp> {
             type: BottomNavigationBarType.fixed,
           ),
         ),
-        home: const SplashScreen(),
+        home: SplashScreen(
+          onLanguageSelected: (locale) {
+            _changeLocale(locale);
+          },
+        ),
         routes: {
           '/main': (context) => AuthWrapper(
             onLocaleChange: _changeLocale, 
             localeNotifier: _localeNotifier,
             onThemeChange: _saveThemeMode,
-            currentThemeMode: _themeMode,
+            currentThemeMode: _appTheme,
           ),
           '/auth': (context) => YokiStyleAuthScreen(
             onLoginSuccess: () {
@@ -411,6 +728,8 @@ class _CommunityAppState extends State<CommunityApp> {
       );
         },
       ),
+        );
+      },
     );
   }
 }
@@ -418,8 +737,8 @@ class _CommunityAppState extends State<CommunityApp> {
 class AuthWrapper extends StatefulWidget {
   final void Function(Locale) onLocaleChange;
   final ValueNotifier<Locale> localeNotifier;
-  final void Function(ThemeMode) onThemeChange;
-  final ThemeMode currentThemeMode;
+  final void Function(AppTheme) onThemeChange;
+  final AppTheme currentThemeMode;
   const AuthWrapper({
     super.key, 
     required this.onLocaleChange, 
@@ -433,65 +752,40 @@ class AuthWrapper extends StatefulWidget {
 }
 
 class _AuthWrapperState extends State<AuthWrapper> {
+  Future<User?>? _redirectResultFuture;
+  
   @override
   void initState() {
     super.initState();
-    _handleGoogleSignInRedirect();
-  }
-
-  /// טיפול ב-Google Sign-In redirect
-  Future<void> _handleGoogleSignInRedirect() async {
-    if (kIsWeb) {
-      try {
-        // בדיקה אם יש redirect result מ-Google Sign-In
-        final result = await FirebaseAuth.instance.getRedirectResult();
-        if (result.user != null) {
-          debugPrint('✅ Google Sign-In redirect successful: ${result.user!.email}');
-          // הצג הודעה על התחברות מוצלחת
-          if (mounted) {
-            ScaffoldMessenger.of(context).showSnackBar(
-              const SnackBar(
-                content: Text('התחברות מוצלחת דרך גוגל!'),
-                backgroundColor: Colors.green,
-                duration: Duration(seconds: 3),
-              ),
-            );
-          }
-        } else {
-          debugPrint('ℹ️ No Google Sign-In redirect result');
-        }
-      } catch (e) {
-        debugPrint('❌ Google Sign-In redirect error: $e');
-        // התעלם משגיאות minified - זה לא קריטי
-        if (e.toString().contains('minified')) {
-          debugPrint('⚠️ Ignoring minified error - this is a known issue with getRedirectResult');
-        }
-      }
-    }
+    // טעינת redirect result פעם אחת ב-initState
+    _redirectResultFuture = _getRedirectResult();
   }
 
   /// קבלת redirect result
   Future<User?> _getRedirectResult() async {
     if (kIsWeb) {
       try {
-        final result = await FirebaseAuth.instance.getRedirectResult();
+        // הוספת timeout כדי למנוע חסימה ארוכה
+        final result = await FirebaseAuth.instance
+            .getRedirectResult()
+            .timeout(
+              const Duration(seconds: 2),
+              onTimeout: () {
+                debugPrint('⏱️ getRedirectResult timeout - returning null');
+                throw TimeoutException('getRedirectResult timeout');
+              },
+            );
         if (result.user != null) {
           debugPrint('✅ Google Sign-In redirect successful: ${result.user!.email}');
-          // הצג הודעה על התחברות מוצלחת
-          if (mounted) {
-            ScaffoldMessenger.of(context).showSnackBar(
-              const SnackBar(
-                content: Text('התחברות מוצלחת דרך גוגל!'),
-                backgroundColor: Colors.green,
-                duration: Duration(seconds: 3),
-              ),
-            );
-          }
+          // הודעה תוצג ב-build method דרך ScaffoldMessenger
           return result.user;
         } else {
           debugPrint('ℹ️ No Google Sign-In redirect result');
           return null;
         }
+      } on TimeoutException {
+        debugPrint('⏱️ getRedirectResult timeout');
+        return null;
       } catch (e) {
         debugPrint('❌ Google Sign-In redirect error: $e');
         // התעלם משגיאות minified - זה לא קריטי
@@ -508,14 +802,27 @@ class _AuthWrapperState extends State<AuthWrapper> {
   Widget build(BuildContext context) {
     return Builder(
       builder: (context) {
+        // בדיקה שה-widget עדיין פעיל
+        if (!mounted || !context.mounted) {
+          return const SizedBox.shrink();
+        }
         return FutureBuilder<User?>(
-          future: _getRedirectResult(),
+          future: _redirectResultFuture,
           builder: (context, redirectSnapshot) {
+            // בדיקה שה-widget עדיין פעיל
+            if (!mounted || !context.mounted) {
+              return const SizedBox.shrink();
+            }
             return StreamBuilder<User?>(
               stream: FirebaseAuth.instance.authStateChanges(),
               builder: (context, snapshot) {
                 // בדיקה שה-widget עדיין פעיל
-                if (!mounted) {
+                if (!mounted || !context.mounted) {
+                  return const SizedBox.shrink();
+                }
+                
+            // בדיקה נוספת שה-widget עדיין פעיל לפני כל פעולה
+            if (!mounted || !context.mounted) {
                   return const SizedBox.shrink();
                 }
                 
@@ -524,11 +831,14 @@ class _AuthWrapperState extends State<AuthWrapper> {
             debugPrint('   - connectionState: ${snapshot.connectionState}');
             debugPrint('   - hasData: ${snapshot.hasData}');
             debugPrint('   - user: ${snapshot.data?.uid}');
+            debugPrint('   - user email: ${snapshot.data?.email}');
             debugPrint('   - redirectUser: ${redirectSnapshot.data?.uid}');
+            debugPrint('   - redirectUser email: ${redirectSnapshot.data?.email}');
             
             // אם יש redirect result, השתמש בו
             if (redirectSnapshot.hasData && redirectSnapshot.data != null) {
               debugPrint('✅ AuthWrapper - Redirect user found, showing MainApp');
+              if (!mounted || !context.mounted) return const SizedBox.shrink();
               return MainApp(
                 onLocaleChange: widget.onLocaleChange, 
                 localeNotifier: widget.localeNotifier,
@@ -539,6 +849,7 @@ class _AuthWrapperState extends State<AuthWrapper> {
             
             if (snapshot.hasError) {
               debugPrint('❌ AuthWrapper - Error in stream, showing YokiStyleAuthScreen');
+              if (!mounted || !context.mounted) return const SizedBox.shrink();
               return YokiStyleAuthScreen(
                 onLoginSuccess: () {
                   // אחרי התחברות מוצלחת - המעבר יתבצע אוטומטית דרך StreamBuilder
@@ -547,30 +858,40 @@ class _AuthWrapperState extends State<AuthWrapper> {
               );
             }
 
-            if (snapshot.connectionState == ConnectionState.waiting) {
-              debugPrint('⏳ AuthWrapper - Waiting for auth state, showing loading');
-              return const Scaffold(
-                body: Center(child: CircularProgressIndicator()),
-              );
-            }
-
+            // אם יש user גם ב-waiting state, נציג את MainApp (למקרה שה-stream לא מתעדכן)
             if (snapshot.hasData && snapshot.data != null) {
-              debugPrint('✅ AuthWrapper - User authenticated, showing MainApp');
+              debugPrint('✅ AuthWrapper - User authenticated (even in waiting state), showing MainApp');
+              if (!mounted || !context.mounted) return const SizedBox.shrink();
               return MainApp(
                 onLocaleChange: widget.onLocaleChange, 
                 localeNotifier: widget.localeNotifier,
                 onThemeChange: widget.onThemeChange,
                 currentThemeMode: widget.currentThemeMode,
               );
-            } else {
-              debugPrint('🔐 AuthWrapper - No user, showing YokiStyleAuthScreen');
-              return YokiStyleAuthScreen(
-                onLoginSuccess: () {
-                  // אחרי התחברות מוצלחת - המעבר יתבצע אוטומטית דרך StreamBuilder
-                  debugPrint('✅ AuthWrapper - Login success callback called');
-                },
+            }
+
+            if (snapshot.connectionState == ConnectionState.waiting) {
+              debugPrint('⏳ AuthWrapper - Waiting for auth state, showing loading');
+              if (!mounted || !context.mounted) return const SizedBox.shrink();
+              // הצג loading - אם זה לוקח יותר מדי זמן, ה-StreamBuilder יתעדכן אוטומטית
+              return const Scaffold(
+                body: Center(child: CircularProgressIndicator()),
               );
             }
+
+            // אם אין user, נציג מסך התחברות
+            debugPrint('🔐 AuthWrapper - No user, showing YokiStyleAuthScreen');
+            if (!mounted || !context.mounted) return const SizedBox.shrink();
+            return YokiStyleAuthScreen(
+              onLoginSuccess: () {
+                // אחרי התחברות מוצלחת - המעבר יתבצע אוטומטית דרך StreamBuilder
+                debugPrint('✅ AuthWrapper - Login success callback called');
+                // כפיית עדכון של ה-widget
+                if (mounted) {
+                  setState(() {});
+                }
+              },
+            );
           },
         );
           },
@@ -583,8 +904,8 @@ class _AuthWrapperState extends State<AuthWrapper> {
 class MainApp extends StatefulWidget {
   final void Function(Locale) onLocaleChange;
   final ValueNotifier<Locale> localeNotifier;
-  final void Function(ThemeMode) onThemeChange;
-  final ThemeMode currentThemeMode;
+  final void Function(AppTheme) onThemeChange;
+  final AppTheme currentThemeMode;
   const MainApp({
     super.key, 
     required this.onLocaleChange, 
@@ -600,11 +921,20 @@ class MainApp extends StatefulWidget {
 class _MainAppState extends State<MainApp> with WidgetsBindingObserver, AudioMixin {
   int _selectedIndex = 0;
   int _pendingPaymentsCount = 0;
+  Timer? _reminderTimer;
+  Timer? _locationServiceCheckTimer;
+  
+  _MainAppState() {
+    debugPrint('🚀 MainApp constructor called');
+  }
 
   @override
   void initState() {
     super.initState();
+    debugPrint('🚀 MainApp - initState called');
     WidgetsBinding.instance.addObserver(this);
+    // אתחול שירות ההתראות
+    _initializePushNotifications();
     // וידוא שהמנהל מוגדר כעסקי
     _ensureAdminProfileIfNeeded();
     // האזנה להתראות
@@ -613,11 +943,183 @@ class _MainAppState extends State<MainApp> with WidgetsBindingObserver, AudioMix
     if (AdminAuthService.isCurrentUserAdmin()) {
       _listenToPendingPayments();
     }
+    
+    // בדיקה אם צריך לפתוח פרופיל אחרי תשלום
+    _checkAndOpenProfileAfterPayment();
+    
+    // אתחול Timer לבדיקת תזכורות כל דקה
+    _startReminderTimer();
+    // הפעלת שירות עדכון מיקום ברקע (לא עובד ב-web)
+    if (!kIsWeb) {
+      BackgroundLocationService.start();
+    }
+    // ✅ בדיקת שירות המיקום כאשר האפליקציה נפתחת - מיידית
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      if (mounted) {
+        // ✅ בדיקה מיידית (ללא delay) אם context זמין
+        if (context.mounted) {
+          _checkLocationService(providedContext: context);
+        } else {
+          // אם context לא זמין, נמתין קצת ונבדוק שוב
+          Future.delayed(const Duration(milliseconds: 300), () {
+            if (mounted && context.mounted) {
+              _checkLocationService(providedContext: context);
+            }
+          });
+        }
+      }
+    });
+    
+    // ✅ בדיקה תקופתית של שירות המיקום כל 30 שניות כאשר האפליקציה פתוחה (להצגת דיאלוג אם נדרש)
+    // הערה: הדיאלוג יוצג רק פעם אחת בכניסה לאפליקציה (forceShow: true), ולאחר מכן רק אם עברה שעה (forceShow: false)
+    _locationServiceCheckTimer = Timer.periodic(const Duration(seconds: 30), (_) {
+      if (mounted && context.mounted) {
+        _checkLocationService(providedContext: context, forceShow: false);
+      }
+    });
   }
+
+  /// אתחול שירות ההתראות
+  Future<void> _initializePushNotifications() async {
+    try {
+      await PushNotificationService.initialize();
+      debugPrint('✅ Push notification service initialized in MainApp');
+    } catch (e) {
+      debugPrint('❌ Error initializing push notification service: $e');
+    }
+  }
+
+  /// בדיקה אם צריך לפתוח פרופיל אחרי תשלום מוצלח
+  void _checkAndOpenProfileAfterPayment() {
+    // המתן קצת כדי שהמסך ייטען
+    Future.delayed(const Duration(milliseconds: 1500), () {
+      if (mounted && AppStateService.shouldOpenProfileAfterPayment()) {
+        debugPrint('✅ Opening profile after payment success');
+        
+        // עדכן את ה-index למסך פרופיל (index 3)
+        setState(() {
+          _selectedIndex = 3;
+        });
+        
+        // איפוס הסמן
+        AppStateService.clearShouldOpenProfileAfterPayment();
+        
+        // הצג הודעה שהמנוי התעדכן
+        Future.delayed(const Duration(milliseconds: 500), () {
+          if (mounted) {
+            ScaffoldMessenger.of(context).showSnackBar(
+              const SnackBar(
+                content: Text('🎉 תשלום אושר! המנוי שלך הופעל בהצלחה'),
+                backgroundColor: Colors.green,
+                duration: Duration(seconds: 5),
+              ),
+            );
+          }
+        });
+      }
+    });
+  }
+
+  /// אתחול Timer לבדיקת תזכורות כל דקה
+  void _startReminderTimer() {
+    debugPrint('🚀 MainApp: Starting reminder timer (every 1 minute)');
+    _reminderTimer = Timer.periodic(const Duration(minutes: 1), (timer) {
+      debugPrint('🚀 MainApp: Timer triggered - checking reminders');
+      _checkRequestReminders();
+      // Guest trial expiry check moved to Cloud Functions
+    });
+    
+    // בדיקה ראשונית
+    _checkRequestReminders();
+    // Guest trial expiry check moved to Cloud Functions
+  }
+
+  /// בדיקת בקשות עם עוזרים במשך יותר משבוע
+  Future<void> _checkRequestReminders() async {
+    try {
+      debugPrint('🚀 MainApp: Starting request reminders check...');
+      await RequestReminderService.checkAndSendReminderNotifications();
+      debugPrint('🚀 MainApp: Request reminders check completed');
+    } catch (e) {
+      debugPrint('❌ Error checking request reminders: $e');
+    }
+  }
+
+  /// ✅ בדיקת שירות המיקום והצגת דיאלוג אם מבוטל
+  /// forceShow: true = הצג תמיד (בכניסה לאפליקציה), false = הצג רק אם לא הוצג לאחרונה
+  Future<void> _checkLocationService({BuildContext? providedContext, bool forceShow = false}) async {
+    try {
+      // בדיקה ראשונית אם שירות המיקום פעיל (לפני המתנה ל-context)
+      bool serviceEnabled = await Geolocator.isLocationServiceEnabled();
+      if (serviceEnabled) {
+        debugPrint('📍 Location service is enabled, no need to show dialog');
+        return;
+      }
+      
+      debugPrint('📍 Location service is disabled, attempting to show dialog (forceShow: $forceShow)');
+      
+      // ✅ המתנה מינימלית (100ms) כדי לוודא שה-context מוכן - מיידי ככל האפשר
+      await Future.delayed(const Duration(milliseconds: 100));
+      
+      // ניסיון לקבל context - ננסה מספר פעמים
+      BuildContext? context = providedContext;
+      
+      if (context == null || !context.mounted) {
+        // ננסה לקבל context מה-State ישירות
+        try {
+          if (mounted) {
+            context = this.context;
+          }
+        } catch (e) {
+          debugPrint('❌ Error getting context from State: $e');
+        }
+      }
+      
+      if (context == null || !context.mounted) {
+        // ננסה לקבל context מ-AppStateService - ננסה פחות פעמים עם delay קצר יותר
+        for (int i = 0; i < 5; i++) {
+          context = AppStateService.currentContext;
+          if (context != null && context.mounted) {
+            break;
+          }
+          if (i < 4) {
+            await Future.delayed(const Duration(milliseconds: 100));
+          }
+        }
+      }
+      
+      if (context == null || !context.mounted) {
+        debugPrint('📍 No active context available for location service check after retries');
+        // אם אין context, נציג התראה במקום דיאלוג
+        await LocationService.checkAndShowLocationServiceNotification();
+        return;
+      }
+      
+      debugPrint('✅ Context available, showing location service dialog');
+      
+      // בדיקה והצגת דיאלוג אם שירות המיקום מבוטל
+      // forceShow = true בכניסה לאפליקציה כדי להציג את הדיאלוג פעם אחת, false בבדיקות תקופתיות
+      await LocationService.checkAndShowLocationServiceDialog(context, forceShow: forceShow);
+    } catch (e) {
+      debugPrint('❌ Error checking location service: $e');
+      // אם יש שגיאה, נציג התראה במקום דיאלוג
+      try {
+        await LocationService.checkAndShowLocationServiceNotification();
+      } catch (e2) {
+        debugPrint('❌ Error showing location service notification: $e2');
+      }
+    }
+  }
+
+  // Guest trial expiry check moved to Cloud Functions
 
   @override
   void dispose() {
     WidgetsBinding.instance.removeObserver(this);
+    _reminderTimer?.cancel();
+    _locationServiceCheckTimer?.cancel();
+    // עצירת שירות עדכון מיקום ברקע
+    BackgroundLocationService.stop();
     super.dispose();
   }
 
@@ -630,9 +1132,22 @@ class _MainAppState extends State<MainApp> with WidgetsBindingObserver, AudioMix
       case AppLifecycleState.detached:
         // כשהאפליקציה עוברת לרקע או נסגרת - ניקוי מצב
         AppStateService.clearUserState();
+        // ✅ בדיקת שירות המיקום כאשר האפליקציה עוברת לרקע
+        BackgroundLocationService.checkLocationServiceWhenBackground();
         break;
       case AppLifecycleState.resumed:
-        // כשהאפליקציה חוזרת לקדמה - אין צורך לעשות כלום
+        // כשהאפליקציה חוזרת לקדמה - בדוק אם שירות המיקום פעיל - מיידית
+        // ✅ בדיקה מיידית (ללא delay) אם context זמין - forceShow: true כדי להציג פעם אחת
+        if (mounted && context.mounted) {
+          _checkLocationService(providedContext: context, forceShow: true);
+        } else {
+          // אם context לא זמין, נמתין קצת ונבדוק שוב
+          Future.delayed(const Duration(milliseconds: 300), () {
+            if (mounted && context.mounted) {
+              _checkLocationService(providedContext: context, forceShow: true);
+            }
+          });
+        }
         break;
       default:
         break;
@@ -676,9 +1191,9 @@ class _MainAppState extends State<MainApp> with WidgetsBindingObserver, AudioMix
           .doc(notificationId)
           .update({'read': true});
 
-      print('Notification shown and marked as read');
+      debugPrint('Notification shown and marked as read');
     } catch (e) {
-      print('Error showing notification: $e');
+      debugPrint('Error showing notification: $e');
     }
   }
 
@@ -707,8 +1222,11 @@ class _MainAppState extends State<MainApp> with WidgetsBindingObserver, AudioMix
 
   @override
   Widget build(BuildContext context) {
-    // שמירת context לניווט התראות
+    debugPrint('🏗️ MainApp - build method called');
+    // שמירת context לניווט התראות - רק אם ה-widget עדיין פעיל
+    if (mounted) {
     AppStateService.setCurrentContext(context);
+    }
     
     return ValueListenableBuilder<Locale>(
       valueListenable: widget.localeNotifier,
@@ -724,7 +1242,7 @@ class _MainAppState extends State<MainApp> with WidgetsBindingObserver, AudioMix
                 // אייקון מדריך
                 IconButton(
                   icon: const Icon(Icons.school),
-                  tooltip: 'מדריך משתמש',
+                  tooltip: l10n.userGuide,
                   onPressed: () {
                     Navigator.push(
                       context,
@@ -737,89 +1255,123 @@ class _MainAppState extends State<MainApp> with WidgetsBindingObserver, AudioMix
                 Builder(
                   builder: (context) {
                     if (!mounted) return const SizedBox.shrink();
+                    // שמירת l10n ב-closure כדי למנוע בעיות עם deactivated widget
+                    final currentL10n = l10n;
                     return PopupMenuButton<Locale>(
+                      key: ValueKey('locale_menu_${locale.languageCode}'), // Force rebuild on locale change
                       icon: const Icon(Icons.language),
-                      tooltip: l10n.selectLanguage,
+                      tooltip: currentL10n.selectLanguage,
                       onSelected: (locale) {
                         if (mounted) {
                           widget.onLocaleChange(locale);
                         }
                       },
-                      itemBuilder: (context) => [
+                      itemBuilder: (context) {
+                        // Guard: בדיקה אם ה-context עדיין valid
+                        if (!mounted) return [];
+                        // יצירת l10n חדש-context כדי לוודא שהוא עדכני
+                        final builderL10n = AppLocalizations.of(context);
+                        return [
                         PopupMenuItem(
                           value: const Locale('he'),
-                          child: Text(l10n.hebrew),
+                            child: Text(builderL10n.hebrew),
                         ),
                         PopupMenuItem(
                           value: const Locale('ar'),
-                          child: Text(l10n.arabic),
+                            child: Text(builderL10n.arabic),
                         ),
                         PopupMenuItem(
                           value: const Locale('en'),
-                          child: Text(l10n.english),
+                            child: Text(builderL10n.english),
                         ),
-                      ],
+                        ];
+                      },
                     );
                   },
                 ),
                 Builder(
                   builder: (context) {
                     if (!mounted) return const SizedBox.shrink();
-                    return PopupMenuButton<ThemeMode>(
+                    // שמירת l10n ב-closure כדי למנוע בעיות עם deactivated widget
+                    final currentL10n = l10n;
+                    final currentThemeMode = widget.currentThemeMode;
+                    return PopupMenuButton<AppTheme>(
+                      key: ValueKey('theme_menu_${locale.languageCode}'), // Force rebuild on locale change
                       icon: const Icon(Icons.palette),
-                      tooltip: l10n.theme,
-                      onSelected: (themeMode) {
+                      tooltip: currentL10n.theme,
+                      onSelected: (appTheme) {
                         if (mounted) {
-                          widget.onThemeChange(themeMode);
+                          widget.onThemeChange(appTheme);
                         }
                       },
-                      itemBuilder: (context) => [
+                      itemBuilder: (context) {
+                        // Guard: בדיקה אם ה-context עדיין valid
+                        if (!mounted) return [];
+                        // יצירת l10n חדש-context כדי לוודא שהוא עדכני
+                        final builderL10n = AppLocalizations.of(context);
+                        return [
                         PopupMenuItem(
-                          value: ThemeMode.system,
+                          value: AppTheme.system,
                           child: Row(
                             children: [
                               Icon(
                                 Icons.brightness_auto,
-                                color: widget.currentThemeMode == ThemeMode.system 
+                                  color: currentThemeMode == AppTheme.system 
                                     ? const Color(0xFF03A9F4) 
                                     : null,
                               ),
                               const SizedBox(width: 8),
-                              Text(l10n.systemTheme),
+                                Text(builderL10n.systemTheme),
                             ],
                           ),
                         ),
                         PopupMenuItem(
-                          value: ThemeMode.light,
+                          value: AppTheme.light,
                           child: Row(
                             children: [
                               Icon(
                                 Icons.brightness_high,
-                                color: widget.currentThemeMode == ThemeMode.light 
+                                  color: currentThemeMode == AppTheme.light 
                                     ? const Color(0xFF03A9F4) 
                                     : null,
                               ),
                               const SizedBox(width: 8),
-                              Text(l10n.lightTheme),
+                                Text(builderL10n.lightTheme),
                             ],
                           ),
                         ),
                         PopupMenuItem(
-                          value: ThemeMode.dark,
+                          value: AppTheme.dark,
                           child: Row(
                             children: [
                               Icon(
                                 Icons.brightness_2,
-                                color: widget.currentThemeMode == ThemeMode.dark 
+                                  color: currentThemeMode == AppTheme.dark 
                                     ? const Color(0xFF03A9F4) 
                                     : null,
                               ),
                               const SizedBox(width: 8),
-                              Text(l10n.darkTheme),
+                                Text(builderL10n.darkTheme),
                             ],
                           ),
                         ),
-                      ],
+                        PopupMenuItem(
+                          value: AppTheme.gold,
+                          child: Row(
+                            children: [
+                              Icon(
+                                Icons.star,
+                                  color: currentThemeMode == AppTheme.gold 
+                                    ? const Color(0xFFFFD700) 
+                                    : null,
+                              ),
+                              const SizedBox(width: 8),
+                                Text(builderL10n.goldTheme),
+                            ],
+                          ),
+                        ),
+                        ];
+                      },
                     );
                   },
                 ),
@@ -859,7 +1411,7 @@ class _MainAppState extends State<MainApp> with WidgetsBindingObserver, AudioMix
                 ),
                 BottomNavigationBarItem(
                   icon: const Icon(Icons.assignment),
-                  label: l10n.myRequests,
+                  label: l10n.myRequestsMenu,
                 ),
                 BottomNavigationBarItem(
                   icon: StreamBuilder<int>(
@@ -938,7 +1490,7 @@ class _MainAppState extends State<MainApp> with WidgetsBindingObserver, AudioMix
                           ),
                       ],
                     ),
-                    label: 'ניהול תשלומים',
+                    label: l10n.managePayments,
                   ),
               ],
             ),
@@ -980,30 +1532,38 @@ class _MainAppState extends State<MainApp> with WidgetsBindingObserver, AudioMix
     // הוספת צליל לכפתור התנתקות
     await AudioService().playSound(AudioEvent.buttonClick);
     
+    // Guard context usage after async gap
+    if (!mounted) return;
+    
     // הצגת דיאלוג אישור התנתקות
     final shouldLogout = await showDialog<bool>(
       context: context,
       barrierDismissible: false,
       builder: (BuildContext context) {
+        final dialogL10n = AppLocalizations.of(context);
         return AlertDialog(
-          title: const Text('התנתקות'),
-          content: const Text('האם אתה בטוח שברצונך להתנתק?'),
+          title: Text(dialogL10n.logoutTitle),
+          content: Text(dialogL10n.logoutMessage),
           actions: [
             TextButton(
               onPressed: () async {
                 await AudioService().playSound(AudioEvent.buttonClick);
+                if (context.mounted) {
                 Navigator.of(context).pop(false);
+                }
               },
-              child: const Text('ביטול'),
+              child: Text(dialogL10n.cancel),
             ),
             TextButton(
               onPressed: () async {
                 await AudioService().playSound(AudioEvent.buttonClick);
+                if (context.mounted) {
                 Navigator.of(context).pop(true);
+                }
               },
-              child: const Text(
-                'התנתקות',
-                style: TextStyle(color: Colors.red),
+              child: Text(
+                dialogL10n.logoutButton,
+                style: const TextStyle(color: Colors.red),
               ),
             ),
           ],
@@ -1017,6 +1577,9 @@ class _MainAppState extends State<MainApp> with WidgetsBindingObserver, AudioMix
         // שימוש ב-AutoLoginService להתנתקות מלאה
         await AutoLoginService.logout();
         
+        // Guard context usage after async gap
+        if (!mounted) return;
+        
         // חזרה למסך התחברות
         Navigator.pushNamedAndRemoveUntil(
           context,
@@ -1024,9 +1587,12 @@ class _MainAppState extends State<MainApp> with WidgetsBindingObserver, AudioMix
           (route) => false,
         );
       } catch (e) {
+        // Guard context usage after async gap
+        if (!mounted) return;
+        final l10n = AppLocalizations.of(context);
         ScaffoldMessenger.of(context).showSnackBar(
           SnackBar(
-            content: Text('שגיאה בהתנתקות: $e'),
+            content: Text(l10n.errorLoggingOut(e.toString())),
             backgroundColor: Colors.red,
           ),
         );
@@ -1040,9 +1606,44 @@ class _MainAppState extends State<MainApp> with WidgetsBindingObserver, AudioMix
     
     debugPrint('🔍 _showNewRequestDialog: Starting monthly request limit check');
     
+    // בדיקה אם המשתמש הוא אורח זמני
+    final user = FirebaseAuth.instance.currentUser;
+    if (user != null) {
+      try {
+        final userDoc = await FirebaseFirestore.instance
+            .collection('users')
+            .doc(user.uid)
+            .get();
+        
+        if (userDoc.exists) {
+          final userData = userDoc.data()!;
+          final isTemporaryGuest = userData['isTemporaryGuest'] ?? false;
+          
+          if (isTemporaryGuest) {
+            debugPrint('🔍 _showNewRequestDialog: Temporary guest detected, blocking request creation');
+            if (!mounted) return;
+            final l10n = AppLocalizations.of(context);
+            ScaffoldMessenger.of(context).showSnackBar(
+              SnackBar(
+                content: Text(l10n.pleaseRegisterFirst),
+                backgroundColor: Colors.orange,
+                duration: const Duration(seconds: 3),
+              ),
+            );
+            return;
+          }
+        }
+      } catch (e) {
+        debugPrint('🔍 _showNewRequestDialog: Error checking temporary guest status: $e');
+      }
+    }
+    
     // בדיקת מגבלת בקשות חודשיות
     final canCreateRequest = await _checkMonthlyRequestLimit();
     debugPrint('🔍 _showNewRequestDialog: canCreateRequest = $canCreateRequest');
+    
+    // Guard context usage after async gap
+    if (!mounted) return;
     
     if (!canCreateRequest) {
       debugPrint('🔍 _showNewRequestDialog: Cannot create request, showing limit dialog');
@@ -1124,19 +1725,22 @@ class _MainAppState extends State<MainApp> with WidgetsBindingObserver, AudioMix
 
   /// הצגת דיאלוג מגבלת בקשות חודשיות
   Future<void> _showMonthlyLimitDialog(String nextMonthDate, int maxRequests) async {
+    if (!mounted) return;
+    
     return showDialog(
       context: context,
       barrierDismissible: false,
       builder: (BuildContext context) {
+        final dialogL10n = AppLocalizations.of(context);
         return AlertDialog(
           title: Row(
             children: [
               Icon(Icons.warning, color: Colors.orange[700], size: 28),
               const SizedBox(width: 8),
-              const Expanded(
+              Expanded(
                 child: Text(
-                  'הגעת למגבלת הבקשות החודשיות',
-                  style: TextStyle(fontSize: 16),
+                  dialogL10n.monthlyLimitReached,
+                  style: const TextStyle(fontSize: 16),
                 ),
               ),
             ],
@@ -1147,12 +1751,12 @@ class _MainAppState extends State<MainApp> with WidgetsBindingObserver, AudioMix
               crossAxisAlignment: CrossAxisAlignment.start,
               children: [
                 Text(
-                  'הגעת למגבלת הבקשות החודשיות שלך ($maxRequests בקשות).',
+                  dialogL10n.monthlyLimitMessage(maxRequests),
                   style: const TextStyle(fontSize: 16, fontWeight: FontWeight.w500),
                 ),
                 const SizedBox(height: 12),
                 Text(
-                  'באפשרותך:',
+                  dialogL10n.youCan,
                   style: const TextStyle(fontSize: 14, fontWeight: FontWeight.w600),
                 ),
                 const SizedBox(height: 8),
@@ -1162,7 +1766,7 @@ class _MainAppState extends State<MainApp> with WidgetsBindingObserver, AudioMix
                     const SizedBox(width: 8),
                     Expanded(
                       child: Text(
-                        'לחכות לחודש הבא שמתחיל ב-$nextMonthDate',
+                        dialogL10n.waitForNextMonth(nextMonthDate),
                         style: const TextStyle(fontSize: 14),
                       ),
                     ),
@@ -1175,7 +1779,7 @@ class _MainAppState extends State<MainApp> with WidgetsBindingObserver, AudioMix
                     const SizedBox(width: 8),
                     Expanded(
                       child: Text(
-                        'לשדרג מנוי לקבלת יותר בקשות חודשיות',
+                        dialogL10n.upgradeSubscription,
                         style: const TextStyle(fontSize: 14),
                       ),
                     ),
@@ -1188,14 +1792,18 @@ class _MainAppState extends State<MainApp> with WidgetsBindingObserver, AudioMix
             TextButton(
               onPressed: () async {
                 await AudioService().playSound(AudioEvent.buttonClick);
+                if (context.mounted) {
                 Navigator.of(context).pop();
+                }
               },
-              child: const Text('הבנתי'),
+              child: Text(dialogL10n.understood),
             ),
             ElevatedButton(
               onPressed: () async {
                 await AudioService().playSound(AudioEvent.buttonClick);
+                if (context.mounted) {
                 Navigator.of(context).pop();
+                }
                 // ניווט ישירות למסך פרופיל
                 if (context.mounted) {
                   Navigator.of(context).push(
@@ -1209,7 +1817,7 @@ class _MainAppState extends State<MainApp> with WidgetsBindingObserver, AudioMix
                 backgroundColor: Colors.green[600],
                 foregroundColor: Colors.white,
               ),
-              child: const Text('שדרג מנוי בפרופיל'),
+              child: Text(dialogL10n.upgradeSubscriptionInProfile),
             ),
           ],
         );
