@@ -95,15 +95,15 @@ class _AnimatedAuthScreenState extends State<AnimatedAuthScreen>
                               shape: BoxShape.circle,
                               gradient: RadialGradient(
                                 colors: [
-                                  Colors.purple.withOpacity(0.4),
-                                  Colors.blue.withOpacity(0.3),
-                                  Colors.pink.withOpacity(0.2),
+                                  Colors.purple.withValues(alpha: 0.4),
+                                  Colors.blue.withValues(alpha: 0.3),
+                                  Colors.pink.withValues(alpha: 0.2),
                                   Colors.transparent,
                                 ],
                               ),
                               boxShadow: [
                                 BoxShadow(
-                                  color: Colors.purple.withOpacity(0.3),
+                                  color: Colors.purple.withValues(alpha: 0.3),
                                   blurRadius: 20,
                                   spreadRadius: 5,
                                 ),
@@ -130,8 +130,8 @@ class _AnimatedAuthScreenState extends State<AnimatedAuthScreen>
                               shape: BoxShape.circle,
                               gradient: RadialGradient(
                                 colors: [
-                                  Colors.cyan.withOpacity(0.3),
-                                  Colors.teal.withOpacity(0.2),
+                                  Colors.cyan.withValues(alpha: 0.3),
+                                  Colors.teal.withValues(alpha: 0.2),
                                   Colors.transparent,
                                 ],
                               ),
@@ -198,16 +198,18 @@ class _AnimatedAuthScreenState extends State<AnimatedAuthScreen>
             borderRadius: BorderRadius.circular(20),
             boxShadow: [
               BoxShadow(
-                color: Colors.black.withOpacity(0.2),
+                color: Colors.black.withValues(alpha: 0.2),
                 blurRadius: 20,
                 offset: const Offset(0, 10),
               ),
             ],
           ),
-          child: const Icon(
-            Icons.favorite,
-            size: 60,
-            color: Color(0xFF6C5CE7),
+          child: Padding(
+            padding: const EdgeInsets.all(20.0),
+            child: Image.asset(
+              'assets/images/logolarge.png',
+              fit: BoxFit.contain,
+            ),
           ),
         ),
         
@@ -299,7 +301,7 @@ class _AnimatedAuthScreenState extends State<AnimatedAuthScreen>
           backgroundColor: color,
           foregroundColor: textColor,
           elevation: 8,
-          shadowColor: Colors.black.withOpacity(0.3),
+          shadowColor: Colors.black.withValues(alpha: 0.3),
           shape: RoundedRectangleBorder(
             borderRadius: BorderRadius.circular(16),
           ),
@@ -314,7 +316,7 @@ class _AnimatedAuthScreenState extends State<AnimatedAuthScreen>
         Expanded(
           child: Container(
             height: 1,
-            color: Colors.white.withOpacity(0.3),
+            color: Colors.white.withValues(alpha: 0.3),
           ),
         ),
         Padding(
@@ -322,7 +324,7 @@ class _AnimatedAuthScreenState extends State<AnimatedAuthScreen>
           child: Text(
             'או',
             style: TextStyle(
-              color: Colors.white.withOpacity(0.7),
+              color: Colors.white.withValues(alpha: 0.7),
               fontSize: 14,
             ),
           ),
@@ -330,7 +332,7 @@ class _AnimatedAuthScreenState extends State<AnimatedAuthScreen>
         Expanded(
           child: Container(
             height: 1,
-            color: Colors.white.withOpacity(0.3),
+            color: Colors.white.withValues(alpha: 0.3),
           ),
         ),
       ],
@@ -350,7 +352,7 @@ class _AnimatedAuthScreenState extends State<AnimatedAuthScreen>
         shape: BoxShape.circle,
         boxShadow: [
           BoxShadow(
-            color: const Color(0xFF6C5CE7).withOpacity(0.4),
+            color: const Color(0xFF6C5CE7).withValues(alpha: 0.4),
             blurRadius: 20,
             spreadRadius: 2,
             offset: const Offset(0, 8),
@@ -380,7 +382,7 @@ class _AnimatedAuthScreenState extends State<AnimatedAuthScreen>
       child: Text(
         'התחבר עם אימייל',
         style: TextStyle(
-          color: Colors.white.withOpacity(0.8),
+          color: Colors.white.withValues(alpha: 0.8),
           fontSize: 16,
           decoration: TextDecoration.underline,
         ),
@@ -400,6 +402,8 @@ class _AnimatedAuthScreenState extends State<AnimatedAuthScreen>
       }
     } catch (e) {
       await playErrorSound();
+      // Guard context usage after async gap
+      if (!mounted) return;
       ScaffoldMessenger.of(context).showSnackBar(
         SnackBar(content: Text('שגיאה בהתחברות: $e')),
       );
@@ -426,6 +430,8 @@ class _AnimatedAuthScreenState extends State<AnimatedAuthScreen>
       }
     } catch (e) {
       await playErrorSound();
+      // Guard context usage after async gap
+      if (!mounted) return;
       ScaffoldMessenger.of(context).showSnackBar(
         SnackBar(content: Text('שגיאה בהתחברות: $e')),
       );
@@ -438,6 +444,8 @@ class _AnimatedAuthScreenState extends State<AnimatedAuthScreen>
     setState(() => _isLoading = true);
     try {
       await playErrorSound();
+      // Guard context usage after async gap
+      if (!mounted) return;
       ScaffoldMessenger.of(context).showSnackBar(
         const SnackBar(
           content: Text('כניסה דרך אינסטגרם תהיה זמינה בקרוב!'),
