@@ -1,15 +1,21 @@
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'dart:async';
 import '../models/chat.dart';
 import '../services/chat_service.dart';
 import '../services/app_state_service.dart';
 import '../l10n/app_localizations.dart';
 import '../services/tutorial_service.dart';
 import '../services/audio_service.dart';
+import '../services/voice_message_service.dart';
 import '../widgets/tutorial_dialog.dart';
+import '../widgets/voice_message_widget.dart';
+import '../widgets/voice_recorder_widget.dart';
+import '../providers/chat_provider.dart';
 
-class ChatScreen extends StatefulWidget {
+class ChatScreen extends ConsumerStatefulWidget {
   final String chatId;
   final String requestTitle;
   
@@ -20,70 +26,54 @@ class ChatScreen extends StatefulWidget {
   });
 
   @override
-  State<ChatScreen> createState() => _ChatScreenState();
+  ConsumerState<ChatScreen> createState() => _ChatScreenState();
 }
 
-class _ChatScreenState extends State<ChatScreen> {
+class _ChatScreenState extends ConsumerState<ChatScreen> {
   final TextEditingController _messageController = TextEditingController();
   final ScrollController _scrollController = ScrollController();
-  late Stream<QuerySnapshot> _messagesStream;
+  final VoiceMessageService _voiceService = VoiceMessageService();
   bool _isChatClosed = false;
+  bool _isDeletedByRequestCreator = false; // האם הצ'אט נמחק על ידי מבקש השירות
+  bool _isDeletedByServiceProvider = false; // האם הצ'אט נמחק על ידי נותן השירות
+  bool _isCurrentUserRequestCreator = false; // האם המשתמש הנוכחי הוא מבקש השירות
+  bool _isRequestCompleted = false; // האם הבקשה במצב "טופל"
+  StreamSubscription<DocumentSnapshot>? _chatStatusSubscription;
+  bool _chatTutorialShown = false; // האם הדיאלוג "תקשורת עם נותן שירות" כבר הוצג
+  bool _isRecordingVoice = false; // האם מקליטים הודעה קולית
+  bool _showVoiceRecorder = false; // האם להציג את ה-voice recorder
 
   @override
   void initState() {
     super.initState();
-    _messagesStream = FirebaseFirestore.instance
-        .collection('chats')
-        .doc(widget.chatId)
-        .collection('messages')
-        .orderBy('sentAt', descending: true)
-        .snapshots();
     
     debugPrint('ChatScreen initialized for chatId: ${widget.chatId}');
     
     // עדכון המצב - המשתמש נמצא בצ'אט (קודם כל!)
     _enterChat();
     
-    // בדיקת סטטוס הצ'אט
+    // בדיקת סטטוס הצ'אט והאזנה לשינויים בזמן אמת
     _checkChatStatus();
+    _listenToChatStatus();
     
     // בדיקת ההודעות הקיימות
-    
     _checkExistingMessages();
     
     // סימון הודעות כנקראות כשנכנסים לצ'אט
     ChatService.markMessagesAsRead(widget.chatId);
     
-    // סימון הודעות כנקראות בזמן אמת
-    _markMessagesAsRead();
+    // סימון הודעות כנקראות בזמן אמת - נעשה ב-build method
+    
+    // הצגת הודעת הדרכה רק כשהמשתמש נכנס למסך הצ'אט (רק פעם אחת)
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      _showChatSpecificTutorial();
+    });
   }
+
   
   // פונקציה להפעלת צליל לחיצה
   Future<void> playButtonSound() async {
     await AudioService().playSound(AudioEvent.buttonClick);
-  }
-  
-  // סימון הודעות כנקראות בזמן אמת
-  void _markMessagesAsRead() {
-    final user = FirebaseAuth.instance.currentUser;
-    if (user == null) return;
-
-    // האזנה לשינויים בהודעות
-    _messagesStream.listen((snapshot) {
-      // בדיקה שהמשתמש עדיין בצ'אט (לא יצא)
-      if (!mounted) return;
-      
-      // רק אם המשתמש באמת נכנס לצ'אט (המסך פעיל)
-      if (ModalRoute.of(context)?.isCurrent == true) {
-        for (var doc in snapshot.docs) {
-          final message = Message.fromFirestore(doc);
-          // אם ההודעה לא נשלחה על ידי המשתמש הנוכחי ולא נקראה על ידו
-          if (message.from != user.uid && !message.readBy.contains(user.uid)) {
-            _markMessageAsRead(message.messageId, user.uid);
-          }
-        }
-      }
-    });
   }
 
   // סימון הודעה ספציפית כנקראה
@@ -108,6 +98,12 @@ class _ChatScreenState extends State<ChatScreen> {
 
   // הודעת הדרכה ספציפית לצ'אט - רק כשצריך
   Future<void> _showChatSpecificTutorial() async {
+    // בדיקה אם הדיאלוג כבר הוצג במהלך הפעלה זו
+    if (_chatTutorialShown) {
+      debugPrint('💬 Chat tutorial already shown in this session, returning');
+      return;
+    }
+    
     // רק אם המשתמש לא ראה את ההדרכה הזו קודם
     final hasSeenTutorial = await TutorialService.hasSeenTutorial('chat_specific_tutorial');
     if (hasSeenTutorial) return;
@@ -132,17 +128,20 @@ class _ChatScreenState extends State<ChatScreen> {
     
     if (!mounted) return;
     
+    // סמן שהדיאלוג הוצג
+    _chatTutorialShown = true;
+    
     showDialog(
       context: context,
       barrierDismissible: true,
       builder: (context) => TutorialDialog(
         tutorialKey: 'chat_specific_tutorial',
-        title: 'תקשורת עם נותן השירות',
-        message: 'כאן תוכל לתקשר עם נותן השירות, לשאול שאלות ולתאם את הפרטים.',
+        title: AppLocalizations.of(context).communicationWithServiceProvider,
+        message: AppLocalizations.of(context).communicationWithServiceProviderMessage,
         features: [
           '💬 שליחת הודעות טקסט',
-          '📞 לחיצה על הטלפון להתקשרות',
-          'ℹ️ מידע על הבקשה והמפרסם',
+          '💬 לדבר על הבקשה לקבל/לתת מידע נוסף',
+          '🤝 לסגור עסקה בין מבקש השירות לנותן השירות',
         ],
       ),
     );
@@ -155,6 +154,9 @@ class _ChatScreenState extends State<ChatScreen> {
 
   Future<void> _checkChatStatus() async {
     try {
+      final user = FirebaseAuth.instance.currentUser;
+      if (user == null) return;
+      
       final chatDoc = await FirebaseFirestore.instance
           .collection('chats')
           .doc(widget.chatId)
@@ -163,6 +165,140 @@ class _ChatScreenState extends State<ChatScreen> {
       if (chatDoc.exists) {
         final chatData = chatDoc.data()!;
         final isClosed = chatData['isClosed'] as bool? ?? false;
+        final deletedBy = List<String>.from(chatData['deletedBy'] ?? []);
+        final requestId = chatData['requestId'] as String?;
+        
+        // בדיקה אם הצ'אט נמחק על ידי מבקש השירות או נותן השירות
+        bool isDeletedByRequestCreator = false;
+        bool isDeletedByServiceProvider = false;
+        if (requestId != null && deletedBy.isNotEmpty) {
+          try {
+            final requestDoc = await FirebaseFirestore.instance
+                .collection('requests')
+                .doc(requestId)
+                .get();
+            
+            if (requestDoc.exists) {
+              final requestData = requestDoc.data()!;
+              final createdBy = requestData['createdBy'] as String?;
+              
+              // אם יוצר הבקשה מחק את הצ'אט, נותן השירות לא יכול לפתוח אותו מחדש
+              if (createdBy != null && deletedBy.contains(createdBy) && createdBy != user.uid) {
+                isDeletedByRequestCreator = true;
+                debugPrint('⚠️ Chat was deleted by request creator $createdBy, cannot reopen');
+              }
+              
+              // אם נותן השירות מחק את הצ'אט, מבקש השירות לא יכול לפתוח אותו מחדש
+              if (createdBy == user.uid) {
+                // המשתמש הנוכחי הוא מבקש השירות
+                isDeletedByServiceProvider = false; // נבדוק מחדש
+                for (final deletedByUid in deletedBy) {
+                  if (deletedByUid != user.uid) {
+                    // מישהו אחר מחק את הצ'אט - זה נותן השירות
+                    isDeletedByServiceProvider = true;
+                    debugPrint('⚠️ Chat was deleted by service provider $deletedByUid, cannot reopen');
+                    break;
+                  }
+                }
+              }
+            }
+          } catch (e) {
+            debugPrint('Error checking request creator: $e');
+          }
+        }
+        
+        // בדיקה אם המשתמש הנוכחי הוא מבקש השירות ואם הבקשה במצב "טופל"
+        bool isCurrentUserRequestCreator = false;
+        bool isRequestCompleted = false;
+        if (requestId != null) {
+          try {
+            final requestDoc = await FirebaseFirestore.instance
+                .collection('requests')
+                .doc(requestId)
+                .get();
+            
+            if (requestDoc.exists) {
+              final requestData = requestDoc.data()!;
+              final createdBy = requestData['createdBy'] as String?;
+              final status = requestData['status'] as String?;
+              isCurrentUserRequestCreator = createdBy == user.uid;
+              isRequestCompleted = status == 'completed';
+            }
+          } catch (e) {
+            debugPrint('Error checking if current user is request creator: $e');
+          }
+        }
+        
+        if (mounted) {
+          setState(() {
+            _isChatClosed = isClosed;
+            _isDeletedByRequestCreator = isDeletedByRequestCreator;
+            _isDeletedByServiceProvider = isDeletedByServiceProvider;
+            _isCurrentUserRequestCreator = isCurrentUserRequestCreator;
+            _isRequestCompleted = isRequestCompleted;
+          });
+        }
+        
+        debugPrint('Chat status: ${isClosed ? "closed" : "open"}, deleted by request creator: $isDeletedByRequestCreator, deleted by service provider: $isDeletedByServiceProvider');
+      }
+    } catch (e) {
+      debugPrint('Error checking chat status: $e');
+    }
+  }
+  
+  void _listenToChatStatus() {
+    _chatStatusSubscription = FirebaseFirestore.instance
+        .collection('chats')
+        .doc(widget.chatId)
+        .snapshots()
+        .listen((snapshot) {
+      if (snapshot.exists && mounted) {
+        final chatData = snapshot.data()!;
+        final isClosed = chatData['isClosed'] as bool? ?? false;
+        final deletedBy = List<String>.from(chatData['deletedBy'] ?? []);
+        final requestId = chatData['requestId'] as String?;
+        
+        // בדיקה אם הצ'אט נמחק על ידי מבקש השירות או נותן השירות (ללא async)
+        if (requestId != null && deletedBy.isNotEmpty) {
+          final user = FirebaseAuth.instance.currentUser;
+          if (user != null) {
+            // נבדוק את זה באופן אסינכרוני בנפרד
+            _checkIfDeletedByRequestCreator(requestId, deletedBy, user.uid).then((result) {
+              if (mounted) {
+                setState(() {
+                  _isDeletedByRequestCreator = result;
+                });
+              }
+            });
+            
+            // בדיקה אם נותן השירות מחק את הצ'אט
+            _checkIfDeletedByServiceProvider(requestId, deletedBy, user.uid).then((result) {
+              if (mounted) {
+                setState(() {
+                  _isDeletedByServiceProvider = result;
+                });
+              }
+            });
+            
+            // בדיקה אם המשתמש הנוכחי הוא מבקש השירות
+            _checkIfCurrentUserIsRequestCreator(requestId, user.uid).then((result) {
+              if (mounted) {
+                setState(() {
+                  _isCurrentUserRequestCreator = result;
+                });
+              }
+            });
+            
+            // בדיקה אם הבקשה במצב "טופל"
+            _checkIfRequestCompleted(requestId).then((result) {
+              if (mounted) {
+                setState(() {
+                  _isRequestCompleted = result;
+                });
+              }
+            });
+          }
+        }
         
         if (mounted) {
           setState(() {
@@ -170,29 +306,126 @@ class _ChatScreenState extends State<ChatScreen> {
           });
         }
         
-        debugPrint('Chat status: ${isClosed ? "closed" : "open"}');
+        debugPrint('🔄 Chat status updated in real-time: ${isClosed ? "closed" : "open"}');
+      }
+    });
+  }
+  
+  Future<bool> _checkIfDeletedByRequestCreator(String requestId, List<String> deletedBy, String currentUserId) async {
+    try {
+      final requestDoc = await FirebaseFirestore.instance
+          .collection('requests')
+          .doc(requestId)
+          .get();
+      
+      if (requestDoc.exists) {
+        final requestData = requestDoc.data()!;
+        final createdBy = requestData['createdBy'] as String?;
+        
+        // אם יוצר הבקשה מחק את הצ'אט, נותן השירות לא יכול לפתוח אותו מחדש
+        if (createdBy != null && deletedBy.contains(createdBy) && createdBy != currentUserId) {
+          return true;
+        }
       }
     } catch (e) {
-      debugPrint('Error checking chat status: $e');
+      debugPrint('Error checking request creator in listener: $e');
     }
+    return false;
+  }
+  
+  Future<bool> _checkIfDeletedByServiceProvider(String requestId, List<String> deletedBy, String currentUserId) async {
+    try {
+      final requestDoc = await FirebaseFirestore.instance
+          .collection('requests')
+          .doc(requestId)
+          .get();
+      
+      if (requestDoc.exists) {
+        final requestData = requestDoc.data()!;
+        final createdBy = requestData['createdBy'] as String?;
+        
+        // אם המשתמש הנוכחי הוא מבקש השירות, נבדוק אם נותן השירות מחק את הצ'אט
+        if (createdBy == currentUserId) {
+          for (final deletedByUid in deletedBy) {
+            if (deletedByUid != currentUserId) {
+              // מישהו אחר מחק את הצ'אט - זה נותן השירות
+              return true;
+            }
+          }
+        }
+      }
+    } catch (e) {
+      debugPrint('Error checking service provider in listener: $e');
+    }
+    return false;
+  }
+  
+  Future<bool> _checkIfCurrentUserIsRequestCreator(String requestId, String currentUserId) async {
+    try {
+      final requestDoc = await FirebaseFirestore.instance
+          .collection('requests')
+          .doc(requestId)
+          .get();
+      
+      if (requestDoc.exists) {
+        final requestData = requestDoc.data()!;
+        final createdBy = requestData['createdBy'] as String?;
+        return createdBy == currentUserId;
+      }
+    } catch (e) {
+      debugPrint('Error checking if current user is request creator: $e');
+    }
+    return false;
+  }
+  
+  Future<bool> _checkIfRequestCompleted(String requestId) async {
+    try {
+      final requestDoc = await FirebaseFirestore.instance
+          .collection('requests')
+          .doc(requestId)
+          .get();
+      
+      if (requestDoc.exists) {
+        final requestData = requestDoc.data()!;
+        final status = requestData['status'] as String?;
+        return status == 'completed';
+      }
+    } catch (e) {
+      debugPrint('Error checking if request is completed: $e');
+    }
+    return false;
   }
   
   Future<void> _checkExistingMessages() async {
     try {
       debugPrint('Checking existing messages for chat: ${widget.chatId}');
       
-      // בדיקה אם יש הודעת מערכת על הגבלת הודעות
+      final l10n = AppLocalizations.of(context);
+      
+      // בדיקה אם יש הודעת מערכת על הגבלת הודעות (כל שפה)
+      // חיפוש לפי isSystemMessage בלבד, לא לפי טקסט
       final systemMessageSnapshot = await FirebaseFirestore.instance
           .collection('chats')
           .doc(widget.chatId)
           .collection('messages')
           .where('isSystemMessage', isEqualTo: true)
-          .where('text', isEqualTo: 'ניתן לשלוח עד 50 הודעות בצ\'אט זה. הודעות מערכת לא נספרות במגבלה.')
           .get();
       
+      // בדיקה אם יש הודעת הגבלת הודעות (כל שפה)
+      bool hasLimitMessage = false;
+      for (var doc in systemMessageSnapshot.docs) {
+        final text = doc.data()['text'] as String? ?? '';
+        if (text.contains('ניתן לשלוח עד 50') || 
+            text.contains('يمكن إرسال ما يصل إلى 50') ||
+            text.contains('You can send up to 50')) {
+          hasLimitMessage = true;
+          break;
+        }
+      }
+      
       // אם אין הודעת מערכת על הגבלת הודעות, נוסיף אותה
-      if (systemMessageSnapshot.docs.isEmpty) {
-        await _addMessageLimitSystemMessage();
+      if (!hasLimitMessage) {
+        await _addMessageLimitSystemMessage(l10n);
       }
       
       // בדיקה ישירה של ההודעות
@@ -230,7 +463,7 @@ class _ChatScreenState extends State<ChatScreen> {
   }
 
   // הוספת הודעת מערכת על הגבלת הודעות
-  Future<void> _addMessageLimitSystemMessage() async {
+  Future<void> _addMessageLimitSystemMessage(AppLocalizations l10n) async {
     try {
       await FirebaseFirestore.instance
           .collection('chats')
@@ -238,7 +471,7 @@ class _ChatScreenState extends State<ChatScreen> {
           .collection('messages')
           .add({
         'from': 'system',
-        'text': 'ניתן לשלוח עד 50 הודעות בצ\'אט זה. הודעות מערכת לא נספרות במגבלה.',
+        'text': l10n.canSendUpTo50Messages,
         'sentAt': Timestamp.fromDate(DateTime.now()),
         'isSystemMessage': true,
         'isDeleted': false,
@@ -259,6 +492,8 @@ class _ChatScreenState extends State<ChatScreen> {
     // עדכון המצב - המשתמש יצא מהצ'אט
     AppStateService.exitAllChats();
     
+    _chatStatusSubscription?.cancel();
+    _voiceService.dispose();
     _messageController.dispose();
     _scrollController.dispose();
     super.dispose();
@@ -267,21 +502,18 @@ class _ChatScreenState extends State<ChatScreen> {
   @override
   Widget build(BuildContext context) {
     final l10n = AppLocalizations.of(context);
-    
-    // הצגת הודעת הדרכה רק כשהמשתמש נכנס למסך הצ'אט
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      _showChatSpecificTutorial();
-    });
     final user = FirebaseAuth.instance.currentUser;
     
-    return WillPopScope(
-      onWillPop: () async {
+    return PopScope(
+      canPop: false,
+      onPopInvokedWithResult: (bool didPop, dynamic result) async {
+        if (didPop) return;
         // סימון כל ההודעות כנקראות לפני יציאה מהצ'אט
         ChatService.markMessagesAsRead(widget.chatId);
         
         // עדכון המצב כשהמשתמש עוזב את הצ'אט
         AppStateService.exitAllChats();
-        return true;
+        Navigator.of(context).pop();
       },
       child: Directionality(
         textDirection: l10n.isRTL ? TextDirection.rtl : TextDirection.ltr,
@@ -306,7 +538,7 @@ class _ChatScreenState extends State<ChatScreen> {
             ],
           ),
           backgroundColor: Theme.of(context).brightness == Brightness.dark 
-              ? const Color(0xFFFF9800) // כתום ענתיק
+              ? const Color(0xFF9C27B0) // סגול יפה
               : Theme.of(context).colorScheme.primary,
           foregroundColor: Colors.white,
           toolbarHeight: 50,
@@ -332,21 +564,24 @@ class _ChatScreenState extends State<ChatScreen> {
                       children: [
                         const Icon(Icons.lock, color: Colors.orange),
                         const SizedBox(width: 8),
-                        const Text('סגור צ\'אט'),
+                        Text(l10n.closeChat),
                       ],
                     ),
                   ),
                 ] else ...[
-                  PopupMenuItem(
-                    value: 'reopen',
-                    child: Row(
-                      children: [
-                        const Icon(Icons.lock_open, color: Colors.green),
-                        const SizedBox(width: 8),
-                        const Text('פתח צ\'אט מחדש'),
-                      ],
+                  // הצגת "פתח צ'אט סגור" רק אם הצ'אט לא נמחק על ידי מבקש השירות או נותן השירות
+                  // ולא אם הבקשה במצב "טופל"
+                  if (!_isDeletedByRequestCreator && !_isDeletedByServiceProvider && !_isRequestCompleted)
+                    PopupMenuItem(
+                      value: 'reopen',
+                      child: Row(
+                        children: [
+                          const Icon(Icons.lock_open, color: Colors.green),
+                          const SizedBox(width: 8),
+                          Text(l10n.reopenChat),
+                        ],
+                      ),
                     ),
-                  ),
                 ],
                 PopupMenuItem(
                   value: 'clear',
@@ -369,7 +604,11 @@ class _ChatScreenState extends State<ChatScreen> {
               stream: FirebaseFirestore.instance
                   .collection('chats')
                   .doc(widget.chatId)
-                  .snapshots(),
+                  .snapshots()
+                  .handleError((error) {
+                    debugPrint('❌ Error in chat status stream: $error');
+                    // Log error and continue - StreamBuilder will handle error state
+                  }),
               builder: (context, chatStatusSnapshot) {
                 if (chatStatusSnapshot.hasData && chatStatusSnapshot.data!.exists) {
                   final chatData = chatStatusSnapshot.data!.data() as Map<String, dynamic>;
@@ -391,16 +630,84 @@ class _ChatScreenState extends State<ChatScreen> {
               },
             ),
             Expanded(
-              child: StreamBuilder<QuerySnapshot>(
-                stream: _messagesStream,
-                builder: (context, snapshot) {
-                  debugPrint('StreamBuilder state: ${snapshot.connectionState}');
-                  debugPrint('StreamBuilder hasError: ${snapshot.hasError}');
-                  debugPrint('StreamBuilder hasData: ${snapshot.hasData}');
-                  debugPrint('StreamBuilder docs count: ${snapshot.data?.docs.length ?? 0}');
+              child: Consumer(
+                builder: (context, ref, child) {
+                  final messagesAsync = ref.watch(chatMessagesProvider(widget.chatId));
                   
-                  if (snapshot.connectionState == ConnectionState.waiting) {
+                  // האזנה לשינויים בהודעות דרך provider - סימון הודעות כנקראות
+                  ref.listen<AsyncValue<List<Message>>>(
+                    chatMessagesProvider(widget.chatId),
+                    (previous, next) {
+                      if (!mounted) return;
+                      
+                      final user = FirebaseAuth.instance.currentUser;
+                      if (user == null) return;
+                      
+                      // רק אם המשתמש באמת נכנס לצ'אט (המסך פעיל)
+                      if (ModalRoute.of(context)?.isCurrent == true) {
+                        next.whenData((messages) {
+                          for (final message in messages) {
+                            // אם ההודעה לא נשלחה על ידי המשתמש הנוכחי ולא נקראה על ידו
+                            if (message.from != user.uid && !message.readBy.contains(user.uid)) {
+                              _markMessageAsRead(message.messageId, user.uid);
+                            }
+                          }
+                          
+                          // גלילה למטה כשההודעות נטענות או מתעדכנות
+                          WidgetsBinding.instance.addPostFrameCallback((_) {
+                            if (_scrollController.hasClients) {
+                              _scrollController.animateTo(
+                                _scrollController.position.maxScrollExtent,
+                                duration: const Duration(milliseconds: 300),
+                                curve: Curves.easeOut,
+                              );
+                            }
+                          });
+                        });
+                      }
+                    },
+                  );
+                  
+                  return messagesAsync.when(
+                    data: (messages) {
+                      if (messages.isEmpty) {
                     return Center(
+                          child: Column(
+                            mainAxisAlignment: MainAxisAlignment.center,
+                            children: [
+                              Icon(
+                                Icons.chat_bubble_outline,
+                                size: 64,
+                                color: Theme.of(context).colorScheme.onSurfaceVariant,
+                              ),
+                              const SizedBox(height: 16),
+                              Text(
+                                l10n.noMessages,
+                                style: TextStyle(
+                                  fontSize: 16,
+                                  color: Theme.of(context).colorScheme.onSurfaceVariant,
+                                ),
+                              ),
+                            ],
+                          ),
+                        );
+                      }
+
+                      debugPrint('Loaded ${messages.length} messages for chat: ${widget.chatId}');
+
+                      return ListView.builder(
+                        controller: _scrollController,
+                        reverse: false, // הודעות חדשות למטה כמו WhatsApp
+                        itemCount: messages.length,
+                        itemBuilder: (context, index) {
+                          final message = messages[index];
+                          final isMe = message.from == user?.uid;
+                          
+                          return _buildMessageBubble(message, isMe, l10n);
+                        },
+                      );
+                    },
+                    loading: () => Center(
                       child: Column(
                         mainAxisAlignment: MainAxisAlignment.center,
                         children: [
@@ -411,7 +718,7 @@ class _ChatScreenState extends State<ChatScreen> {
                               borderRadius: BorderRadius.circular(20),
                               boxShadow: [
                                 BoxShadow(
-                                  color: Colors.black.withOpacity(0.1),
+                                  color: Colors.black.withValues(alpha: 0.1),
                                   blurRadius: 10,
                                   offset: const Offset(0, 5),
                                 ),
@@ -432,11 +739,11 @@ class _ChatScreenState extends State<ChatScreen> {
                                 ),
                                 const SizedBox(height: 16),
                                 Text(
-                                  'טוען הודעות...',
+                                  l10n.loadingMessages,
                                   style: TextStyle(
                                     fontSize: 16,
                                     fontWeight: FontWeight.w500,
-                                    color: Colors.grey[700],
+                                    color: Theme.of(context).colorScheme.onSurfaceVariant,
                                   ),
                                 ),
                               ],
@@ -444,58 +751,12 @@ class _ChatScreenState extends State<ChatScreen> {
                           ),
                         ],
                       ),
-                    );
-                  }
-
-                  if (snapshot.hasError) {
-                    debugPrint('StreamBuilder error: ${snapshot.error}');
+                    ),
+                    error: (error, stack) {
+                      debugPrint('Error loading messages: $error');
                     return Center(
-                      child: Text('שגיאה: ${snapshot.error}'),
+                        child: Text(l10n.errorLoadingMessages(error.toString())),
                     );
-                  }
-
-                  if (!snapshot.hasData || snapshot.data!.docs.isEmpty) {
-                    debugPrint('No messages found for chat: ${widget.chatId}');
-                    return Center(
-                      child: Column(
-                        mainAxisAlignment: MainAxisAlignment.center,
-                        children: [
-                          Icon(
-                            Icons.chat_bubble_outline,
-                            size: 64,
-                            color: Colors.grey[400],
-                          ),
-                          const SizedBox(height: 16),
-                          Text(
-                            l10n.noMessages,
-                            style: TextStyle(
-                              fontSize: 16,
-                              color: Colors.grey[600],
-                            ),
-                          ),
-                        ],
-                      ),
-                    );
-                  }
-
-                  final messages = snapshot.data!.docs
-                      .map((doc) => Message.fromFirestore(doc))
-                      .toList();
-                  
-                  // מיון ידני לפי זמן שליחה (חדשות למטה, הישנות למעלה)
-                  messages.sort((a, b) => b.sentAt.compareTo(a.sentAt));
-                  
-                  debugPrint('Loaded ${messages.length} messages for chat: ${widget.chatId}');
-
-                  return ListView.builder(
-                    controller: _scrollController,
-                    reverse: true,
-                    itemCount: messages.length,
-                    itemBuilder: (context, index) {
-                      final message = messages[index];
-                      final isMe = message.from == user?.uid;
-                      
-                      return _buildMessageBubble(message, isMe, l10n);
                     },
                   );
                 },
@@ -512,23 +773,29 @@ class _ChatScreenState extends State<ChatScreen> {
 
   // הצגת דיאלוג אפשרויות הודעה (עריכה ומחיקה)
   void _showMessageOptionsDialog(Message message) {
+    // הודעות קוליות לא ניתן לערוך
+    final canEdit = message.type != MessageType.voice;
+    final l10n = AppLocalizations.of(context);
+    
     showDialog(
       context: context,
       builder: (context) => AlertDialog(
-        title: const Text('אפשרויות הודעה'),
-        content: const Text('מה תרצה לעשות עם ההודעה?'),
+        title: Text(l10n.messageOptions),
+        content: Text(l10n.whatDoYouWantToDoWithMessage),
         actions: [
+          // כפתור עריכה - רק אם ההודעה לא קולית
+          if (canEdit)
           TextButton(
             onPressed: () {
               Navigator.of(context).pop();
               _showEditMessageDialog(message);
             },
-            child: const Row(
+              child: Row(
               mainAxisSize: MainAxisSize.min,
               children: [
                 Icon(Icons.edit, size: 16),
                 SizedBox(width: 4),
-                Text('ערוך'),
+                  Text(l10n.edit),
               ],
             ),
           ),
@@ -537,18 +804,18 @@ class _ChatScreenState extends State<ChatScreen> {
               Navigator.of(context).pop();
               _showDeleteMessageDialog(message);
             },
-            child: const Row(
+            child: Row(
               mainAxisSize: MainAxisSize.min,
               children: [
                 Icon(Icons.delete, size: 16, color: Colors.red),
                 SizedBox(width: 4),
-                Text('מחק', style: TextStyle(color: Colors.red)),
+                Text(l10n.delete, style: TextStyle(color: Colors.red)),
               ],
             ),
           ),
           TextButton(
             onPressed: () => Navigator.of(context).pop(),
-            child: const Text('ביטול'),
+            child: Text(l10n.cancel),
           ),
         ],
       ),
@@ -558,15 +825,16 @@ class _ChatScreenState extends State<ChatScreen> {
   // הצגת דיאלוג עריכת הודעה
   void _showEditMessageDialog(Message message) {
     final TextEditingController editController = TextEditingController(text: message.text);
+    final l10n = AppLocalizations.of(context);
     
     showDialog(
       context: context,
       builder: (context) => AlertDialog(
-        title: const Text('עריכת הודעה'),
+        title: Text(l10n.editMessage),
         content: TextField(
           controller: editController,
-          decoration: const InputDecoration(
-            hintText: 'הקלד את ההודעה החדשה...',
+          decoration: InputDecoration(
+            hintText: l10n.typeNewMessage,
             border: OutlineInputBorder(),
           ),
           maxLines: 3,
@@ -574,14 +842,14 @@ class _ChatScreenState extends State<ChatScreen> {
         actions: [
           TextButton(
             onPressed: () => Navigator.of(context).pop(),
-            child: const Text('ביטול'),
+            child: Text(l10n.cancel),
           ),
           TextButton(
             onPressed: () {
               Navigator.of(context).pop();
               _editMessage(message, editController.text.trim());
             },
-            child: const Text('שמור'),
+            child: Text(l10n.save),
           ),
         ],
       ),
@@ -604,9 +872,10 @@ class _ChatScreenState extends State<ChatScreen> {
       });
 
       if (mounted) {
+        final l10n = AppLocalizations.of(context);
         ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(
-            content: Text('ההודעה נערכה בהצלחה'),
+          SnackBar(
+            content: Text(l10n.messageEditedSuccessfully),
             backgroundColor: Colors.green,
           ),
         );
@@ -614,9 +883,10 @@ class _ChatScreenState extends State<ChatScreen> {
     } catch (e) {
       debugPrint('Error editing message: $e');
       if (mounted) {
+        final l10n = AppLocalizations.of(context);
         ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(
-            content: Text('שגיאה בעריכת ההודעה'),
+          SnackBar(
+            content: Text(l10n.errorEditingMessage(e.toString())),
             backgroundColor: Colors.red,
           ),
         );
@@ -626,15 +896,16 @@ class _ChatScreenState extends State<ChatScreen> {
 
   // הצגת דיאלוג מחיקת הודעה
   void _showDeleteMessageDialog(Message message) {
+    final l10n = AppLocalizations.of(context);
     showDialog(
       context: context,
       builder: (context) => AlertDialog(
-        title: const Text('מחיקת הודעה'),
-        content: const Text('האם אתה בטוח שברצונך למחוק את ההודעה?'),
+        title: Text(l10n.deleteMessageTitle),
+        content: Text(l10n.deleteMessageConfirm),
         actions: [
           TextButton(
             onPressed: () => Navigator.of(context).pop(),
-            child: const Text('ביטול'),
+            child: Text(l10n.cancel),
           ),
           TextButton(
             onPressed: () async {
@@ -652,6 +923,7 @@ class _ChatScreenState extends State<ChatScreen> {
   // מחיקת הודעה
   Future<void> _deleteMessage(Message message) async {
     try {
+      final l10n = AppLocalizations.of(context);
       final success = await ChatService.deleteMessage(
         chatId: widget.chatId,
         messageId: message.messageId,
@@ -660,17 +932,18 @@ class _ChatScreenState extends State<ChatScreen> {
       if (success) {
         if (mounted) {
           ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(
-              content: Text('ההודעה נמחקה בהצלחה'),
+            SnackBar(
+              content: Text(l10n.messageDeletedSuccessfully),
               backgroundColor: Colors.green,
             ),
           );
         }
       } else {
         if (mounted) {
+          final l10n = AppLocalizations.of(context);
           ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(
-              content: Text('שגיאה במחיקת ההודעה'),
+            SnackBar(
+              content: Text(l10n.errorDeletingMessage),
               backgroundColor: Colors.red,
             ),
           );
@@ -679,9 +952,10 @@ class _ChatScreenState extends State<ChatScreen> {
     } catch (e) {
       debugPrint('Error deleting message: $e');
       if (mounted) {
+        final l10n = AppLocalizations.of(context);
         ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(
-            content: Text('שגיאה במחיקת ההודעה'),
+          SnackBar(
+            content: Text(l10n.errorDeletingMessage),
             backgroundColor: Colors.red,
           ),
         );
@@ -692,15 +966,23 @@ class _ChatScreenState extends State<ChatScreen> {
   Widget _buildMessageBubble(Message message, bool isMe, AppLocalizations l10n) {
     // הצגת הודעות מערכת בצורה מיוחדת
     if (message.isSystemMessage) {
+      // אם זו הודעת הגבלת הודעות, הצג בשפה הנוכחית
+      String displayText = message.text;
+      if (message.text.contains('ניתן לשלוח עד 50') || 
+          message.text.contains('يمكن إرسال ما يصل إلى 50') ||
+          message.text.contains('You can send up to 50')) {
+        displayText = l10n.canSendUpTo50Messages;
+      }
+      
       return Padding(
         padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
         child: Center(
           child: Container(
             padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
             decoration: BoxDecoration(
-              color: Colors.blue[50],
+              color: Theme.of(context).colorScheme.primaryContainer,
               borderRadius: BorderRadius.circular(20),
-              border: Border.all(color: Colors.blue[200]!, width: 1),
+              border: Border.all(color: Theme.of(context).colorScheme.primary, width: 1),
             ),
             child: Row(
               mainAxisSize: MainAxisSize.min,
@@ -708,14 +990,14 @@ class _ChatScreenState extends State<ChatScreen> {
                 Icon(
                   Icons.info_outline,
                   size: 16,
-                  color: Colors.blue[600],
+                  color: Theme.of(context).colorScheme.primary,
                 ),
                 const SizedBox(width: 8),
                 Flexible(
                   child: Text(
-                    message.text,
+                    displayText,
                     style: TextStyle(
-                      color: Colors.blue[800],
+                      color: Theme.of(context).colorScheme.onPrimaryContainer,
                       fontSize: 14,
                       fontWeight: FontWeight.w500,
                     ),
@@ -738,7 +1020,7 @@ class _ChatScreenState extends State<ChatScreen> {
             CircleAvatar(
               radius: 16,
               backgroundColor: Theme.of(context).brightness == Brightness.dark 
-              ? const Color(0xFFFF9800) // כתום ענתיק
+              ? const Color(0xFF9C27B0) // סגול יפה
               : Theme.of(context).colorScheme.primary,
               child: const Icon(Icons.person, size: 16, color: Colors.white),
             ),
@@ -751,10 +1033,10 @@ class _ChatScreenState extends State<ChatScreen> {
                 padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
                 decoration: BoxDecoration(
 color: message.isDeleted 
-                      ? Colors.grey[300]
+                      ? Theme.of(context).colorScheme.surfaceContainerHighest
                       : (isMe 
                           ? Theme.of(context).colorScheme.primary
-                          : Colors.grey[200]),
+                          : Theme.of(context).colorScheme.surfaceContainer),
                   borderRadius: BorderRadius.circular(20).copyWith(
                     bottomLeft: isMe ? const Radius.circular(20) : const Radius.circular(4),
                     bottomRight: isMe ? const Radius.circular(4) : const Radius.circular(20),
@@ -766,20 +1048,21 @@ color: message.isDeleted
                     FutureBuilder<String>(
                       future: _getUserName(message.from),
                       builder: (context, snapshot) {
-                        String displayName;
-                        if (isMe) {
+                        // תמיד נציג את השם מהמסד הנתונים, גם אם זו ההודעה שלי
+                        String displayName = snapshot.data ?? l10n.otherUser;
+                        
+                        // אם זו ההודעה שלי ולא קיבלנו שם מהמסד הנתונים, ננסה לקבל מ-Firebase Auth
+                        if (isMe && (snapshot.data == null || snapshot.data == l10n.otherUser)) {
                           final currentUser = FirebaseAuth.instance.currentUser;
                           displayName = currentUser?.displayName ?? currentUser?.email?.split('@')[0] ?? l10n.you;
-                        } else {
-                          displayName = snapshot.data ?? l10n.otherUser;
                         }
                         
                         return Text(
                           displayName,
                           style: TextStyle(
                             color: message.isDeleted 
-                                ? Colors.grey[500]
-                                : (isMe ? Colors.white70 : Colors.grey[600]),
+                                ? Theme.of(context).colorScheme.onSurfaceVariant
+                                : (isMe ? Theme.of(context).colorScheme.onPrimary : Theme.of(context).colorScheme.onSurfaceVariant),
                             fontSize: 12,
                             fontWeight: FontWeight.bold,
                           ),
@@ -787,15 +1070,112 @@ color: message.isDeleted
                       },
                     ),
                     const SizedBox(height: 2),
-                    Text(
-                      message.isDeleted ? 'הודעה נמחקה' : message.text,
-                      style: TextStyle(
-                        color: message.isDeleted 
-                            ? Colors.grey[500]
-                            : (isMe ? Colors.white : Colors.black87),
-                        fontSize: 16,
-                        fontStyle: message.isDeleted ? FontStyle.italic : FontStyle.normal,
-                      ),
+                    FutureBuilder<String>(
+                      future: _getUserName(message.from),
+                      builder: (context, userNameSnapshot) {
+                        final currentL10n = AppLocalizations.of(context);
+                        String displayText = message.isDeleted ? currentL10n.messageDeleted : message.text;
+                        
+                        // אם זו הודעת "שלום! אני.." - הצג בשפה הנוכחית
+                        if (!message.isDeleted && !message.isSystemMessage) {
+                          // בדיקה אם זו הודעת "שלום! אני.." (כל שפה)
+                          if (message.text.contains('שלום! אני') || 
+                              message.text.contains('Hello! I am') ||
+                              message.text.contains('مرحباً! أنا')) {
+                            // קבלת שם המשתמש מהפרופיל
+                            final userName = userNameSnapshot.data ?? message.from;
+                            
+                            if (userName.isNotEmpty && userName != message.from) {
+                              // חילוץ קטגוריה ודירוג מההודעה המקורית
+                              String? category;
+                              bool hasRating = false;
+                              String? ratingPart;
+                              
+                              // חילוץ מההודעה המקורית - כל שפה
+                              if (message.text.contains('⭐')) {
+                                hasRating = true;
+                                final ratingMatch = RegExp(r'\(([^)]+⭐[^)]+)\)').firstMatch(message.text);
+                                if (ratingMatch != null) {
+                                  ratingPart = ratingMatch.group(1);
+                                }
+                              } else {
+                                // חילוץ קטגוריה
+                                if (message.text.contains('מתחום')) {
+                                  final categoryMatch = RegExp(r'מתחום\s+([^\s]+)').firstMatch(message.text);
+                                  if (categoryMatch != null) {
+                                    category = categoryMatch.group(1)?.trim();
+                                  }
+                                } else if (message.text.contains('from field')) {
+                                  final categoryMatch = RegExp(r'from field\s+([^\s]+)').firstMatch(message.text);
+                                  if (categoryMatch != null) {
+                                    category = categoryMatch.group(1)?.trim();
+                                  }
+                                } else if (message.text.contains('من مجال')) {
+                                  final categoryMatch = RegExp(r'من مجال\s+([^\s]+)').firstMatch(message.text);
+                                  if (categoryMatch != null) {
+                                    category = categoryMatch.group(1)?.trim();
+                                  }
+                                }
+                              }
+                              
+                              // חילוץ badge אם קיים
+                              String badge = '';
+                              if (message.text.contains('🏆 מומחה') || 
+                                  message.text.contains('🏆') ||
+                                  message.text.contains('خبير')) {
+                                badge = ' 🏆 מומחה'; // נשתמש באותה badge לכל השפות
+                              }
+                              
+                              // בניית ההודעה בשפה הנוכחית
+                              displayText = currentL10n.helloIAm(userName, badge);
+                              
+                              // הוספת חלק הדירוג או "חדש בתחום"
+                              if (hasRating && ratingPart != null) {
+                                displayText += ' ($ratingPart)';
+                              } else if (category != null && category.isNotEmpty) {
+                                displayText += ' ${currentL10n.newInField(category)}';
+                              }
+                              
+                              // הוספת "מעוניין לעזור"
+                              displayText += ' ${currentL10n.interestedInHelping}';
+                            }
+                          }
+                        }
+                        
+                        // Check if this is a voice message
+                        if (message.type == MessageType.voice) {
+                          // אם ההודעה הקולית נמחקה, הצג "הודעה נמחקה" (גם למשתמש השני)
+                          if (message.isDeleted) {
+                            return Text(
+                              currentL10n.messageDeleted,
+                              style: TextStyle(
+                                color: isMe ? Theme.of(context).colorScheme.onSurfaceVariant : Theme.of(context).colorScheme.onSurfaceVariant,
+                                fontStyle: FontStyle.italic,
+                                fontSize: 16,
+                              ),
+                            );
+                          }
+                          // אחרת, הצג את ה-VoiceMessageWidget
+                          return VoiceMessageWidget(
+                            data: message.data, // Can be Base64 or URL
+                            duration: message.duration,
+                            isMe: isMe,
+                          );
+                        }
+                        
+                        return Text(
+                          displayText,
+                          style: TextStyle(
+                            color: message.isDeleted 
+                                ? Theme.of(context).colorScheme.onSurfaceVariant
+                                : (isMe 
+                                    ? Colors.white 
+                                    : Theme.of(context).colorScheme.onSurface),
+                            fontSize: 16,
+                            fontStyle: message.isDeleted ? FontStyle.italic : FontStyle.normal,
+                          ),
+                        );
+                      },
                     ),
                     const SizedBox(height: 4),
                     Row(
@@ -805,8 +1185,8 @@ color: message.isDeleted
                           _formatTime(message.sentAt),
                           style: TextStyle(
                             color: message.isDeleted 
-                                ? Colors.grey[500]
-                                : (isMe ? Colors.white70 : Colors.grey[600]),
+                                ? Theme.of(context).colorScheme.onSurfaceVariant
+                                : (isMe ? Theme.of(context).colorScheme.onPrimary : Theme.of(context).colorScheme.onSurfaceVariant),
                             fontSize: 12,
                           ),
                         ),
@@ -825,7 +1205,7 @@ color: message.isDeleted
             const SizedBox(width: 8),
             CircleAvatar(
               radius: 16,
-              backgroundColor: Colors.grey[300],
+              backgroundColor: Theme.of(context).colorScheme.surfaceContainerHighest,
               child: const Icon(Icons.person, size: 16, color: Colors.white),
             ),
           ],
@@ -875,7 +1255,7 @@ color: message.isDeleted
           return Icon(
             Icons.done,
             size: 18,
-            color: Colors.grey[600],
+            color: Theme.of(context).colorScheme.onSurfaceVariant,
           );
         }
       },
@@ -906,7 +1286,7 @@ color: message.isDeleted
       return Container(
         padding: const EdgeInsets.all(16),
         decoration: BoxDecoration(
-          color: Colors.grey[100],
+          color: Theme.of(context).colorScheme.surfaceContainer,
           boxShadow: [
             BoxShadow(
               color: Colors.grey.withValues(alpha: 0.2),
@@ -918,13 +1298,13 @@ color: message.isDeleted
         ),
         child: Row(
           children: [
-            Icon(Icons.lock, color: Colors.grey[600]),
+            Icon(Icons.lock, color: Theme.of(context).colorScheme.onSurfaceVariant),
             const SizedBox(width: 8),
             Expanded(
               child: Text(
                 'הצ\'אט נסגר - הטיפול בבקשה הסתיים',
                 style: TextStyle(
-                  color: Colors.grey[600],
+                  color: Theme.of(context).colorScheme.onSurfaceVariant,
                   fontStyle: FontStyle.italic,
                 ),
               ),
@@ -936,9 +1316,9 @@ color: message.isDeleted
                 _showDeleteChatDialog(l10n);
               },
               icon: const Icon(Icons.delete, size: 16),
-              label: const Text('מחק צ\'אט'),
+              label: Text(l10n.deleteChat),
               style: ElevatedButton.styleFrom(
-                backgroundColor: Colors.red[600],
+                backgroundColor: Theme.of(context).colorScheme.error,
                 foregroundColor: Colors.white,
                 padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
               ),
@@ -961,38 +1341,156 @@ color: message.isDeleted
           ),
         ],
       ),
-      child: Row(
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
         children: [
-          Expanded(
-            child: TextField(
-              controller: _messageController,
-              decoration: InputDecoration(
-                hintText: l10n.sendMessage,
-                border: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(25),
+          // הודעה שהצ'אט סגור (אם הצ'אט סגור)
+          if (_isChatClosed)
+            Container(
+              padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
+              margin: const EdgeInsets.only(bottom: 8),
+              decoration: BoxDecoration(
+                color: Theme.of(context).colorScheme.tertiaryContainer,
+                borderRadius: BorderRadius.circular(8),
+                border: Border.all(color: Theme.of(context).colorScheme.tertiary),
+              ),
+              child: Row(
+                children: [
+                  Icon(Icons.lock, color: Theme.of(context).colorScheme.tertiary, size: 16),
+                  const SizedBox(width: 8),
+                  Expanded(
+                    child: Text(
+                      'הצ\'אט סגור - ניתן לצפות בהיסטוריית השיחה בלבד',
+                      style: TextStyle(
+                        color: Theme.of(context).colorScheme.onTertiaryContainer,
+                        fontSize: 12,
+                        fontWeight: FontWeight.w500,
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ),
+          // Voice recorder widget
+          if (_showVoiceRecorder)
+            Padding(
+              padding: const EdgeInsets.only(bottom: 8),
+              child: VoiceRecorderWidget(
+                onRecordingComplete: (filePath) async {
+                  if (filePath != null) {
+                    await _sendVoiceMessage(filePath);
+                  }
+                  setState(() {
+                    _showVoiceRecorder = false;
+                    _isRecordingVoice = false;
+                  });
+                },
+                onCancel: () {
+                  setState(() {
+                    _showVoiceRecorder = false;
+                    _isRecordingVoice = false;
+                  });
+                },
+              ),
+            ),
+          Row(
+            children: [
+              // Microphone button for voice messages
+              if (!_isChatClosed && 
+                  !(_isCurrentUserRequestCreator && _isDeletedByServiceProvider) &&
+                  !(!_isCurrentUserRequestCreator && _isDeletedByRequestCreator))
+                GestureDetector(
+                  onLongPressStart: (_) {
+                    setState(() {
+                      _showVoiceRecorder = true;
+                      _isRecordingVoice = true;
+                    });
+                  },
+                  onLongPressEnd: (_) {
+                    // Recording will be stopped by VoiceRecorderWidget
+                  },
+                  child: Container(
+                    padding: const EdgeInsets.all(12),
+                    decoration: BoxDecoration(
+                      color: _isRecordingVoice ? Theme.of(context).colorScheme.error : Theme.of(context).colorScheme.surfaceContainerHighest,
+                      shape: BoxShape.circle,
+                    ),
+                    child: Icon(
+                      Icons.mic,
+                      color: _isRecordingVoice ? Theme.of(context).colorScheme.onError : Theme.of(context).colorScheme.onSurfaceVariant,
+                    ),
+                  ),
                 ),
-                contentPadding: const EdgeInsets.symmetric(
-                  horizontal: 16,
-                  vertical: 12,
+              if (!_isChatClosed && 
+                  !(_isCurrentUserRequestCreator && _isDeletedByServiceProvider) &&
+                  !(!_isCurrentUserRequestCreator && _isDeletedByRequestCreator))
+                const SizedBox(width: 8),
+              Expanded(
+                child: TextField(
+                  controller: _messageController,
+                  enabled: !_isChatClosed && 
+                      !(_isCurrentUserRequestCreator && _isDeletedByServiceProvider) &&
+                      !(!_isCurrentUserRequestCreator && _isDeletedByRequestCreator), // disabled כאשר הצ'אט סגור או נמחק על ידי המשתמש השני
+                  decoration: InputDecoration(
+                    hintText: _getChatDisabledHint(l10n),
+                    border: OutlineInputBorder(
+                      borderRadius: BorderRadius.circular(25),
+                    ),
+                    contentPadding: const EdgeInsets.symmetric(
+                      horizontal: 16,
+                      vertical: 12,
+                    ),
+                    filled: _isChatClosed || 
+                        (_isCurrentUserRequestCreator && _isDeletedByServiceProvider) ||
+                        (!_isCurrentUserRequestCreator && _isDeletedByRequestCreator),
+                    fillColor: (_isChatClosed || 
+                        (_isCurrentUserRequestCreator && _isDeletedByServiceProvider) ||
+                        (!_isCurrentUserRequestCreator && _isDeletedByRequestCreator)) 
+                        ? Theme.of(context).colorScheme.surfaceContainer 
+                        : null,
+                  ),
+                  maxLines: null,
+                  textInputAction: TextInputAction.send,
+                  onSubmitted: (_isChatClosed || 
+                      (_isCurrentUserRequestCreator && _isDeletedByServiceProvider) ||
+                      (!_isCurrentUserRequestCreator && _isDeletedByRequestCreator)) 
+                      ? null 
+                      : (_) => _sendMessage(),
                 ),
               ),
-              maxLines: null,
-              textInputAction: TextInputAction.send,
-              onSubmitted: (_) => _sendMessage(),
-            ),
-          ),
-          const SizedBox(width: 8),
-          FloatingActionButton(
-            onPressed: _sendMessage,
-            mini: true,
-            backgroundColor: Theme.of(context).brightness == Brightness.dark 
-              ? const Color(0xFFFF9800) // כתום ענתיק
-              : Theme.of(context).colorScheme.primary,
-            child: const Icon(Icons.send, color: Colors.white),
+              const SizedBox(width: 8),
+              FloatingActionButton(
+                onPressed: (_isChatClosed || 
+                    (_isCurrentUserRequestCreator && _isDeletedByServiceProvider) ||
+                    (!_isCurrentUserRequestCreator && _isDeletedByRequestCreator)) 
+                    ? null 
+                    : _sendMessage, // disabled כאשר הצ'אט סגור או נמחק על ידי המשתמש השני
+                mini: true,
+                backgroundColor: (_isChatClosed || 
+                    (_isCurrentUserRequestCreator && _isDeletedByServiceProvider) ||
+                    (!_isCurrentUserRequestCreator && _isDeletedByRequestCreator))
+                  ? Theme.of(context).colorScheme.onSurfaceVariant
+                  : (Theme.of(context).brightness == Brightness.dark 
+                      ? const Color(0xFF9C27B0) // סגול יפה
+                      : Theme.of(context).colorScheme.primary),
+                child: const Icon(Icons.send, color: Colors.white),
+              ),
+            ],
           ),
         ],
       ),
     );
+  }
+
+  String _getChatDisabledHint(AppLocalizations l10n) {
+    if (_isChatClosed) {
+      return 'הצ\'אט סגור - לא ניתן לשלוח הודעות';
+    } else if (_isCurrentUserRequestCreator && _isDeletedByServiceProvider) {
+      return 'הצ\'אט נמחק על ידי נותן השירות - לא ניתן לשלוח הודעות';
+    } else if (!_isCurrentUserRequestCreator && _isDeletedByRequestCreator) {
+      return 'הצ\'אט נמחק על ידי מבקש השירות - לא ניתן לשלוח הודעות';
+    }
+    return l10n.sendMessage;
   }
 
   // ספירת הודעות (ללא הודעות מערכת)
@@ -1012,32 +1510,102 @@ color: message.isDeleted
     }
   }
 
+  /// Send voice message
+  Future<void> _sendVoiceMessage(String filePath) async {
+    try {
+      // Process voice message (convert to Base64 or upload to file.io)
+      final processed = await _voiceService.processVoiceMessage(filePath);
+      
+      // Determine if we should use data (Base64) or url (file.io)
+      final voiceData = processed['data'] as String?;
+      final voiceUrl = processed['url'] as String?;
+      final voiceDataOrUrl = voiceData ?? voiceUrl; // Use Base64 if available, otherwise URL
+      
+      // Send voice message via ChatService (includes notification and message creation)
+      await ChatService.sendMessageWithNotification(
+        chatId: widget.chatId,
+        text: '🎤 הודעה קולית',
+        type: 'voice',
+        data: voiceDataOrUrl,
+        duration: processed['duration'] as int?,
+      );
+      
+      if (mounted) {
+        // Scroll to bottom (new messages at bottom)
+        WidgetsBinding.instance.addPostFrameCallback((_) {
+          if (_scrollController.hasClients) {
+            _scrollController.animateTo(
+              _scrollController.position.maxScrollExtent,
+              duration: const Duration(milliseconds: 300),
+              curve: Curves.easeOut,
+            );
+          }
+        });
+      }
+    } catch (e) {
+      debugPrint('❌ Error sending voice message: $e');
+      if (mounted) {
+        final l10n = AppLocalizations.of(context);
+        ScaffoldMessenger.of(context).showSnackBar(
+          SnackBar(
+            content: Text(l10n.errorSendingVoiceMessage(e.toString())),
+            backgroundColor: Colors.red,
+          ),
+        );
+      }
+    }
+  }
+
   Future<void> _sendMessage() async {
     final text = _messageController.text.trim();
     if (text.isEmpty) return;
 
-    // בדיקה אם הצ'אט סגור
+    final l10n = AppLocalizations.of(context);
+    final user = FirebaseAuth.instance.currentUser;
+    if (user == null) return;
+    
+    // בדיקה אם הצ'אט נמחק על ידי המשתמש השני
+    // אם מבקש השירות מחק את הצ'אט, נותן השירות לא יוכל לשלוח הודעות
+    // אם נותן השירות מחק את הצ'אט, מבקש השירות לא יוכל לשלוח הודעות
+    bool cannotSendMessage = false;
+    String? reasonMessage;
+    
     if (_isChatClosed) {
+      cannotSendMessage = true;
+      reasonMessage = l10n.chatClosedCannotSend;
+    } else {
+      // בדיקה אם המשתמש השני מחק את הצ'אט
+      // אם המשתמש הנוכחי הוא מבקש השירות והצ'אט נמחק על ידי נותן השירות
+      if (_isCurrentUserRequestCreator && _isDeletedByServiceProvider) {
+        cannotSendMessage = true;
+        reasonMessage = 'לא ניתן לשלוח הודעות - הצ\'אט נמחק על ידי נותן השירות';
+      }
+      // אם המשתמש הנוכחי הוא נותן השירות והצ'אט נמחק על ידי מבקש השירות
+      else if (!_isCurrentUserRequestCreator && _isDeletedByRequestCreator) {
+        cannotSendMessage = true;
+        reasonMessage = 'לא ניתן לשלוח הודעות - הצ\'אט נמחק על ידי מבקש השירות';
+      }
+    }
+    
+    if (cannotSendMessage) {
       if (!mounted) return;
       ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(
-          content: Text('הצ\'אט נסגר - לא ניתן לשלוח הודעות'),
+        SnackBar(
+          content: Text(reasonMessage ?? l10n.chatClosedCannotSend),
           backgroundColor: Colors.red,
         ),
       );
       return;
     }
 
-    final user = FirebaseAuth.instance.currentUser;
-    if (user == null) return;
-
     // בדיקת הגבלת הודעות
     final messageCount = await _getMessageCount();
     if (messageCount >= 50) {
       if (!mounted) return;
+      final l10n = AppLocalizations.of(context);
       ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(
-          content: Text('הגעת למגבלת 50 הודעות - לא ניתן לשלוח הודעות נוספות'),
+        SnackBar(
+          content: Text(l10n.reached50MessageLimit),
           backgroundColor: Colors.orange,
           duration: Duration(seconds: 3),
         ),
@@ -1048,9 +1616,10 @@ color: message.isDeleted
     // הצגת אזהרה כשהולכים להגיע למגבלה
     if (messageCount >= 45) {
       if (!mounted) return;
+      final l10n = AppLocalizations.of(context);
       ScaffoldMessenger.of(context).showSnackBar(
         SnackBar(
-          content: Text('אזהרה: נותרו ${50 - messageCount} הודעות בלבד'),
+          content: Text(l10n.warningMessagesRemaining(50 - messageCount)),
           backgroundColor: Colors.amber,
           duration: const Duration(seconds: 2),
         ),
@@ -1058,23 +1627,26 @@ color: message.isDeleted
     }
 
     try {
-      // שימוש בפונקציה החדשה עם התראות
+      // Send message via ChatService (includes notification, message creation, and 50 message limit check)
       final success = await ChatService.sendMessageWithNotification(
         chatId: widget.chatId,
         text: text,
+        type: 'text',
       );
 
       if (success) {
         _messageController.clear();
         
-        // גלילה למטה
+        // גלילה למטה (הודעות חדשות למטה)
+        WidgetsBinding.instance.addPostFrameCallback((_) {
         if (_scrollController.hasClients) {
           _scrollController.animateTo(
-            0,
+              _scrollController.position.maxScrollExtent,
             duration: const Duration(milliseconds: 300),
             curve: Curves.easeOut,
           );
         }
+        });
       } else {
         throw Exception('Failed to send message');
       }
@@ -1082,7 +1654,7 @@ color: message.isDeleted
       if (!mounted) return;
       ScaffoldMessenger.of(context).showSnackBar(
         SnackBar(
-          content: Text('שגיאה בשליחת הודעה: $e'),
+          content: Text(l10n.errorSendingMessage(e.toString())),
           backgroundColor: Colors.red,
         ),
       );
@@ -1095,20 +1667,22 @@ color: message.isDeleted
     final result = await showDialog<bool>(
       context: context,
       builder: (context) => AlertDialog(
-        title: const Text('מחיקת צ\'אט'),
-        content: const Text('האם אתה בטוח שברצונך למחוק את הצ\'אט? פעולה זו לא ניתנת לביטול.'),
+        title: Text(l10n.deleteChatTitle),
+        content: Text(l10n.deleteChatConfirm),
         actions: [
           TextButton(
             onPressed: () => Navigator.of(context).pop(false),
-            child: const Text('ביטול'),
+            child: Text(l10n.cancel),
           ),
           ElevatedButton(
             onPressed: () async {
               await playButtonSound();
+              // Guard context usage after async gap - check context.mounted for builder context
+              if (!context.mounted) return;
               Navigator.of(context).pop(true);
             },
             style: ElevatedButton.styleFrom(
-              backgroundColor: Colors.red[600],
+              backgroundColor: Theme.of(context).colorScheme.error,
               foregroundColor: Colors.white,
             ),
             child: const Text('מחק'),
@@ -1135,23 +1709,67 @@ color: message.isDeleted
       
       if (!chatDoc.exists) return;
       
+      final chatData = chatDoc.data()!;
+      final requestId = chatData['requestId'] as String?;
+      
+      // קבלת פרטי הבקשה כדי לבדוק מי יוצר הבקשה
+      final requestDoc = requestId != null 
+          ? await FirebaseFirestore.instance
+              .collection('requests')
+              .doc(requestId)
+              .get()
+          : null;
+      
+      final createdBy = requestDoc?.exists == true 
+          ? requestDoc!.data()!['createdBy'] as String?
+          : null;
+      
+      final isRequestCreator = createdBy == user.uid;
+      final isServiceProvider = !isRequestCreator;
+      
       // הוספת המשתמש לרשימת המחיקות
+      // אם מבקש השירות מוחק את הצ'אט, נסמן אותו כ"סגור" אצל נותן השירות
+      // אם נותן השירות מוחק את הצ'אט, לא נסמן אותו כ"סגור" (מבקש השירות עדיין יכול לראות אותו)
       await FirebaseFirestore.instance
           .collection('chats')
           .doc(widget.chatId)
           .update({
         'deletedBy': FieldValue.arrayUnion([user.uid]),
         'deletedAt': FieldValue.arrayUnion([DateTime.now()]),
+        'isClosed': isRequestCreator, // סגירת הצ'אט רק אם מבקש השירות מוחק אותו
         'updatedAt': DateTime.now(),
       });
 
-      debugPrint('✅ Chat ${widget.chatId} marked as deleted by user ${user.uid}');
+      if (isRequestCreator) {
+        debugPrint('✅ Chat ${widget.chatId} marked as deleted by request creator ${user.uid} and closed for service provider');
+      } else {
+        debugPrint('✅ Chat ${widget.chatId} marked as deleted by service provider ${user.uid} - request will remain in "My Inquiries"');
+      }
+      
+      // אם יש requestId, עדכן את הבקשה
+      // אם נותן השירות מוחק את הצ'אט, הבקשה תישאר ב"פניות שלי" שלו (לא נסיר אותו מ-helpers)
+      // אם מבקש השירות מוחק את הצ'אט, הבקשה תישאר ב"פניות שלי" של נותן השירות (לא נסיר אותו מ-helpers)
+      if (requestId != null && requestDoc?.exists == true) {
+        try {
+          // לא נסיר את נותן השירות מ-helpers גם אם הוא מוחק את הצ'אט
+          // הבקשה תישאר ב"פניות שלי" שלו
+          if (isServiceProvider) {
+            debugPrint('ℹ️ Service provider deleted chat - request will remain in "My Inquiries" for service provider');
+          } else {
+            debugPrint('ℹ️ Request creator deleted chat - request will remain in "My Inquiries" for service provider');
+          }
+        } catch (e) {
+          debugPrint('❌ Error updating request after chat deletion: $e');
+          // נמשיך גם אם יש שגיאה בעדכון הבקשה
+        }
+      }
       
       if (!mounted) return;
+      final l10n = AppLocalizations.of(context);
       
       ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(
-          content: Text('הצ\'אט נמחק בהצלחה'),
+        SnackBar(
+          content: Text(l10n.chatDeletedSuccessfully),
           backgroundColor: Colors.green,
         ),
       );
@@ -1162,10 +1780,11 @@ color: message.isDeleted
     } catch (e) {
       debugPrint('❌ Error deleting chat: $e');
       if (!mounted) return;
+      final l10n = AppLocalizations.of(context);
       
       ScaffoldMessenger.of(context).showSnackBar(
         SnackBar(
-          content: Text('שגיאה במחיקת הצ\'אט: $e'),
+          content: Text(l10n.errorDeletingChat(e.toString())),
           backgroundColor: Colors.red,
         ),
       );
@@ -1196,12 +1815,12 @@ color: message.isDeleted
     showDialog(
       context: context,
       builder: (context) => AlertDialog(
-        title: const Text('סגירת צ\'אט'),
-        content: const Text('האם אתה בטוח שברצונך לסגור את הצ\'אט? לאחר הסגירה לא ניתן יהיה לשלוח הודעות נוספות.'),
+        title: Text(l10n.closeChatTitle),
+        content: Text(l10n.closeChatMessage),
         actions: [
           TextButton(
             onPressed: () => Navigator.of(context).pop(),
-            child: const Text('ביטול'),
+            child: Text(l10n.cancel),
           ),
           TextButton(
             onPressed: () {
@@ -1209,7 +1828,7 @@ color: message.isDeleted
               _closeChat();
             },
             style: TextButton.styleFrom(foregroundColor: Colors.orange),
-            child: const Text('סגור צ\'אט'),
+            child: Text(l10n.closeChat),
           ),
         ],
       ),
@@ -1221,6 +1840,9 @@ color: message.isDeleted
       final user = FirebaseAuth.instance.currentUser;
       if (user == null) return;
 
+      final l10n = AppLocalizations.of(context);
+      final userName = user.displayName ?? user.email?.split('@')[0] ?? l10n.otherUser;
+      
       // שליחת הודעת מערכת
       await FirebaseFirestore.instance
           .collection('chats')
@@ -1228,7 +1850,7 @@ color: message.isDeleted
           .collection('messages')
           .add({
         'from': 'system',
-        'text': 'הצ\'אט נסגר על ידי ${user.displayName ?? user.email?.split('@')[0] ?? 'משתמש'}. לא ניתן לשלוח הודעות נוספות.',
+        'text': l10n.chatClosedBy(userName),
         'sentAt': Timestamp.fromDate(DateTime.now()),
         'isSystemMessage': true,
         'messageType': 'chat_closed',
@@ -1242,14 +1864,14 @@ color: message.isDeleted
         'isClosed': true,
         'closedAt': Timestamp.fromDate(DateTime.now()),
         'closedBy': user.uid,
-        'lastMessage': 'הצ\'אט נסגר',
+        'lastMessage': l10n.chatClosedStatus,
         'updatedAt': Timestamp.fromDate(DateTime.now()),
       });
 
       if (mounted) {
         ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(
-            content: Text('הצ\'אט נסגר בהצלחה'),
+          SnackBar(
+            content: Text(l10n.chatClosedSuccessfully),
             backgroundColor: Colors.orange,
           ),
         );
@@ -1257,9 +1879,10 @@ color: message.isDeleted
     } catch (e) {
       debugPrint('Error closing chat: $e');
       if (mounted) {
+        final l10n = AppLocalizations.of(context);
         ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(
-            content: Text('שגיאה בסגירת הצ\'אט'),
+          SnackBar(
+            content: Text(l10n.errorClosingChat),
             backgroundColor: Colors.red,
           ),
         );
@@ -1272,6 +1895,35 @@ color: message.isDeleted
       final user = FirebaseAuth.instance.currentUser;
       if (user == null) return;
 
+      final l10n = AppLocalizations.of(context);
+      
+      // בדיקה אם הצ'אט נמחק על ידי מבקש השירות או נותן השירות
+      if (_isDeletedByRequestCreator) {
+        if (mounted) {
+          ScaffoldMessenger.of(context).showSnackBar(
+            SnackBar(
+              content: Text(l10n.cannotReopenChatDeletedByRequester),
+              backgroundColor: Colors.red,
+            ),
+          );
+        }
+        return;
+      }
+      
+      if (_isDeletedByServiceProvider) {
+        if (mounted) {
+          ScaffoldMessenger.of(context).showSnackBar(
+            SnackBar(
+              content: Text(l10n.cannotReopenChatDeletedByProvider),
+              backgroundColor: Colors.red,
+            ),
+          );
+        }
+        return;
+      }
+      
+      final userName = user.displayName ?? user.email?.split('@')[0] ?? l10n.otherUser;
+      
       // עדכון הצ'אט כפתוח
       await FirebaseFirestore.instance
           .collection('chats')
@@ -1280,7 +1932,7 @@ color: message.isDeleted
         'isClosed': false,
         'reopenedAt': Timestamp.fromDate(DateTime.now()),
         'reopenedBy': user.uid,
-        'lastMessage': 'הצ\'אט נפתח מחדש',
+        'lastMessage': l10n.chatReopened,
         'updatedAt': Timestamp.fromDate(DateTime.now()),
       });
 
@@ -1291,7 +1943,7 @@ color: message.isDeleted
           .collection('messages')
           .add({
         'from': 'system',
-        'text': 'הצ\'אט נפתח מחדש על ידי ${user.displayName ?? user.email?.split('@')[0] ?? 'משתמש'}.',
+        'text': l10n.chatReopenedBy(userName),
         'sentAt': Timestamp.fromDate(DateTime.now()),
         'isSystemMessage': true,
         'messageType': 'chat_reopened',
@@ -1299,8 +1951,8 @@ color: message.isDeleted
 
       if (mounted) {
         ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(
-            content: Text('הצ\'אט נפתח מחדש בהצלחה'),
+          SnackBar(
+            content: Text(l10n.chatReopened),
             backgroundColor: Colors.green,
           ),
         );
@@ -1308,9 +1960,10 @@ color: message.isDeleted
     } catch (e) {
       debugPrint('Error reopening chat: $e');
       if (mounted) {
+        final l10n = AppLocalizations.of(context);
         ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(
-            content: Text('שגיאה בפתיחת הצ\'אט מחדש'),
+          SnackBar(
+            content: Text(l10n.errorGeneral(e.toString())),
             backgroundColor: Colors.red,
           ),
         );
@@ -1323,7 +1976,7 @@ color: message.isDeleted
       context: context,
       builder: (context) => AlertDialog(
         title: Text(l10n.clearChat),
-        content: const Text('האם אתה בטוח שברצונך למחוק את ההודעות שלך?\nהמשתמש השני ימשיך לראות את ההודעות שלו.'),
+        content: Text(l10n.deleteMyMessagesConfirm),
         actions: [
           TextButton(
             onPressed: () => Navigator.pop(context),
@@ -1371,17 +2024,19 @@ color: message.isDeleted
       }
 
       if (!mounted) return;
+      final l10n = AppLocalizations.of(context);
       ScaffoldMessenger.of(context).showSnackBar(
         SnackBar(
-          content: Text('ההודעות שלך נמחקו בהצלחה'),
+          content: Text(l10n.myMessagesDeletedSuccessfully),
           backgroundColor: Colors.green,
         ),
       );
     } catch (e) {
       if (!mounted) return;
+      final l10n = AppLocalizations.of(context);
       ScaffoldMessenger.of(context).showSnackBar(
         SnackBar(
-          content: Text('שגיאה במחיקת ההודעות: $e'),
+          content: Text(l10n.errorDeletingMyMessages(e.toString())),
           backgroundColor: Colors.red,
         ),
       );
