@@ -4,12 +4,14 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:permission_handler/permission_handler.dart';
+import 'package:cached_network_image/cached_network_image.dart';
 import 'dart:io';
 import '../models/request.dart';
 import '../l10n/app_localizations.dart';
 import '../widgets/phone_input_widget.dart';
 import '../widgets/two_level_category_selector.dart';
 import 'location_picker_screen.dart';
+import '../services/app_sharing_service.dart';
 
 class EditRequestScreen extends StatefulWidget {
   final Request request;
@@ -28,7 +30,7 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
   String _selectedPhonePrefix = '';
   String _selectedPhoneNumber = '';
   
-  RequestCategory _selectedCategory = RequestCategory.maintenance;
+  RequestCategory _selectedCategory = RequestCategory.plumbing;
   RequestLocation _selectedLocation = RequestLocation.custom;
   final List<String> _selectedImages = [];
   final List<File> _selectedImageFiles = [];
@@ -41,6 +43,17 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
   double? _minRating;
   List<RequestCategory> _selectedTargetCategories = [];
   
+  // מחיר (אופציונאלי) - רק לבקשות בתשלום
+  final _priceController = TextEditingController();
+  double? _price;
+  
+  // דירוגים מינימליים מפורטים
+  double? _minReliability;
+  double? _minAvailability;
+  double? _minAttitude;
+  double? _minFairPrice;
+  bool _useDetailedRatings = false; // האם להשתמש בדירוגים מפורטים
+  
   // שדות דחיפות חדשים
   UrgencyLevel _selectedUrgency = UrgencyLevel.normal;
   final List<RequestTag> _selectedTags = [];
@@ -51,6 +64,12 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
   double? _selectedLongitude;
   String? _selectedAddress;
   double? _exposureRadius; // רדיוס חשיפה בקילומטרים
+  bool? _showToProvidersOutsideRange; // null = לא נבחר, true = כן, false = לא
+  bool _showToProvidersOutsideRangeError = false; // האם להציג שגיאה על השדה
+  
+  // האם להציג בקשה לכל המשתמשים או רק לנותני שירות מתחום X
+  bool? _showToAllUsers; // null = לא נבחר, true = לכל המשתמשים, false = רק לנותני שירות מתחום X
+  bool _showToAllUsersError = false; // האם להציג שגיאה על השדה
 
   @override
   void initState() {
@@ -76,6 +95,18 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
     _selectedDeadline = widget.request.deadline;
     _minRating = widget.request.minRating;
     _selectedTargetCategories = widget.request.targetCategories ?? [];
+    _price = widget.request.price;
+    if (_price != null) {
+      _priceController.text = _price!.toStringAsFixed(0);
+    }
+    
+    // דירוגים מפורטים
+    _minReliability = widget.request.minReliability;
+    _minAvailability = widget.request.minAvailability;
+    _minAttitude = widget.request.minAttitude;
+    _minFairPrice = widget.request.minFairPrice;
+    _useDetailedRatings = _minReliability != null || _minAvailability != null || 
+                         _minAttitude != null || _minFairPrice != null;
     
     // שדות דחיפות חדשים
     _selectedUrgency = widget.request.urgencyLevel;
@@ -87,12 +118,15 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
     _selectedLongitude = widget.request.longitude;
     _selectedAddress = widget.request.address;
     _exposureRadius = widget.request.exposureRadius;
+    _showToProvidersOutsideRange = widget.request.showToProvidersOutsideRange;
+    _showToAllUsers = widget.request.showToAllUsers;
   }
 
   @override
   void dispose() {
     _titleController.dispose();
     _descriptionController.dispose();
+    _priceController.dispose();
     super.dispose();
   }
 
@@ -139,6 +173,7 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
   }
 
   Future<void> _pickImages() async {
+    final l10n = AppLocalizations.of(context);
     try {
       // בדיקת הרשאות
       PermissionStatus permission = PermissionStatus.denied;
@@ -180,8 +215,8 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
       if (permission != PermissionStatus.granted) {
         if (mounted) {
           ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(
-              content: Text('נדרשת הרשאת גישה לתמונות'),
+            SnackBar(
+              content: Text(l10n.imageAccessPermissionRequired),
               duration: Duration(seconds: 2),
             ),
           );
@@ -208,7 +243,7 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
       if (mounted) {
         ScaffoldMessenger.of(context).showSnackBar(
           SnackBar(
-            content: Text('שגיאה בבחירת תמונות: $e'),
+            content: Text('${l10n.errorSelectingImages}: $e'),
             duration: const Duration(seconds: 2),
           ),
         );
@@ -217,6 +252,7 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
   }
 
   Future<void> _takePhoto() async {
+    final l10n = AppLocalizations.of(context);
     try {
       // בדיקת הרשאות מצלמה
       PermissionStatus permission = await Permission.camera.status;
@@ -227,8 +263,8 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
       if (permission != PermissionStatus.granted) {
         if (mounted) {
           ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(
-              content: Text('נדרשת הרשאת גישה למצלמה'),
+            SnackBar(
+              content: Text(l10n.cameraAccessPermissionRequired),
               duration: Duration(seconds: 2),
             ),
           );
@@ -252,8 +288,8 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
         // הצגת הודעה על הצלחה
         if (mounted) {
           ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(
-              content: Text('תמונה נוספה בהצלחה'),
+            SnackBar(
+              content: Text(l10n.imageAddedSuccessfully),
               duration: Duration(seconds: 1),
               backgroundColor: Colors.green,
             ),
@@ -265,7 +301,7 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
       if (mounted) {
         ScaffoldMessenger.of(context).showSnackBar(
           SnackBar(
-            content: Text('שגיאה בצילום תמונה: $e'),
+            content: Text('${l10n.errorTakingPhoto}: $e'),
             duration: const Duration(seconds: 2),
           ),
         );
@@ -274,6 +310,7 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
   }
 
   Future<void> _uploadImages() async {
+    final l10n = AppLocalizations.of(context);
     if (_selectedImageFiles.isEmpty) {
       debugPrint('No images to upload');
       return;
@@ -326,7 +363,7 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
       if (mounted) {
         ScaffoldMessenger.of(context).showSnackBar(
           SnackBar(
-            content: Text('שגיאה בהעלאת תמונות: $e'),
+            content: Text('${l10n.errorUploadingImages}: $e'),
             duration: const Duration(seconds: 2),
           ),
         );
@@ -337,6 +374,7 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
 
   /// מחיקת תמונות מ-Firebase Storage
   Future<void> _deleteImagesFromStorage(List<String> imageUrls) async {
+    final l10n = AppLocalizations.of(context);
     try {
       debugPrint('🗑️ Starting to delete ${imageUrls.length} images from Storage');
       
@@ -360,7 +398,7 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
       if (mounted) {
         ScaffoldMessenger.of(context).showSnackBar(
           SnackBar(
-            content: Text('נמחקו $deletedCount תמונות מ-Storage'),
+            content: Text(l10n.imagesDeletedFromStorage(deletedCount)),
             backgroundColor: Colors.blue,
             duration: const Duration(seconds: 2),
           ),
@@ -371,7 +409,7 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
       if (mounted) {
         ScaffoldMessenger.of(context).showSnackBar(
           SnackBar(
-            content: Text('שגיאה במחיקת תמונות: $e'),
+            content: Text(l10n.errorDeletingImages(e.toString())),
             backgroundColor: Colors.orange,
             duration: const Duration(seconds: 2),
           ),
@@ -381,8 +419,38 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
   }
 
   Future<void> _updateRequest() async {
-    if (!_formKey.currentState!.validate()) return;
-    if (_isLoading) return;
+    final l10n = AppLocalizations.of(context);
+    debugPrint('🔍 _updateRequest called');
+    debugPrint('🔍 Form key: $_formKey');
+    debugPrint('🔍 Form key current state: ${_formKey.currentState}');
+    debugPrint('🔍 Form validation: ${_formKey.currentState?.validate()}');
+    debugPrint('🔍 Is loading: $_isLoading');
+    debugPrint('🔍 Title: ${_titleController.text}');
+    debugPrint('🔍 Description: ${_descriptionController.text}');
+    debugPrint('🔍 Selected category: ${_selectedCategory.name}');
+    debugPrint('🔍 Selected location: ${_selectedLocation.name}');
+    debugPrint('🔍 Selected latitude: $_selectedLatitude');
+    debugPrint('🔍 Selected longitude: $_selectedLongitude');
+    
+    // בדיקת שדות ידנית במקום Form validation
+    if (_titleController.text.trim().isEmpty) {
+      debugPrint('❌ Title is empty');
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(
+          content: Text(l10n.pleaseEnterTitle),
+          backgroundColor: Colors.red,
+        ),
+      );
+      return;
+    }
+    
+    // ✅ השדה "תיאור" הוא אופציונאלי - אין בדיקה
+    
+    debugPrint('✅ Manual validation passed');
+    if (_isLoading) {
+      debugPrint('❌ Already loading');
+      return;
+    }
 
     setState(() => _isLoading = true);
 
@@ -391,8 +459,8 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
       if (user == null) {
         if (!mounted) return;
         ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(
-            content: Text('משתמש לא מחובר'),
+          SnackBar(
+            content: Text(l10n.userNotLoggedIn),
             backgroundColor: Colors.red,
           ),
         );
@@ -402,8 +470,36 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
       // בדיקת מיקום וטווח חשיפה
       if (_selectedLatitude == null || _selectedLongitude == null) {
         ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(
-            content: Text('נא לבחור מיקום לבקשה'),
+          SnackBar(
+            content: Text(l10n.pleaseSelectLocationForRequest),
+            backgroundColor: Colors.red,
+          ),
+        );
+        return;
+      }
+      
+      // ✅ בדיקה שהשדה showToProvidersOutsideRange נבחר (חובה)
+      if (_showToProvidersOutsideRange == null) {
+        setState(() {
+          _showToProvidersOutsideRangeError = true; // הצגת שגיאה על השדה
+        });
+        ScaffoldMessenger.of(context).showSnackBar(
+          SnackBar(
+            content: const Text('אנא בחר האם להציג את הבקשה לנותני שירות מחוץ לטווח שהגדרת'),
+            backgroundColor: Colors.red,
+          ),
+        );
+        return;
+      }
+      
+      // ✅ בדיקה שהשדה showToAllUsers נבחר (חובה אם יש קטגוריה)
+      if (_showToAllUsers == null) {
+        setState(() {
+          _showToAllUsersError = true; // הצגת שגיאה על השדה
+        });
+        ScaffoldMessenger.of(context).showSnackBar(
+          SnackBar(
+            content: const Text('אנא בחר האם להציג את הבקשה לכל נותני השירות מכל התחומים או רק לנותני שירות מתחום שבחרת'),
             backgroundColor: Colors.red,
           ),
         );
@@ -441,6 +537,10 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
         'type': _selectedType.name,
         'deadline': _selectedDeadline != null ? Timestamp.fromDate(_selectedDeadline!) : null,
         'minRating': _minRating,
+        'minReliability': _minReliability,
+        'minAvailability': _minAvailability,
+        'minAttitude': _minAttitude,
+        'minFairPrice': _minFairPrice,
         'targetAudience': TargetAudience.all.name,
         'maxDistance': null,
         'targetVillage': null,
@@ -449,6 +549,9 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
         'longitude': _selectedLongitude,
         'address': _selectedAddress,
         'exposureRadius': _exposureRadius,
+        'price': _price, // מחיר (אופציונאלי) - רק לבקשות בתשלום
+        'showToProvidersOutsideRange': _showToProvidersOutsideRange,
+        'showToAllUsers': _showToAllUsers, // null = לא נבחר, true = לכל המשתמשים, false = רק לנותני שירות מתחום X
         'updatedAt': FieldValue.serverTimestamp(),
       };
 
@@ -471,7 +574,7 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
       if (!mounted) return;
       ScaffoldMessenger.of(context).showSnackBar(
         SnackBar(
-          content: Text('שגיאה בעדכון: $e'),
+            content: Text(l10n.errorUpdatingRequest(e.toString())),
           backgroundColor: Colors.red,
         ),
       );
@@ -519,24 +622,157 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
         });
         
         if (mounted) {
+          final l10n = AppLocalizations.of(context);
           ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(
-              content: Text('כל התמונות נמחקו בהצלחה'),
+            SnackBar(
+              content: Text(l10n.allImagesDeletedSuccessfully),
               backgroundColor: Colors.green,
             ),
           );
         }
       } catch (e) {
         if (mounted) {
+          final l10n = AppLocalizations.of(context);
           ScaffoldMessenger.of(context).showSnackBar(
             SnackBar(
-              content: Text('שגיאה במחיקת תמונות: $e'),
+              content: Text(l10n.errorDeletingImages(e.toString())),
               backgroundColor: Colors.red,
             ),
           );
         }
       }
     }
+  }
+
+  // פונקציה לבניית שדה דירוג מפורט עם סליידר
+  Widget _buildDetailedRatingField(
+    String title,
+    String description,
+    double? currentValue,
+    Function(double?) onChanged,
+    IconData icon,
+    Color color,
+  ) {
+    // אם אין ערך, נגדיר 0.0
+    final ratingValue = currentValue ?? 0.0;
+    
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Row(
+          children: [
+            Icon(icon, color: color, size: 18),
+            const SizedBox(width: 8),
+            Text(
+              title,
+              style: TextStyle(
+                fontSize: 14,
+                fontWeight: FontWeight.w600,
+                color: color,
+              ),
+            ),
+            const Spacer(),
+            Text(
+              '${ratingValue.toStringAsFixed(1)} ⭐',
+              style: TextStyle(
+                fontSize: 12,
+                fontWeight: FontWeight.bold,
+                color: color,
+              ),
+            ),
+          ],
+        ),
+        if (description.isNotEmpty) ...[
+          const SizedBox(height: 4),
+          Text(
+            description,
+            style: TextStyle(
+              fontSize: 12,
+              color: Theme.of(context).colorScheme.onSurfaceVariant,
+            ),
+          ),
+        ],
+        const SizedBox(height: 8),
+        SliderTheme(
+          data: SliderTheme.of(context).copyWith(
+            activeTrackColor: color,
+            inactiveTrackColor: color.withValues(alpha: 0.3),
+            thumbColor: color,
+            overlayColor: color.withValues(alpha: 0.2),
+            valueIndicatorColor: color,
+            valueIndicatorTextStyle: const TextStyle(
+              color: Colors.white,
+              fontSize: 12,
+              fontWeight: FontWeight.bold,
+            ),
+          ),
+          child: Slider(
+            value: ratingValue,
+            min: 0.0,
+            max: 5.0,
+            divisions: 50, // 0.1 increments
+            label: '${ratingValue.toStringAsFixed(1)} ⭐',
+            onChanged: (value) {
+              onChanged(value);
+            },
+          ),
+        ),
+        const SizedBox(height: 8),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            Text(
+              '0.0 ⭐',
+              style: TextStyle(
+                fontSize: 10,
+                color: Theme.of(context).colorScheme.onSurfaceVariant,
+              ),
+            ),
+            Text(
+              '5.0 ⭐',
+              style: TextStyle(
+                fontSize: 10,
+                color: Theme.of(context).colorScheme.onSurfaceVariant,
+              ),
+            ),
+          ],
+        ),
+      ],
+    );
+  }
+
+  // פונקציה לבניית כפתור דחיפות
+  Widget _buildUrgencyButton(UrgencyLevel level, String label) {
+    final isSelected = _selectedUrgency == level;
+    return GestureDetector(
+      onTap: () {
+        setState(() {
+          _selectedUrgency = level;
+        });
+      },
+      child: Container(
+        padding: const EdgeInsets.symmetric(vertical: 12, horizontal: 8),
+        decoration: BoxDecoration(
+          color: isSelected ? level.color : Theme.of(context).colorScheme.surfaceContainer,
+          borderRadius: BorderRadius.circular(8),
+          border: Border.all(
+            color: isSelected ? level.color : Theme.of(context).colorScheme.outlineVariant,
+            width: 2,
+          ),
+        ),
+        child: Text(
+          label,
+          textAlign: TextAlign.center,
+          style: TextStyle(
+            color: isSelected 
+                ? Colors.white 
+                : Theme.of(context).colorScheme.onSurface,
+            fontWeight: isSelected ? FontWeight.bold : FontWeight.normal,
+            fontSize: 12,
+          ),
+        ),
+      ),
+    );
   }
 
   @override
@@ -555,7 +791,7 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
             ),
           ),
           backgroundColor: Theme.of(context).brightness == Brightness.dark 
-              ? const Color(0xFFFF9800) // כתום ענתיק
+              ? const Color(0xFF9C27B0) // סגול יפה
               : Theme.of(context).colorScheme.primary,
           foregroundColor: Colors.white,
           toolbarHeight: 50,
@@ -584,7 +820,7 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                         borderRadius: BorderRadius.circular(20),
                         boxShadow: [
                           BoxShadow(
-                            color: Colors.black.withOpacity(0.1),
+                            color: Colors.black.withValues(alpha: 0.1),
                             blurRadius: 10,
                             offset: const Offset(0, 5),
                           ),
@@ -605,11 +841,11 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                           ),
                           const SizedBox(height: 16),
                           Text(
-                            'מעדכן בקשה...',
+                            l10n.updatingRequest,
                             style: TextStyle(
                               fontSize: 16,
                               fontWeight: FontWeight.w500,
-                              color: Colors.grey[700],
+                              color: Theme.of(context).colorScheme.onSurfaceVariant,
                             ),
                           ),
                         ],
@@ -635,8 +871,8 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                         child: TwoLevelCategorySelector(
                           selectedCategories: [_selectedCategory],
                           maxSelections: 1,
-                          title: 'בחירת קטגוריה',
-                          instruction: 'בחר תחום ראשי ואז תחום משנה:',
+                          title: l10n.selectCategory,
+                          instruction: l10n.selectMainCategoryThenSubcategory,
                           onSelectionChanged: (categories) {
                             if (categories.isNotEmpty) {
                               setState(() {
@@ -651,14 +887,17 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                       // כותרת
                       TextFormField(
                         controller: _titleController,
-                        decoration: const InputDecoration(
-                          labelText: 'כותרת',
+                        decoration: InputDecoration(
+                          labelText: l10n.title,
                           border: OutlineInputBorder(),
                         ),
                         validator: (value) {
+                          debugPrint('🔍 Title validation: value="$value"');
                           if (value == null || value.isEmpty) {
-                            return 'אנא הזן כותרת';
+                            debugPrint('❌ Title validation failed: empty');
+                            return l10n.pleaseEnterTitle;
                           }
+                          debugPrint('✅ Title validation passed');
                           return null;
                         },
                       ),
@@ -667,19 +906,90 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                       // תיאור
                       TextFormField(
                         controller: _descriptionController,
-                        decoration: const InputDecoration(
-                          labelText: 'תיאור',
+                        decoration: InputDecoration(
+                          labelText: l10n.description,
                           border: OutlineInputBorder(),
                         ),
                         maxLines: 3,
-                        validator: (value) {
-                          if (value == null || value.isEmpty) {
-                            return 'אנא הזן תיאור';
-                          }
-                          return null;
-                        },
+                        // ✅ השדה "תיאור" הוא אופציונאלי - אין וולידציה
                       ),
                       const SizedBox(height: 16),
+                      
+                      // ✅ שאלה: האם להציג לכל נותני השירות מכל התחומים או רק לנותני שירות מתחום X
+                      // מופיע רק אם נבחרה קטגוריה
+                      ...[
+                        Builder(
+                          builder: (context) {
+                            final l10n = AppLocalizations.of(context);
+                            final categoryName = _selectedCategory.categoryDisplayName;
+                            
+                            return Column(
+                              crossAxisAlignment: CrossAxisAlignment.start,
+                              children: [
+                                Text(
+                                  l10n.showToAllUsersOrProviders(categoryName),
+                                  style: TextStyle(
+                                    fontSize: 14,
+                                    fontWeight: FontWeight.w500,
+                                    color: _showToAllUsersError 
+                                        ? Theme.of(context).colorScheme.error
+                                        : (Theme.of(context).brightness == Brightness.dark 
+                                            ? Colors.white 
+                                            : Colors.black87),
+                                  ),
+                                ),
+                                const SizedBox(height: 12),
+                                Row(
+                                  children: [
+                                    Radio<bool>(
+                                      value: true,
+                                      groupValue: _showToAllUsers,
+                                      onChanged: (bool? value) {
+                                        setState(() {
+                                          _showToAllUsers = value;
+                                          _showToAllUsersError = false; // הסרת שגיאה אחרי בחירה
+                                        });
+                                      },
+                                    ),
+                                    Expanded(
+                                      child: Text(l10n.yesToAllUsers),
+                                    ),
+                                    const SizedBox(width: 24),
+                                    Radio<bool>(
+                                      value: false,
+                                      groupValue: _showToAllUsers,
+                                      onChanged: (bool? value) async {
+                                        setState(() {
+                                          _showToAllUsers = value;
+                                          _showToAllUsersError = false; // הסרת שגיאה אחרי בחירה
+                                        });
+                                        // ✅ הצגת דיאלוג עם מספר נותני שירות כשמשתמש בוחר "רק לנותני שירות מתחום X"
+                                        if (value == false) {
+                                          await _checkAvailableHelpers();
+                                        }
+                                      },
+                                    ),
+                                    Expanded(
+                                      child: Text(l10n.onlyToProvidersInCategory(categoryName)),
+                                    ),
+                                  ],
+                                ),
+                                if (_showToAllUsersError) ...[
+                                  const SizedBox(height: 8),
+                                  Text(
+                                    'אנא בחר האם להציג את הבקשה לכל נותני השירות מכל התחומים או רק לנותני שירות מתחום $categoryName',
+                                    style: TextStyle(
+                                      fontSize: 12,
+                                      color: Theme.of(context).colorScheme.error,
+                                    ),
+                                  ),
+                                ],
+                              ],
+                            );
+                          },
+                        ),
+                        const SizedBox(height: 16),
+                      ],
 
                       // תמונות
                       Card(
@@ -699,7 +1009,7 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                                     child: ElevatedButton.icon(
                                     onPressed: _pickImages,
                                       icon: const Icon(Icons.photo_library),
-                                      label: const Text('בחר תמונות'),
+                                      label: Text(l10n.selectImages),
                                     style: ElevatedButton.styleFrom(
                                         backgroundColor: const Color(0xFF03A9F4),
                                       foregroundColor: Colors.white,
@@ -711,7 +1021,7 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                                     child: ElevatedButton.icon(
                                       onPressed: _takePhoto,
                                       icon: const Icon(Icons.camera_alt),
-                                      label: const Text('צלם תמונה'),
+                                      label: Text(l10n.takePhoto),
                                       style: ElevatedButton.styleFrom(
                                         backgroundColor: const Color(0xFFE91E63),
                                         foregroundColor: Colors.white,
@@ -727,7 +1037,7 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                                     child: ElevatedButton.icon(
                                     onPressed: _clearImages,
                                     icon: const Icon(Icons.clear),
-                                      label: const Text('נקה הכל'),
+                                      label: Text(l10n.clearAll),
                                     style: ElevatedButton.styleFrom(
                                         backgroundColor: Colors.red,
                                       foregroundColor: Colors.white,
@@ -739,8 +1049,8 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                               if (_selectedImages.isNotEmpty || _selectedImageFiles.isNotEmpty) ...[
                                 const SizedBox(height: 8),
                                 Text(
-                                  '${_selectedImages.length + _selectedImageFiles.length} תמונות נבחרו',
-                                  style: TextStyle(color: Colors.grey[600]),
+                                  l10n.imagesSelected(_selectedImages.length + _selectedImageFiles.length),
+                                  style: TextStyle(color: Theme.of(context).colorScheme.onSurfaceVariant),
                                 ),
                                 const SizedBox(height: 8),
                                 SizedBox(
@@ -792,9 +1102,9 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                         }(),
                         builder: (context, snapshot) {
                           return DropdownButtonFormField<RequestType>(
-                            value: _selectedType,
-                            decoration: const InputDecoration(
-                              labelText: 'סוג בקשה',
+                            initialValue: _selectedType,
+                            decoration: InputDecoration(
+                              labelText: l10n.requestType,
                               border: OutlineInputBorder(),
                               prefixIcon: Icon(Icons.payment),
                             ),
@@ -812,21 +1122,132 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                           );
                         },
                       ),
+                      // שדה מחיר (רק אם סוג הבקשה הוא בתשלום)
+                      if (_selectedType == RequestType.paid) ...[
+                        const SizedBox(height: 16),
+                        TextFormField(
+                          controller: _priceController,
+                          decoration: InputDecoration(
+                            labelText: l10n.howMuchWillingToPay,
+                            hintText: 'לדוגמה: 100',
+                            border: const OutlineInputBorder(),
+                            suffixText: '₪',
+                          ),
+                          keyboardType: TextInputType.numberWithOptions(decimal: true),
+                          onChanged: (value) {
+                            if (value.isEmpty) {
+                              setState(() {
+                                _price = null;
+                              });
+                            } else {
+                              final parsedPrice = double.tryParse(value);
+                              setState(() {
+                                _price = parsedPrice;
+                              });
+                            }
+                          },
+                          validator: (value) {
+                            if (value != null && value.isNotEmpty) {
+                              final parsedPrice = double.tryParse(value);
+                              if (parsedPrice == null || parsedPrice < 0) {
+                                return 'אנא הזן מחיר תקין';
+                              }
+                            }
+                            return null;
+                          },
+                        ),
+                      ],
                       const SizedBox(height: 16),
 
                       // בחירת מיקום
                       Card(
                         child: ListTile(
                           leading: const Icon(Icons.location_on),
-                          title: Text(_selectedAddress ?? 'בחר מיקום'),
+                          title: Text(_selectedAddress ?? l10n.selectLocation),
                           subtitle: _selectedLatitude != null 
                               ? Text('${_selectedLatitude!.toStringAsFixed(4)}, ${_selectedLongitude!.toStringAsFixed(4)}${_exposureRadius != null ? ' • רדיוס: ${_exposureRadius!.toStringAsFixed(1)} ק"מ' : ''}')
-                              : const Text('לחץ לבחירת מיקום'),
+                              : Text(l10n.clickToSelectLocation),
                           trailing: const Icon(Icons.arrow_forward_ios),
                           onTap: _selectLocation,
                         ),
                       ),
                       const SizedBox(height: 16),
+                      
+                      // ✅ האם להציג בקשה לנותני שירות שלא בטווח שהגדרת
+                      // מופיע רק אם נבחר מיקום
+                      if (_selectedLatitude != null && _selectedLongitude != null) ...[
+                        Builder(
+                          builder: (context) {
+                            final l10n = AppLocalizations.of(context);
+                            // זיהוי האיזור לפי קו רוחב
+                            final region = getGeographicRegion(_selectedLatitude);
+                            final regionName = region.getDisplayName(l10n);
+                            // אם יש קטגוריה נבחרת, נציג את שמה, אחרת נציג "התחום שבחרת"
+                            final categoryName = _selectedCategory.categoryDisplayName;
+                            
+                            return Column(
+                              crossAxisAlignment: CrossAxisAlignment.start,
+                              children: [
+                                Text(
+                                  l10n.showToProvidersOutsideRange(regionName, categoryName),
+                                  style: TextStyle(
+                                    fontSize: 14,
+                                    fontWeight: FontWeight.w500,
+                                    color: _showToProvidersOutsideRangeError 
+                                        ? Theme.of(context).colorScheme.error
+                                        : (Theme.of(context).brightness == Brightness.dark 
+                                            ? Colors.white 
+                                            : Colors.black87),
+                                  ),
+                                ),
+                                const SizedBox(height: 12),
+                                Row(
+                                  children: [
+                                    Radio<bool>(
+                                      value: true,
+                                      groupValue: _showToProvidersOutsideRange,
+                                      onChanged: (bool? value) {
+                                        setState(() {
+                                          _showToProvidersOutsideRange = value;
+                                          _showToProvidersOutsideRangeError = false; // הסרת שגיאה אחרי בחירה
+                                        });
+                                      },
+                                    ),
+                                    Expanded(
+                                      child: Text(l10n.yesAllProvidersInRegion(regionName)),
+                                    ),
+                                    const SizedBox(width: 24),
+                                    Radio<bool>(
+                                      value: false,
+                                      groupValue: _showToProvidersOutsideRange,
+                                      onChanged: (bool? value) {
+                                        setState(() {
+                                          _showToProvidersOutsideRange = value;
+                                          _showToProvidersOutsideRangeError = false; // הסרת שגיאה אחרי בחירה
+                                        });
+                                      },
+                                    ),
+                                    Expanded(
+                                      child: Text(l10n.noOnlyInRange),
+                                    ),
+                                  ],
+                                ),
+                                if (_showToProvidersOutsideRangeError) ...[
+                                  const SizedBox(height: 8),
+                                  Text(
+                                    'אנא בחר האם להציג את הבקשה לנותני שירות מחוץ לטווח שהגדרת',
+                                    style: TextStyle(
+                                      fontSize: 12,
+                                      color: Theme.of(context).colorScheme.error,
+                                    ),
+                                  ),
+                                ],
+                              ],
+                            );
+                          },
+                        ),
+                        const SizedBox(height: 16),
+                      ],
 
                       // תאריך יעד
                       Card(
@@ -834,8 +1255,8 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                           leading: const Icon(Icons.calendar_today),
                           title: Text(
                             _selectedDeadline != null
-                                ? 'תאריך יעד: ${_selectedDeadline!.day}/${_selectedDeadline!.month}/${_selectedDeadline!.year}'
-                                : 'תאריך יעד (אופציונלי)',
+                                ? l10n.deadlineDateSelected(_selectedDeadline!.day, _selectedDeadline!.month, _selectedDeadline!.year)
+                                : l10n.deadlineLabel,
                           ),
                           trailing: const Icon(Icons.arrow_forward_ios),
                           onTap: _selectDeadline,
@@ -844,79 +1265,40 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                       const SizedBox(height: 16),
 
                       // רמת דחיפות
-                      Card(
-                        child: Padding(
-                          padding: const EdgeInsets.all(16.0),
-                          child: Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text(
-                                'רמת דחיפות',
-                                style: Theme.of(context).textTheme.titleMedium,
+                      Container(
+                        decoration: BoxDecoration(
+                          border: Border.all(color: Colors.grey),
+                          borderRadius: BorderRadius.circular(8),
+                        ),
+                        padding: const EdgeInsets.all(16),
+                        child: Column(
+                          crossAxisAlignment: CrossAxisAlignment.start,
+                          children: [
+                            Text(
+                              l10n.urgencyLevel,
+                              style: TextStyle(
+                                fontSize: 16, 
+                                fontWeight: FontWeight.bold,
+                                color: Theme.of(context).colorScheme.onSurface,
                               ),
-                              const SizedBox(height: 12),
-                              Row(
-                                children: [
-                                  Expanded(
-                                    child: ElevatedButton(
-                                      onPressed: () {
-                                        setState(() {
-                                          _selectedUrgency = UrgencyLevel.normal;
-                                        });
-                                      },
-                                      style: ElevatedButton.styleFrom(
-                                        backgroundColor: _selectedUrgency == UrgencyLevel.normal 
-                                            ? Colors.green 
-                                            : Colors.grey[300],
-                                        foregroundColor: _selectedUrgency == UrgencyLevel.normal 
-                                            ? Colors.white 
-                                            : Colors.black87,
-                                      ),
-                                      child: const Text('רגיל'),
-                                    ),
-                                  ),
-                                  const SizedBox(width: 8),
-                                  Expanded(
-                                    child: ElevatedButton(
-                                      onPressed: () {
-                                        setState(() {
-                                          _selectedUrgency = UrgencyLevel.urgent24h;
-                                        });
-                                      },
-                                      style: ElevatedButton.styleFrom(
-                                        backgroundColor: _selectedUrgency == UrgencyLevel.urgent24h 
-                                            ? Colors.orange 
-                                            : Colors.grey[300],
-                                        foregroundColor: _selectedUrgency == UrgencyLevel.urgent24h 
-                                            ? Colors.white 
-                                            : Colors.black87,
-                                      ),
-                                      child: const Text('תוך 24 שעות'),
-                                    ),
-                                  ),
-                                  const SizedBox(width: 8),
-                                  Expanded(
-                                    child: ElevatedButton(
-                                      onPressed: () {
-                                        setState(() {
-                                          _selectedUrgency = UrgencyLevel.emergency;
-                                        });
-                                      },
-                                      style: ElevatedButton.styleFrom(
-                                        backgroundColor: _selectedUrgency == UrgencyLevel.emergency 
-                                            ? Colors.red 
-                                            : Colors.grey[300],
-                                        foregroundColor: _selectedUrgency == UrgencyLevel.emergency 
-                                            ? Colors.white 
-                                            : Colors.black87,
-                                      ),
-                                      child: const Text('עכשיו'),
-                                    ),
-                                  ),
-                                ],
-                              ),
-                            ],
-                          ),
+                            ),
+                            const SizedBox(height: 12),
+                            Row(
+                              children: [
+                                Expanded(
+                                  child: _buildUrgencyButton(UrgencyLevel.normal, '🕓 ${l10n.normal}'),
+                                ),
+                                const SizedBox(width: 8),
+                                Expanded(
+                                  child: _buildUrgencyButton(UrgencyLevel.urgent24h, '⏰ ${l10n.within24Hours}'),
+                                ),
+                                const SizedBox(width: 8),
+                                Expanded(
+                                  child: _buildUrgencyButton(UrgencyLevel.emergency, '🚨 ${l10n.now}'),
+                                ),
+                              ],
+                            ),
+                          ],
                         ),
                       ),
                       const SizedBox(height: 16),
@@ -929,14 +1311,14 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                             crossAxisAlignment: CrossAxisAlignment.start,
                             children: [
                               Text(
-                                'תגיות דחיפות',
+                                l10n.urgencyTags,
                                 style: Theme.of(context).textTheme.titleMedium,
                               ),
                               const SizedBox(height: 8),
                               Text(
-                                'בחר תגיות המתאימות לבקשה שלך:',
+                                l10n.selectTagsForRequest,
                                 style: TextStyle(
-                                  color: Colors.grey[600],
+                                  color: Theme.of(context).colorScheme.onSurfaceVariant,
                                   fontSize: 14,
                                 ),
                               ),
@@ -947,19 +1329,19 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                                 children: _getTagsForCategory(_selectedCategory).map((tag) {
                                   final isSelected = _selectedTags.contains(tag);
                                   return FilterChip(
-                                    label: Text(tag.displayName),
+                                    label: Text(tag.displayName(l10n)),
                                     selected: isSelected,
                                     onSelected: (selected) {
-                                      setState(() {
+                                  setState(() {
                                         if (selected) {
                                           _selectedTags.add(tag);
                                         } else {
                                           _selectedTags.remove(tag);
                                         }
-                                      });
-                                    },
-                                    backgroundColor: Colors.grey[200],
-                                    selectedColor: tag.color.withOpacity(0.3),
+                                  });
+                                },
+                                    backgroundColor: Theme.of(context).colorScheme.surfaceContainer,
+                                    selectedColor: tag.color.withValues(alpha: 0.3),
                                     checkmarkColor: tag.color,
                                     labelStyle: TextStyle(
                                       color: isSelected ? tag.color : Colors.black87,
@@ -968,23 +1350,11 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                                   );
                                 }).toList(),
                               ),
-                              const SizedBox(height: 12),
-                              TextFormField(
-                                initialValue: _customTag,
-                                decoration: const InputDecoration(
-                                  labelText: 'תגית מותאמת אישית',
-                                  hintText: 'כתוב תגית קצרה משלך',
-                                  border: OutlineInputBorder(),
-                                ),
-                                onChanged: (value) {
-                                  _customTag = value;
-                                },
-                              ),
                             ],
                           ),
                         ),
-                      ),
-                      const SizedBox(height: 16),
+                        ),
+                        const SizedBox(height: 16),
 
                       // דירוג מינימלי (רק לבקשות בתשלום)
                       if (_selectedType == RequestType.paid) ...[
@@ -995,7 +1365,7 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                               crossAxisAlignment: CrossAxisAlignment.start,
                               children: [
                                 Text(
-                                  'דירוג מינימלי של עוזרים',
+                                  l10n.minRatingForHelpers,
                                   style: Theme.of(context).textTheme.titleMedium,
                                 ),
                                 const SizedBox(height: 12),
@@ -1006,17 +1376,23 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                                         onPressed: () {
                                           setState(() {
                                             _minRating = null;
+                                            _useDetailedRatings = false;
+                                            // איפוס כל השדות המפורטים
+                                            _minReliability = null;
+                                            _minAvailability = null;
+                                            _minAttitude = null;
+                                            _minFairPrice = null;
                                           });
                                         },
                                         style: ElevatedButton.styleFrom(
                                           backgroundColor: _minRating == null 
                                               ? Colors.blue 
-                                              : Colors.grey[300],
+                                              : Theme.of(context).colorScheme.surfaceContainerHighest,
                                           foregroundColor: _minRating == null 
                                               ? Colors.white 
                                               : Colors.black87,
                                         ),
-                                        child: const Text('כל הדירוגים'),
+                                        child: Text(l10n.allRatings),
                                       ),
                                     ),
                                     const SizedBox(width: 8),
@@ -1024,30 +1400,67 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                                       child: ElevatedButton(
                                         onPressed: () {
                                           setState(() {
+                                            _useDetailedRatings = true;
                                             _minRating = 4.0;
                                           });
                                         },
                                         style: ElevatedButton.styleFrom(
                                           backgroundColor: _minRating == 4.0 
                                               ? Colors.orange 
-                                              : Colors.grey[300],
+                                              : Theme.of(context).colorScheme.surfaceContainerHighest,
                                           foregroundColor: _minRating == 4.0 
                                               ? Colors.white 
                                               : Colors.black87,
                                         ),
-                                        child: const Text('דירוגים מפורטים'),
+                                        child: Text(l10n.detailedRatings),
                                       ),
                                     ),
                                   ],
                                 ),
-                                if (_minRating == 4.0) ...[
+                                if (_useDetailedRatings && _minRating == 4.0) ...[
+                                  const SizedBox(height: 20),
+                                  
+                                  // אמינות
+                                  _buildDetailedRatingField(
+                                    l10n.reliability,
+                                    '',
+                                    _minReliability,
+                                    (value) => setState(() => _minReliability = value),
+                                    Icons.verified_user,
+                                    Colors.blue,
+                                  ),
                                   const SizedBox(height: 12),
-                                  Text(
-                                    'דירוגים מפורטים: 4 כוכבים ומעלה',
-                                    style: TextStyle(
-                                      color: Colors.grey[600],
-                                      fontSize: 14,
-                                    ),
+                                  
+                                  // זמינות
+                                  _buildDetailedRatingField(
+                                    l10n.availability,
+                                    '',
+                                    _minAvailability,
+                                    (value) => setState(() => _minAvailability = value),
+                                    Icons.access_time,
+                                    Colors.green,
+                                  ),
+                                  const SizedBox(height: 12),
+                                  
+                                  // יחס
+                                  _buildDetailedRatingField(
+                                    l10n.attitude,
+                                    '',
+                                    _minAttitude,
+                                    (value) => setState(() => _minAttitude = value),
+                                    Icons.people,
+                                    Colors.orange,
+                                  ),
+                                  const SizedBox(height: 12),
+                                  
+                                  // מחיר הוגן
+                                  _buildDetailedRatingField(
+                                    l10n.fairPrice,
+                                    '',
+                                    _minFairPrice,
+                                    (value) => setState(() => _minFairPrice = value),
+                                    Icons.attach_money,
+                                    Colors.purple,
                                   ),
                                 ],
                               ],
@@ -1058,12 +1471,48 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                       ],
                       const SizedBox(height: 16),
 
+                      // תגית מותאמת אישית
+                      if (_selectedTags.isNotEmpty) ...[
+                        Container(
+                          padding: const EdgeInsets.all(16),
+                          decoration: BoxDecoration(
+                            border: Border.all(color: Theme.of(context).colorScheme.outlineVariant),
+                            borderRadius: BorderRadius.circular(8),
+                          ),
+                          child: Column(
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              Text(
+                                l10n.customTag,
+                                style: TextStyle(
+                                  fontSize: 16,
+                                  fontWeight: FontWeight.w500,
+                                ),
+                              ),
+                              const SizedBox(height: 8),
+                              TextFormField(
+                                initialValue: _customTag,
+                                decoration: InputDecoration(
+                                  labelText: l10n.customTag,
+                                  hintText: l10n.writeCustomTag,
+                                  border: OutlineInputBorder(),
+                                ),
+                        onChanged: (value) {
+                                  _customTag = value;
+                        },
+                              ),
+                            ],
+                          ),
+                      ),
+                      const SizedBox(height: 16),
+                      ],
+
                       // כפתור שמירה
                       ElevatedButton(
                         onPressed: _isLoading ? null : _updateRequest,
                         style: ElevatedButton.styleFrom(
                           backgroundColor: Theme.of(context).brightness == Brightness.dark 
-              ? const Color(0xFFFF9800) // כתום ענתיק
+              ? const Color(0xFF9C27B0) // סגול יפה
               : Theme.of(context).colorScheme.primary,
                           foregroundColor: Colors.white,
                           padding: const EdgeInsets.symmetric(vertical: 16),
@@ -1094,33 +1543,28 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
         children: [
           ClipRRect(
             borderRadius: BorderRadius.circular(8),
-            child: Image.network(
-              _selectedImages[index],
+            child: CachedNetworkImage(
+              imageUrl: _selectedImages[index],
               width: 80,
               height: 80,
               fit: BoxFit.cover,
-              errorBuilder: (context, error, stackTrace) {
-                return Container(
+              placeholder: (context, url) => Container(
+                width: 80,
+                height: 80,
+                color: Theme.of(context).colorScheme.surfaceContainer,
+                child: const Center(
+                  child: CircularProgressIndicator(strokeWidth: 2),
+                ),
+              ),
+              errorWidget: (context, url, error) => Container(
                   width: 80,
                   height: 80,
-                  color: Colors.grey[300],
+                  color: Theme.of(context).colorScheme.surfaceContainerHighest,
                   child: const Icon(
                     Icons.error,
                     color: Colors.red,
                   ),
-                );
-              },
-              loadingBuilder: (context, child, loadingProgress) {
-                if (loadingProgress == null) return child;
-                return Container(
-                  width: 80,
-                  height: 80,
-                  color: Colors.grey[200],
-                  child: const Center(
-                    child: CircularProgressIndicator(),
-                  ),
-                );
-              },
+              ),
             ),
           ),
           Positioned(
@@ -1131,13 +1575,15 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                 // מחיקת תמונה בודדת
                 final bool? shouldDelete = await showDialog<bool>(
                   context: context,
-                  builder: (context) => AlertDialog(
-                    title: const Text('מחיקת תמונה'),
-                    content: const Text('האם אתה בטוח שברצונך למחוק את התמונה?'),
+                  builder: (context) {
+                    final l10n = AppLocalizations.of(context);
+                    return AlertDialog(
+                      title: Text(l10n.deleteImage),
+                      content: Text(l10n.deleteImageConfirm),
                     actions: [
                       TextButton(
                         onPressed: () => Navigator.pop(context, false),
-                        child: const Text('ביטול'),
+                          child: Text(l10n.cancel),
                       ),
                       ElevatedButton(
                         onPressed: () => Navigator.pop(context, true),
@@ -1145,13 +1591,15 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                           backgroundColor: Colors.red,
                           foregroundColor: Colors.white,
                         ),
-                        child: const Text('מחק'),
+                          child: Text(l10n.delete),
                       ),
                     ],
-                  ),
+                    );
+                  },
                 );
 
                 if (shouldDelete == true) {
+                  final l10n = AppLocalizations.of(context);
                   try {
                     // מחיקת התמונה מ-Storage
                     await _deleteImagesFromStorage([_selectedImages[index]]);
@@ -1162,8 +1610,8 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                     
                     if (mounted) {
     ScaffoldMessenger.of(context).showSnackBar(
-      const SnackBar(
-                          content: Text('תמונה נמחקה בהצלחה'),
+      SnackBar(
+                          content: Text(l10n.imageDeletedSuccessfully),
                           backgroundColor: Colors.green,
                         ),
                       );
@@ -1172,7 +1620,7 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                     if (mounted) {
                       ScaffoldMessenger.of(context).showSnackBar(
                         SnackBar(
-                          content: Text('שגיאה במחיקת תמונה: $e'),
+                          content: Text(l10n.errorDeletingImage(e.toString())),
                           backgroundColor: Colors.red,
                         ),
                       );
@@ -1223,9 +1671,10 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
                 });
                 
                 if (mounted) {
+                  final l10n = AppLocalizations.of(context);
                   ScaffoldMessenger.of(context).showSnackBar(
-                    const SnackBar(
-                      content: Text('תמונה הוסרה מהרשימה'),
+                    SnackBar(
+                      content: Text(l10n.imageRemovedFromList),
                       backgroundColor: Colors.orange,
                     ),
                   );
@@ -1250,6 +1699,252 @@ class _EditRequestScreenState extends State<EditRequestScreen> {
     );
   }
 
+
+  // ✅ בדיקת מספר נותני שירות זמינים בתחום
+  Future<void> _checkAvailableHelpers() async {
+    debugPrint('🔍 Checking available helpers for category: ${_selectedCategory.toString()}');
+    
+    try {
+      // בדיקה מקיפה - נספור כל סוגי המשתמשים שיכולים לספק שירות בתחום הרלוונטי
+      // 1. משתמשים עסקיים עם מנוי פעיל
+      final businessQuery = await FirebaseFirestore.instance
+          .collection('users')
+          .where('userType', isEqualTo: 'business')
+          .where('isSubscriptionActive', isEqualTo: true)
+          .get();
+      
+      // 2. משתמשי אורח (עם מנוי פעיל)
+      final guestQuery = await FirebaseFirestore.instance
+          .collection('users')
+          .where('userType', isEqualTo: 'guest')
+          .where('isSubscriptionActive', isEqualTo: true)
+          .get();
+      
+      // 3. מנהלים
+      final adminQuery = await FirebaseFirestore.instance
+          .collection('users')
+          .where('userType', isEqualTo: 'admin')
+          .get();
+      
+      // איחוד כל התוצאות
+      final allUsers = [
+        ...businessQuery.docs,
+        ...guestQuery.docs,
+        ...adminQuery.docs,
+      ];
+      
+      int count = 0;
+      final selectedCategoryName = _selectedCategory.name;
+      final selectedCategoryDisplayName = _selectedCategory.categoryDisplayName;
+      
+      for (var doc in allUsers) {
+        final data = doc.data();
+        final businessCategories = data['businessCategories'] as List<dynamic>? ?? [];
+        
+        // בדיקה אם המשתמש הוא משתמש אמיתי (לא משתמש בדיקה עם כל הקטגוריות)
+        bool isRealUser = businessCategories.length < 20;
+        
+        if (!isRealUser) {
+          continue;
+        }
+        
+        bool canProvideService = false;
+        
+        if (businessCategories.isNotEmpty) {
+          for (var category in businessCategories) {
+            bool matches = false;
+            
+            if (category is Map) {
+              final mapCategoryName = category['category']?.toString() ?? '';
+              final mapCategoryDisplayName = category['categoryDisplayName']?.toString();
+              
+              if (mapCategoryName == selectedCategoryName) {
+                matches = true;
+              } else if (mapCategoryDisplayName != null && mapCategoryDisplayName == selectedCategoryDisplayName) {
+                matches = true;
+              }
+            } else if (category is String) {
+              if (category == selectedCategoryName || category == selectedCategoryDisplayName) {
+                matches = true;
+              } else {
+                try {
+                  final cat = RequestCategory.values.firstWhere(
+                    (c) => c.name == category || c.categoryDisplayName == category,
+                    orElse: () => RequestCategory.plumbing,
+                  );
+                  if (cat == _selectedCategory) {
+                    matches = true;
+                  }
+                } catch (e) {
+                  // Continue
+                }
+              }
+            }
+            
+            if (matches) {
+              canProvideService = true;
+              break;
+            }
+          }
+        }
+        
+        if (canProvideService) {
+          count++;
+        }
+      }
+      
+      debugPrint('🎯 Total helpers found: $count');
+      
+      // ✅ הצגת דיאלוג עם מספר נותני שירות רק כשמשתמש בוחר "רק לנותני שירות מתחום X"
+      debugPrint('📊 Showing dialog with helpers count: $count');
+      _showHelpersCountDialog(count);
+    } catch (e) {
+      debugPrint('Error checking available helpers: $e');
+    }
+  }
+  
+  // ✅ הצגת דיאלוג עם מספר נותני שירות בתחום
+  void _showHelpersCountDialog(int count) {
+    final l10n = AppLocalizations.of(context);
+    final hasHelpers = count > 0;
+    
+    showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(16),
+        ),
+        title: Row(
+          children: [
+            Icon(
+              hasHelpers ? Icons.check_circle_outline : Icons.info_outline,
+              color: hasHelpers ? Theme.of(context).colorScheme.primary : Theme.of(context).colorScheme.primary,
+              size: 28,
+            ),
+            const SizedBox(width: 8),
+            Expanded(
+              child: Text(
+                hasHelpers
+                    ? l10n.serviceProvidersInCategory(count)
+                    : l10n.noServiceProvidersInCategory,
+                style: TextStyle(
+                  fontSize: 18,
+                  fontWeight: FontWeight.bold,
+                ),
+                overflow: TextOverflow.ellipsis,
+                maxLines: 2,
+              ),
+            ),
+          ],
+        ),
+        content: SingleChildScrollView(
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Container(
+                padding: const EdgeInsets.all(12),
+                decoration: BoxDecoration(
+                  color: hasHelpers ? Theme.of(context).colorScheme.primaryContainer : Theme.of(context).colorScheme.primaryContainer,
+                  borderRadius: BorderRadius.circular(8),
+                  border: Border.all(
+                    color: hasHelpers ? Theme.of(context).colorScheme.primary : Theme.of(context).colorScheme.primary,
+                  ),
+                ),
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  mainAxisSize: MainAxisSize.min,
+                  children: [
+                    Text(
+                      hasHelpers
+                          ? l10n.serviceProvidersInCategoryMessage(count)
+                          : l10n.noServiceProvidersInCategoryMessage,
+                      style: TextStyle(
+                        fontSize: 14,
+                        fontWeight: FontWeight.w600,
+                        color: hasHelpers ? Theme.of(context).colorScheme.onPrimaryContainer : Theme.of(context).colorScheme.onPrimaryContainer,
+                      ),
+                    ),
+                    if (!hasHelpers) ...[
+                      const SizedBox(height: 8),
+                      Text(
+                        l10n.continueCreatingRequestMessage,
+                        style: TextStyle(
+                          fontSize: 13,
+                          color: Theme.of(context).colorScheme.onPrimaryContainer,
+                        ),
+                      ),
+                    ],
+                  ],
+                ),
+              ),
+              const SizedBox(height: 16),
+              // ✅ חלק "עזור לנו למצוא נותני שירות, שתף את האפליקציה"
+              Container(
+                padding: const EdgeInsets.all(12),
+                decoration: BoxDecoration(
+                  color: Theme.of(context).colorScheme.primaryContainer,
+                  borderRadius: BorderRadius.circular(8),
+                  border: Border.all(color: Theme.of(context).colorScheme.primary),
+                ),
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  mainAxisSize: MainAxisSize.min,
+                  children: [
+                    Row(
+                      children: [
+                        Icon(Icons.share, color: Theme.of(context).colorScheme.primary, size: 20),
+                        const SizedBox(width: 8),
+                        Expanded(
+                          child: Text(
+                            l10n.helpGrowCommunity,
+                            style: TextStyle(
+                              fontSize: 14,
+                              fontWeight: FontWeight.w600,
+                              color: Theme.of(context).colorScheme.onPrimaryContainer,
+                            ),
+                            overflow: TextOverflow.ellipsis,
+                            maxLines: 2,
+                          ),
+                        ),
+                      ],
+                    ),
+                    const SizedBox(height: 8),
+                    Text(
+                      l10n.shareAppToGrowProviders,
+                      style: TextStyle(
+                        fontSize: 13,
+                        color: Theme.of(context).colorScheme.primary,
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+            ],
+          ),
+        ),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.pop(context),
+            child: Text(l10n.understood),
+          ),
+          ElevatedButton.icon(
+            onPressed: () {
+              Navigator.pop(context);
+              // פתיחת מסך שיתוף
+              AppSharingService.shareApp(context);
+            },
+            icon: const Icon(Icons.share, size: 18),
+            label: Text(l10n.shareNow),
+            style: ElevatedButton.styleFrom(
+              backgroundColor: Theme.of(context).colorScheme.primary,
+              foregroundColor: Colors.white,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
 
   String _getTypeDisplayName(RequestType type) {
     switch (type) {
