@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import '../services/tutorial_service.dart';
+import '../l10n/app_localizations.dart';
 
 class TutorialCenterScreen extends StatefulWidget {
   const TutorialCenterScreen({super.key});
@@ -11,272 +12,97 @@ class TutorialCenterScreen extends StatefulWidget {
 class _TutorialCenterScreenState extends State<TutorialCenterScreen>
     with SingleTickerProviderStateMixin {
   late TabController _tabController;
-  final List<TutorialCategory> _categories = [
+  
+  List<TutorialCategory> _buildCategories(AppLocalizations l10n) {
+    return [
     TutorialCategory(
       id: 'home',
-      title: 'מסך הבית',
+      title: l10n.tutorialCategoryHome,
       icon: Icons.home,
       color: Colors.blue,
       tutorials: [
         TutorialItem(
           id: 'home_basics',
-          title: 'תכולת מסך הבית',
-          description: 'למד כיצד לנווט במסך הבית ולהשתמש בפונקציות הבסיסיות',
-          content: '''
-# יסודות מסך הבית
-
-## מה תמצא במסך הבית:
-• **רשימת בקשות** - כל הבקשות הזמינות בקהילה
-• **חיפוש** - חפש בקשות לפי מילות מפתח
-• **סינון** - סנן בקשות לפי קטגוריה, מיקום ומחיר
-• **בקשות שלי** - בקשות שפרסמת או פנית אליהן
-
-## איך להשתמש:
-1. **לצפייה בבקשה** - לחץ על בקשה מהרשימה
-2. **לחיפוש** - השתמש בשורת החיפוש העליונה
-3. **לסינון** - לחץ על כפתור "סינון מתקדם"
-4. **לבקשות שלי** - לחץ על "בקשות שפניתי אליהם"
-          ''',
+          title: l10n.tutorialHomeBasicsTitle,
+          description: l10n.tutorialHomeBasicsDescription,
+          content: l10n.tutorialHomeBasicsContent,
         ),
         TutorialItem(
           id: 'home_search',
-          title: 'חיפוש וסינון',
-          description: 'איך למצוא בקשות ספציפיות במהירות',
-          content: '''
-# חיפוש וסינון
-
-## חיפוש:
-• **מילות מפתח** - הזן מילים רלוונטיות
-• **חיפוש בזמן אמת** - התוצאות מתעדכנות תוך כדי הקלדה
-• **חיפוש בכל השדות** - שם, תיאור, קטגוריה
-
-## סינון מתקדם:
-• **קטגוריות** - בחר תחומי עיסוק ספציפיים
-• **טווח מחיר** - הגדר מינימום ומקסימום
-• **מיקום** - חפש בקרבת מקום מסוים
-• **תאריך** - בקשות חדשות או ישנות
-
-## טיפים:
-• השתמש במילות מפתח קצרות וברורות
-• נסה חיפושים שונים לאותה בקשה
-• שמור סינונים מועדפים
-          ''',
+          title: l10n.tutorialHomeSearchTitle,
+          description: l10n.tutorialHomeSearchDescription,
+          content: l10n.tutorialHomeSearchContent,
         ),
       ],
     ),
     TutorialCategory(
       id: 'requests',
-      title: 'בקשות',
+      title: l10n.tutorialCategoryRequests,
       icon: Icons.assignment,
       color: Colors.green,
       tutorials: [
         TutorialItem(
           id: 'create_request',
-          title: 'יצירת בקשה חדשה',
-          description: 'איך ליצור בקשה לעזרה או שירות',
-          content: '''
-# יצירת בקשה חדשה
-
-## שלבים ליצירת בקשה:
-1. **לחץ על +** - בפינה הימנית התחתונה
-2. **בחר קטגוריה** - תחום העיסוק המתאים
-3. **כתוב תיאור** - הסבר מה אתה צריך
-4. **הגדר מחיר** - אם רלוונטי
-5. **בחר מיקום** - איפה אתה נמצא
-6. **פרסם** - שלח את הבקשה לקהילה
-
-## טיפים לכתיבה טובה:
-• **תיאור ברור** - הסבר בדיוק מה אתה צריך
-• **פרטים חשובים** - זמן, מקום, דחיפות
-• **תמונה** - הוסף תמונה אם זה עוזר להסביר
-• **מחיר הוגן** - הצע מחיר סביר
-
-## דוגמה טובה:
-"צריך עזרה בהעברת רהיטים מדירה לדירה בירושלים. 
-יש 3 ארונות, 2 שולחנות ומיטות. 
-מוכן לשלם 200-300 שקל. 
-מועד: סוף השבוע הקרוב."
-          ''',
+          title: l10n.tutorialCreateRequestTitle,
+          description: l10n.tutorialCreateRequestDescription,
+          content: l10n.tutorialCreateRequestContent,
         ),
         TutorialItem(
           id: 'manage_requests',
-          title: 'ניהול בקשות',
-          description: 'איך לנהל את הבקשות שלך',
-          content: '''
-# ניהול בקשות
-
-## בקשות שפרסמת:
-• **צפייה בפניות** - ראה מי פנה אליך
-• **עריכת בקשה** - עדכן פרטים או מחיר
-• **סגירת בקשה** - כשהתקבל עזרה
-• **מחיקת בקשה** - אם כבר לא רלוונטית
-
-## בקשות שפנית אליהן:
-• **מעקב סטטוס** - ראה אם התקבלת
-• **צ'אט עם המפרסם** - תקשורת ישירה
-• **ביטול פנייה** - אם שינית דעתך
-
-## טיפים לניהול:
-• **עדכן סטטוס** - סמן כשהבקשה הושלמה
-• **תקשר בצ'אט** - שאל שאלות לפני התחייבות
-• **בדוק פרופילים** - ראה דירוגים של נותני שירות
-• **היה מנומס** - תגיב בזמן ובכבוד
-          ''',
+          title: l10n.tutorialManageRequestsTitle,
+          description: l10n.tutorialManageRequestsDescription,
+          content: l10n.tutorialManageRequestsContent,
         ),
       ],
     ),
     TutorialCategory(
       id: 'chat',
-      title: 'צ\'אט',
+      title: l10n.tutorialCategoryChat,
       icon: Icons.chat,
       color: Colors.orange,
       tutorials: [
         TutorialItem(
           id: 'chat_basics',
-          title: 'תכולת הצ\'אט',
-          description: 'איך להשתמש במערכת הצ\'אט',
-          content: '''
-# יסודות הצ'אט
-
-## איך נכנסים לצ'אט:
-1. **מבקשה שפרסמת** - לחץ על "צ'אט" ליד פנייה
-2. **מבקשה שפנית אליה** - לחץ על "צ'אט" בבקשה
-3. **מהמסך "בקשות שלי"** - לחץ על כפתור הצ'אט
-
-## פונקציות הצ'אט:
-• **שליחת הודעות** - טקסט, תמונות, קבצים
-• **עריכת הודעות** - לחץ ארוך על הודעה ששלחת
-• **מחיקת הודעות** - לחץ ארוך ובחר "מחק"
-• **סגירת צ'אט** - אם לא רוצה יותר לתקשר
-
-## סימני קריאה:
-• **וי אחד** - הודעה נשלחה
-• **שני וי** - הודעה נקראה על ידי הנמען
-• **וי זוהר** - הודעה נקראה לאחרונה
-
-## כללי התנהגות:
-• **היה מנומס** - השתמש בשפה נאותה
-• **תגיב בזמן** - אל תשאיר הודעות ללא מענה
-• **היה ברור** - הסבר בדיוק מה אתה צריך
-• **שמור על פרטיות** - אל תעביר מידע אישי
-          ''',
+          title: l10n.tutorialChatBasicsTitle,
+          description: l10n.tutorialChatBasicsDescription,
+          content: l10n.tutorialChatBasicsContent,
         ),
         TutorialItem(
           id: 'chat_advanced',
-          title: 'פונקציות מתקדמות',
-          description: 'פונקציות מתקדמות של הצ\'אט',
-          content: '''
-# פונקציות מתקדמות
-
-## עריכת הודעות:
-1. **לחץ ארוך** על הודעה ששלחת
-2. **בחר "ערוך"** מהתפריט
-3. **ערוך את הטקסט** והשמור
-4. **ההודעה תסומן** כ"נערכה"
-
-## מחיקת הודעות:
-1. **לחץ ארוך** על הודעה ששלחת
-2. **בחר "מחק"** מהתפריט
-3. **אשר את המחיקה**
-4. **ההודעה תימחק** לצמיתות
-
-## סגירת צ'אט:
-• **מתי לסגור** - כשאתה לא רוצה יותר לתקשר
-• **איך לסגור** - תפריט (3 נקודות) > "סגור צ'אט"
-• **פתיחה מחדש** - אפשר לפתוח צ'אט סגור
-• **הודעת מערכת** - תישלח הודעה על סגירת הצ'אט
-
-## ניקוי צ'אט:
-• **מחיקת היסטוריה** - תפריט > "נקה צ'אט"
-• **השפעה** - מוחק את כל ההודעות
-• **לא ניתן לשחזר** - פעולה סופית
-          ''',
+          title: l10n.tutorialChatAdvancedTitle,
+          description: l10n.tutorialChatAdvancedDescription,
+          content: l10n.tutorialChatAdvancedContent,
         ),
       ],
     ),
     TutorialCategory(
       id: 'profile',
-      title: 'פרופיל',
+      title: l10n.tutorialCategoryProfile,
       icon: Icons.person,
       color: Colors.purple,
       tutorials: [
         TutorialItem(
           id: 'profile_setup',
-          title: 'הגדרת פרופיל',
-          description: 'איך להגדיר את הפרופיל שלך',
-          content: '''
-# הגדרת פרופיל
-
-## מידע בסיסי:
-• **שם** - השם שלך (חובה)
-• **אימייל** - כתובת אימייל (חובה)
-• **תמונה** - תמונת פרופיל (אופציונלי)
-• **מיקום** - איפה אתה נמצא (חובה)
-
-## מידע עסקי (למנויים):
-• **תחומי עיסוק** - מה אתה מציע
-• **תיאור** - הסבר על השירותים שלך
-• **מחירים** - מחירון כללי
-• **זמינות** - מתי אתה זמין
-
-## הגדרות פרטיות:
-• **מספר טלפון** - אם רוצה שיוצג
-• **הצגת טלפון** - הסכמה להציג במפה
-• **הגדרות התראות** - איזה התראות לקבל
-
-## טיפים לפרופיל טוב:
-• **תמונה מקצועית** - תמונה ברורה ונעימה
-• **תיאור מפורט** - הסבר מה אתה מציע
-• **מידע אמיתי** - אל תכתוב דברים לא נכונים
-• **עדכון קבוע** - עדכן מידע כשמשתנה
-          ''',
+          title: l10n.tutorialProfileSetupTitle,
+          description: l10n.tutorialProfileSetupDescription,
+          content: l10n.tutorialProfileSetupContent,
         ),
         TutorialItem(
           id: 'subscription',
-          title: 'מנויים ותשלומים',
-          description: 'איך לנהל מנוי ותשלומים',
-          content: '''
-# מנויים ותשלומים
-
-## סוגי מנויים:
-• **פרטי חינם** - גישה בסיסית לאפליקציה
-• **פרטי מנוי** - פונקציות מתקדמות
-• **עסקי מנוי** - פרסום שירותים
-• **מנהל** - גישה מלאה לכל הפונקציות
-
-## מה כלול בכל מנוי:
-### פרטי חינם:
-• צפייה בבקשות
-• פנייה לבקשות
-• צ'אט בסיסי
-
-### פרטי מנוי:
-• כל מה שבחינם +
-• פרסום בקשות
-• סינון מתקדם
-• התראות מותאמות
-
-### עסקי מנוי:
-• כל מה שבפרטי מנוי +
-• פרסום שירותים
-• הופעה במפה
-• דירוגים מפורטים
-
-## תשלומים:
-• **אמצעי תשלום** - כרטיס אשראי, העברה בנקאית
-• **חיוב חודשי** - אוטומטי
-• **ביטול** - אפשר לבטל בכל עת
-• **החזר כספי** - לפי מדיניות החברה
-          ''',
+          title: l10n.tutorialSubscriptionTitle,
+          description: l10n.tutorialSubscriptionDescription,
+          content: l10n.tutorialSubscriptionContent,
         ),
       ],
     ),
   ];
+  }
 
   @override
   void initState() {
     super.initState();
-    _tabController = TabController(length: _categories.length, vsync: this);
+    // TabController will be initialized in build method
+    _tabController = TabController(length: 4, vsync: this);
   }
 
   @override
@@ -287,9 +113,18 @@ class _TutorialCenterScreenState extends State<TutorialCenterScreen>
 
   @override
   Widget build(BuildContext context) {
+    final l10n = AppLocalizations.of(context);
+    final categories = _buildCategories(l10n);
+    
+    // Update TabController length if needed
+    if (_tabController.length != categories.length) {
+      _tabController.dispose();
+      _tabController = TabController(length: categories.length, vsync: this);
+    }
+    
     return Scaffold(
       appBar: AppBar(
-        title: const Text('מדריך משתמש'),
+        title: Text(l10n.userGuide),
         backgroundColor: Theme.of(context).colorScheme.primary,
         foregroundColor: Colors.white,
         bottom: TabBar(
@@ -298,7 +133,7 @@ class _TutorialCenterScreenState extends State<TutorialCenterScreen>
           indicatorColor: Colors.white,
           labelColor: Colors.white,
           unselectedLabelColor: Colors.white70,
-          tabs: _categories.map((category) => Tab(
+          tabs: categories.map((category) => Tab(
             icon: Icon(category.icon),
             text: category.title,
           )).toList(),
@@ -306,12 +141,12 @@ class _TutorialCenterScreenState extends State<TutorialCenterScreen>
       ),
       body: TabBarView(
         controller: _tabController,
-        children: _categories.map((category) => _buildCategoryContent(category)).toList(),
+        children: categories.map((category) => _buildCategoryContent(category, l10n)).toList(),
       ),
     );
   }
 
-  Widget _buildCategoryContent(TutorialCategory category) {
+  Widget _buildCategoryContent(TutorialCategory category, AppLocalizations l10n) {
     return SingleChildScrollView(
       padding: const EdgeInsets.all(16),
       child: Column(
@@ -323,7 +158,7 @@ class _TutorialCenterScreenState extends State<TutorialCenterScreen>
             padding: const EdgeInsets.all(20),
             decoration: BoxDecoration(
               gradient: LinearGradient(
-                colors: [category.color, category.color.withOpacity(0.7)],
+                colors: [category.color, category.color.withValues(alpha: 0.7)],
                 begin: Alignment.topLeft,
                 end: Alignment.bottomRight,
               ),
@@ -343,7 +178,7 @@ class _TutorialCenterScreenState extends State<TutorialCenterScreen>
                 ),
                 const SizedBox(height: 8),
                 Text(
-                  '${category.tutorials.length} הדרכות זמינות',
+                  l10n.tutorialsAvailable(category.tutorials.length),
                   style: const TextStyle(
                     fontSize: 16,
                     color: Colors.white70,
@@ -380,7 +215,7 @@ class _TutorialCenterScreenState extends State<TutorialCenterScreen>
                   Container(
                     padding: const EdgeInsets.all(8),
                     decoration: BoxDecoration(
-                      color: categoryColor.withOpacity(0.1),
+                      color: categoryColor.withValues(alpha: 0.1),
                       borderRadius: BorderRadius.circular(8),
                     ),
                     child: Icon(
@@ -406,7 +241,7 @@ class _TutorialCenterScreenState extends State<TutorialCenterScreen>
                           tutorial.description,
                           style: TextStyle(
                             fontSize: 14,
-                            color: Colors.grey[600],
+                            color: Theme.of(context).colorScheme.onSurfaceVariant,
                           ),
                         ),
                       ],
@@ -414,7 +249,7 @@ class _TutorialCenterScreenState extends State<TutorialCenterScreen>
                   ),
                   Icon(
                     Icons.arrow_forward_ios,
-                    color: Colors.grey[400],
+                    color: Theme.of(context).colorScheme.onSurfaceVariant,
                     size: 16,
                   ),
                 ],
@@ -441,7 +276,7 @@ class _TutorialCenterScreenState extends State<TutorialCenterScreen>
                 padding: const EdgeInsets.all(20),
                 decoration: BoxDecoration(
                   gradient: LinearGradient(
-                    colors: [categoryColor, categoryColor.withOpacity(0.7)],
+                    colors: [categoryColor, categoryColor.withValues(alpha: 0.7)],
                     begin: Alignment.topLeft,
                     end: Alignment.bottomRight,
                   ),
@@ -491,16 +326,19 @@ class _TutorialCenterScreenState extends State<TutorialCenterScreen>
                   children: [
                     TextButton(
                       onPressed: () => Navigator.pop(context),
-                      child: const Text('סגור'),
+                      child: Text(AppLocalizations.of(context).tutorialClose),
                     ),
                     ElevatedButton(
                       onPressed: () async {
                         // סמן כנקרא
                         await TutorialService.markTutorialAsRead(tutorial.id);
+                        // Guard context usage after async gap - check context.mounted for builder context
+                        if (!context.mounted) return;
                         Navigator.pop(context);
+                        if (!context.mounted) return;
                         ScaffoldMessenger.of(context).showSnackBar(
-                          const SnackBar(
-                            content: Text('הדרכה סומנה כנקראה'),
+                          SnackBar(
+                            content: Text(AppLocalizations.of(context).tutorialMarkedAsRead),
                             backgroundColor: Colors.green,
                           ),
                         );
@@ -509,7 +347,7 @@ class _TutorialCenterScreenState extends State<TutorialCenterScreen>
                         backgroundColor: categoryColor,
                         foregroundColor: Colors.white,
                       ),
-                      child: const Text('קראתי'),
+                      child: Text(AppLocalizations.of(context).tutorialRead),
                     ),
                   ],
                 ),
