@@ -5,10 +5,16 @@ import '../services/audio_service.dart';
 import '../services/google_auth_service.dart';
 import '../services/auto_login_service.dart';
 import '../services/terms_service.dart';
+import '../services/push_notification_service.dart';
+import '../services/cloud_function_service.dart';
+import '../services/permission_service.dart';
 import '../widgets/remember_me_dialog.dart';
 import 'terms_and_privacy_screen.dart';
-import 'guest_category_selection_screen.dart';
+import 'about_app_screen.dart';
+import '../l10n/app_localizations.dart';
 import 'dart:math' as math;
+import 'dart:io' show Platform;
+import 'package:flutter/foundation.dart' show kIsWeb;
 
 class YokiStyleAuthScreen extends StatefulWidget {
   final VoidCallback? onLoginSuccess;
@@ -24,6 +30,8 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
   bool _isLoading = false;
   late AnimationController _animationController;
   late Animation<double> _animation;
+  String? _pendingVerificationEmail;
+  String? _pendingVerificationPassword;
 
   @override
   void initState() {
@@ -43,25 +51,73 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
     
     // בדיקת התחברות אוטומטית
     _checkAutoLogin();
+    
+    // בקשות הרשאות - רק במסך התחברות
+    _requestNotificationPermission();
+    _requestLocationPermission();
+  }
+  
+  Future<void> _requestNotificationPermission() async {
+    // המתן קצת לפני בקשת ההרשאה
+    await Future.delayed(const Duration(seconds: 2));
+    
+    if (mounted) {
+      // בקשת הרשאות דרך PermissionService (דיאלוג)
+      await PermissionService.requestNotificationPermission(context);
+      
+      // בקשת הרשאות דרך PushNotificationService (FCM)
+      await PushNotificationService.requestPermissions();
+    }
+  }
+
+  Future<void> _requestLocationPermission() async {
+    // המתן יותר זמן לפני בקשת ההרשאה
+    await Future.delayed(const Duration(seconds: 5));
+    
+    if (mounted) {
+      await PermissionService.requestLocationPermission(context);
+    }
   }
   
   /// בדיקת התחברות אוטומטית
   Future<void> _checkAutoLogin() async {
     try {
-      // בדיקה אם המשתמש בחר לא לזכור פרטי כניסה
-      final shouldRemember = await AutoLoginService.shouldRememberMe();
-      if (!shouldRemember) {
-        debugPrint('User chose not to remember login, signing out');
-        // אם המשתמש בחר לא לזכור, נתנתק אותו
+      // ✅ בדיקה אם המשתמש התנתק מפורשות
+      final userLoggedOut = await AutoLoginService.hasUserLoggedOut();
+      if (userLoggedOut) {
+        debugPrint('User logged out explicitly, signing out');
         await FirebaseAuth.instance.signOut();
         return;
       }
 
+      // ✅ בדיקה אם המשתמש בחר "זכור אותי"
+      final shouldRemember = await AutoLoginService.shouldRememberMe();
+      if (!shouldRemember) {
+        debugPrint('User chose not to remember login - signing out');
+        // אם המשתמש לא בחר "זכור אותי" (לחץ "לא תודה"), נתנתק אותו
+        // כך שבפעם הבאה יצטרך להתחבר שוב
+        await FirebaseAuth.instance.signOut();
+        return;
+      }
+
+      // ✅ אם המשתמש בחר "זכור אותי", נבדוק אם הוא כבר מחובר
+      final currentUser = FirebaseAuth.instance.currentUser;
+      if (currentUser != null) {
+        debugPrint('✅ User already logged in (${currentUser.uid}) and chose to remember - keeping logged in');
+        // המשתמש כבר מחובר ובחר "זכור אותי" - ה-StreamBuilder ב-main.dart כבר יציג את MainApp
+        return;
+      }
+
+      // ✅ המשתמש לא מחובר אבל בחר "זכור אותי" - ננסה auto-login
+      debugPrint('Attempting auto login (user chose to remember login)');
       final userCredential = await AutoLoginService.autoLogin();
       if (userCredential != null && mounted) {
         // התחברות אוטומטית הצליחה
+        debugPrint('✅ Auto login successful');
         await playSuccessSound();
         widget.onLoginSuccess?.call();
+      } else {
+        debugPrint('Auto login failed or returned null');
       }
     } catch (e) {
       debugPrint('Auto login failed: $e');
@@ -79,87 +135,50 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
     return Scaffold(
       body: Stack(
         children: [
-          // רקע אנימטיבי - תמונות זזות משמאל לימין
+          // רקע Full Screen עם אנימציית Rainbow Gradient
           AnimatedBuilder(
             animation: _animation,
             builder: (context, child) {
               return Container(
-                decoration: const BoxDecoration(
+                decoration: BoxDecoration(
                   gradient: LinearGradient(
                     begin: Alignment.topLeft,
                     end: Alignment.bottomRight,
                     colors: [
-                      Color(0xFF1A1A2E),
-                      Color(0xFF16213E),
-                      Color(0xFF0F3460),
+                      HSVColor.fromAHSV(
+                        1,
+                        (_animation.value * 360) % 360, // מחזור צבעים
+                        0.75,
+                        0.85,
+                      ).toColor(),
+                      HSVColor.fromAHSV(
+                        1,
+                        ((_animation.value * 360) + 60) % 360,
+                        0.75,
+                        0.85,
+                      ).toColor(),
+                      HSVColor.fromAHSV(
+                        1,
+                        ((_animation.value * 360) + 120) % 360,
+                        0.75,
+                        0.85,
+                      ).toColor(),
                     ],
                   ),
                 ),
-                child: Stack(
-                  children: [
-                    // תמונות רקע אנימטיביות - תנועה משמאל לימין
-                    ...List.generate(8, (index) {
-                      final screenWidth = MediaQuery.of(context).size.width;
-                      final animationOffset = _animation.value * (screenWidth + 200) - 100;
-                      
-                      return Positioned(
-                        left: animationOffset + (index * 150),
-                        top: 50 + (index * 80) % 300,
-                        child: Transform.rotate(
-                          angle: _animation.value * 0.5 * math.pi + index,
+                // שכבה שקופה קלה להוספת עומק בעיצוב
                           child: Container(
-                            width: 120 + (index * 20),
-                            height: 120 + (index * 20),
                             decoration: BoxDecoration(
-                              shape: BoxShape.circle,
-                              gradient: RadialGradient(
+                    gradient: LinearGradient(
+                      begin: Alignment.topCenter,
+                      end: Alignment.bottomCenter,
                                 colors: [
-                                  Colors.purple.withOpacity(0.4),
-                                  Colors.blue.withOpacity(0.3),
-                                  Colors.pink.withOpacity(0.2),
+                        Colors.white.withOpacity(0.06),
+                        Colors.black.withOpacity(0.05),
                                   Colors.transparent,
                                 ],
                               ),
-                              boxShadow: [
-                                BoxShadow(
-                                  color: Colors.purple.withOpacity(0.3),
-                                  blurRadius: 20,
-                                  spreadRadius: 5,
-                                ),
-                              ],
-                            ),
-                          ),
-                        ),
-                      );
-                    }),
-                    // תמונות מטושטשות נוספות - תנועה משמאל לימין
-                    ...List.generate(6, (index) {
-                      final screenWidth = MediaQuery.of(context).size.width;
-                      final animationOffset = (_animation.value * (screenWidth + 300) - 150) * (index % 2 == 0 ? 1 : -1);
-                      
-                      return Positioned(
-                        left: animationOffset + (index * 200),
-                        top: 100 + (index * 120) % 400,
-                        child: Transform.scale(
-                          scale: 0.5 + (math.sin(_animation.value * 2 * math.pi + index) * 0.3),
-                          child: Container(
-                            width: 80,
-                            height: 80,
-                            decoration: BoxDecoration(
-                              shape: BoxShape.circle,
-                              gradient: RadialGradient(
-                                colors: [
-                                  Colors.cyan.withOpacity(0.3),
-                                  Colors.teal.withOpacity(0.2),
-                                  Colors.transparent,
-                                ],
-                              ),
-                            ),
-                          ),
-                        ),
-                      );
-                    }),
-                  ],
+                  ),
                 ),
               );
             },
@@ -178,9 +197,6 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
                     
                     const SizedBox(height: 25),
                     
-                    // הודעה זמנית על פיתוח
-                    _buildDevelopmentNotice(),
-                    
                     const SizedBox(height: 20),
                     
                     // כפתורי כניסה חברתית (רק גוגל)
@@ -198,11 +214,13 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
                     
                     const SizedBox(height: 15),
                     
+                    // המשך ללא הרשמה
+                    _buildContinueWithoutRegistrationButton(),
                     
                     const SizedBox(height: 20),
                     
-                    // קישורים לתנאי שימוש ומדיניות פרטיות
-                    _buildTermsAndPrivacyLinks(),
+                    // קישורים לתנאי שימוש, מדיניות פרטיות ואודות
+                    _buildLegalLinks(),
                     
                     // רווח נוסף בתחתית
                     const SizedBox(height: 30),
@@ -216,54 +234,9 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
     );
   }
 
-  Widget _buildDevelopmentNotice() {
-    return Container(
-      margin: const EdgeInsets.symmetric(horizontal: 20),
-      padding: const EdgeInsets.all(16),
-      decoration: BoxDecoration(
-        color: Colors.orange.withOpacity(0.1),
-        borderRadius: BorderRadius.circular(12),
-        border: Border.all(
-          color: Colors.orange.withOpacity(0.3),
-          width: 1,
-        ),
-      ),
-      child: Column(
-        children: [
-          Row(
-            children: [
-              Icon(
-                Icons.construction,
-                color: Colors.orange[700],
-                size: 20,
-              ),
-              const SizedBox(width: 8),
-              Text(
-                'האפליקציה בתהליך פיתוח',
-                style: TextStyle(
-                  color: Colors.orange[700],
-                  fontWeight: FontWeight.bold,
-                  fontSize: 16,
-                ),
-              ),
-            ],
-          ),
-          const SizedBox(height: 8),
-          Text(
-            'כרגע חלק מהפונקציות (כמו תשלומים) עדיין בפיתוח. אנחנו עובדים על זה!',
-            style: TextStyle(
-              color: Colors.orange[600],
-              fontSize: 14,
-              height: 1.4,
-            ),
-            textAlign: TextAlign.center,
-          ),
-        ],
-      ),
-    );
-  }
 
   Widget _buildLogoAndTitle() {
+    final l10n = AppLocalizations.of(context);
     return Column(
       children: [
         // לוגו שכונתי
@@ -275,58 +248,27 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
             borderRadius: BorderRadius.circular(15),
             boxShadow: [
               BoxShadow(
-                color: Colors.black.withOpacity(0.2),
+                color: Colors.black.withValues(alpha: 0.2),
                 blurRadius: 15,
                 offset: const Offset(0, 8),
               ),
             ],
           ),
-          child: Stack(
-            alignment: Alignment.center,
-            children: [
-              // לב אדום במרכז
-              const Icon(
-                Icons.favorite,
-                size: 20,
-                color: Colors.red,
-              ),
-              // ידיים צבעוניות סביב הלב
-              ...List.generate(6, (index) {
-                final angle = (index * 60) * (3.14159 / 180);
-                final radius = 25.0;
-                final x = radius * math.cos(angle);
-                final y = radius * math.sin(angle);
-                
-                return Positioned(
-                  left: 40 + x - 6,
-                  top: 40 + y - 6,
-                  child: Container(
-                    width: 12,
-                    height: 12,
-                    decoration: BoxDecoration(
-                      color: [
-                        Colors.blue,
-                        Colors.green,
-                        Colors.orange,
-                        Colors.pink,
-                        Colors.purple,
-                        Colors.yellow,
-                      ][index],
-                      shape: BoxShape.circle,
+          child: Padding(
+            padding: const EdgeInsets.all(12.0),
+            child: Image.asset(
+              'assets/images/logolarge.png',
+              fit: BoxFit.contain,
                     ),
-                  ),
-                );
-              }),
-            ],
           ),
         ),
         
         const SizedBox(height: 12),
         
         // כותרת
-        const Text(
-          'שכונתי',
-          style: TextStyle(
+        Text(
+          l10n.appTitle,
+          style: const TextStyle(
             fontSize: 24,
             fontWeight: FontWeight.bold,
             color: Colors.white,
@@ -335,22 +277,35 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
         
         const SizedBox(height: 4),
         
-        const Text(
-          'ברוך הבא',
-          style: TextStyle(
-            fontSize: 14,
+        Text(
+          l10n.welcome,
+          style: const TextStyle(
+            fontSize: 18,
+            fontWeight: FontWeight.bold,
+            color: Colors.white,
+          ),
+        ),
+        
+        const SizedBox(height: 8),
+        
+        Text(
+          l10n.welcomeSubtitle,
+          style: const TextStyle(
+            fontSize: 13,
             color: Colors.white70,
           ),
+          textAlign: TextAlign.center,
         ),
       ],
     );
   }
 
   Widget _buildGoogleButton() {
+    final l10n = AppLocalizations.of(context);
     return _buildSocialButton(
       icon: null, // נשתמש בלוגו מותאם אישית
       customIcon: _buildGoogleLogo(),
-      label: 'המשך עם גוגל',
+      label: l10n.continueWithGoogle,
       color: Colors.white,
       textColor: Colors.black87,
       onPressed: _handleGoogleLogin,
@@ -389,7 +344,7 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
           backgroundColor: color,
           foregroundColor: textColor,
           elevation: 8,
-          shadowColor: Colors.black.withOpacity(0.3),
+          shadowColor: Colors.black.withValues(alpha: 0.3),
           shape: RoundedRectangleBorder(
             borderRadius: BorderRadius.circular(16),
           ),
@@ -422,23 +377,28 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
         Expanded(
           child: Container(
             height: 1,
-            color: Colors.white.withOpacity(0.3),
+            color: Colors.white.withValues(alpha: 0.3),
           ),
         ),
         Padding(
           padding: const EdgeInsets.symmetric(horizontal: 16),
-          child: Text(
-            'או',
+          child: Builder(
+            builder: (context) {
+              final l10n = AppLocalizations.of(context);
+              return Text(
+                l10n.or,
             style: TextStyle(
-              color: Colors.white.withOpacity(0.7),
+              color: Colors.white.withValues(alpha: 0.7),
               fontSize: 14,
             ),
+              );
+            },
           ),
         ),
         Expanded(
           child: Container(
             height: 1,
-            color: Colors.white.withOpacity(0.3),
+            color: Colors.white.withValues(alpha: 0.3),
           ),
         ),
       ],
@@ -447,6 +407,7 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
 
 
   Widget _buildEmailLoginOption() {
+    final l10n = AppLocalizations.of(context);
     return Column(
       children: [
         SizedBox(
@@ -458,19 +419,19 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
               _showEmailLoginDialog();
             },
             icon: const Icon(Icons.email, size: 20, color: Colors.white),
-            label: const Text(
-              'התחבר עם שכונתי',
-              style: TextStyle(
+            label: Text(
+              l10n.loginWithShchunati,
+              style: const TextStyle(
                 fontSize: 14,
                 fontWeight: FontWeight.w600,
                 color: Colors.white,
               ),
             ),
             style: ElevatedButton.styleFrom(
-              backgroundColor: Colors.purple.withOpacity(0.8),
+              backgroundColor: Colors.purple.withValues(alpha: 0.8),
               foregroundColor: Colors.white,
               elevation: 6,
-              shadowColor: Colors.purple.withOpacity(0.3),
+              shadowColor: Colors.purple.withValues(alpha: 0.3),
               shape: RoundedRectangleBorder(
                 borderRadius: BorderRadius.circular(12),
               ),
@@ -481,182 +442,198 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
     );
   }
 
+  Widget _buildContinueWithoutRegistrationButton() {
+    final l10n = AppLocalizations.of(context);
+    return SizedBox(
+      width: double.infinity,
+      height: 48,
+      child: TextButton(
+        onPressed: _isLoading ? null : () async {
+          await playButtonSound();
+          await _handleContinueWithoutRegistration();
+        },
+        style: TextButton.styleFrom(
+          foregroundColor: Colors.white,
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(12),
+            side: BorderSide(color: Colors.white.withValues(alpha: 0.5), width: 1),
+          ),
+        ),
+        child: Text(
+          l10n.continueWithoutRegistration,
+          style: const TextStyle(
+            fontSize: 14,
+            fontWeight: FontWeight.w600,
+            color: Colors.white,
+          ),
+        ),
+      ),
+    );
+  }
 
-  Widget _buildTermsAndPrivacyLinks() {
+  /// טיפול בכניסה ללא הרשמה - יצירת משתמש אורח עם email וסיסמה זמניים
+  Future<void> _handleContinueWithoutRegistration() async {
+    if (_isLoading) return;
+    
+    setState(() {
+      _isLoading = true;
+    });
+
+    try {
+      // יצירת email וסיסמה זמניים ייחודיים
+      final timestamp = DateTime.now().millisecondsSinceEpoch;
+      final randomId = timestamp.toString();
+      final tempEmail = 'guest_$randomId@temp.shchunati.com';
+      final tempPassword = 'temp_${timestamp}_${math.Random().nextInt(10000)}';
+      
+      debugPrint('🔐 Creating temporary guest user with email: $tempEmail');
+
+      // יצירת משתמש עם email וסיסמה זמניים
+      final userCredential = await FirebaseAuth.instance.createUserWithEmailAndPassword(
+        email: tempEmail,
+        password: tempPassword,
+      );
+      
+      final user = userCredential.user;
+      
+      if (user == null) {
+        throw Exception('לא ניתן ליצור משתמש אורח');
+      }
+
+      debugPrint('✅ Guest user created: ${user.uid}');
+
+      // יצירת פרופיל אורח אוטומטי
+      final now = DateTime.now();
+      final trialEndDate = now.add(const Duration(days: 30));
+      
+      final guestProfile = {
+        'userId': user.uid,
+        'displayName': 'אורח',
+        'email': tempEmail,
+        'userType': 'guest',
+        'createdAt': Timestamp.fromDate(now),
+        'isSubscriptionActive': true,
+        'subscriptionStatus': 'guest_trial',
+        'businessCategories': [],
+        'guestTrialStartDate': Timestamp.fromDate(now),
+        'guestTrialEndDate': Timestamp.fromDate(trialEndDate),
+        'maxRequestsPerMonth': 10,
+        'maxRadius': 3.0,
+        'canCreatePaidRequests': true,
+        'hasAcceptedTerms': true,
+        'isTemporaryGuest': true, // סימון שזה משתמש זמני
+      };
+
+      // שמירת הפרופיל ב-Firestore
+      await FirebaseFirestore.instance
+          .collection('users')
+          .doc(user.uid)
+          .set(guestProfile);
+
+      debugPrint('✅ Guest profile created for temporary user');
+
+      // הצגת הודעת הצלחה
+      if (mounted) {
+        await playSuccessSound();
+        widget.onLoginSuccess?.call();
+      }
+    } catch (e) {
+      debugPrint('❌ Error in continue without registration: $e');
+      if (mounted) {
+        ScaffoldMessenger.of(context).showSnackBar(
+          SnackBar(
+            content: Text('שגיאה בכניסה ללא הרשמה: $e'),
+            backgroundColor: Colors.red,
+            duration: const Duration(seconds: 3),
+          ),
+        );
+      }
+    } finally {
+      if (mounted) {
+        setState(() {
+          _isLoading = false;
+        });
+      }
+    }
+  }
+
+  Widget _buildLegalLinks() {
+    final l10n = AppLocalizations.of(context);
     return Container(
       margin: const EdgeInsets.symmetric(horizontal: 20),
       child: Column(
         children: [
           Text(
-            'על ידי המשך השימוש באפליקציה, אתה מסכים ל:',
+            l10n.byContinuingYouAgree,
             textAlign: TextAlign.center,
             style: TextStyle(
               fontSize: 12,
-              color: Colors.white.withOpacity(0.8),
+              color: Colors.white.withValues(alpha: 0.8),
             ),
           ),
           const SizedBox(height: 8),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
+          Wrap(
+            alignment: WrapAlignment.center,
+            spacing: 8,
+            runSpacing: 4,
             children: [
-              GestureDetector(
-                onTap: () => _showTermsAndPrivacyDialog(),
-                child: Text(
-                  'תנאי שימוש',
-                  style: TextStyle(
-                    fontSize: 12,
-                    color: Colors.blue[300],
-                    decoration: TextDecoration.underline,
-                    fontWeight: FontWeight.w500,
+              _buildLegalLink(
+                l10n.termsAndPrivacyButton,
+                () => Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                    builder: (context) => TermsAndPrivacyScreen(
+                      onAccept: () {},
+                      onDecline: () {},
+                      readOnly: true, // קריאה בלבד - לא להציג לחצנים
+                    ),
                   ),
                 ),
               ),
               Text(
-                ' ו-',
+                ' • ',
                 style: TextStyle(
                   fontSize: 12,
-                  color: Colors.white.withOpacity(0.8),
+                  color: Colors.white.withValues(alpha: 0.6),
                 ),
               ),
-              GestureDetector(
-                onTap: () => _showTermsAndPrivacyDialog(),
-                child: Text(
-                  'מדיניות פרטיות',
-                  style: TextStyle(
-                    fontSize: 12,
-                    color: Colors.blue[300],
-                    decoration: TextDecoration.underline,
-                    fontWeight: FontWeight.w500,
+              _buildLegalLink(
+                l10n.aboutButton,
+                () => Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                    builder: (context) => const AboutAppScreen(),
                   ),
                 ),
               ),
             ],
           ),
-        ],
-      ),
-    );
-  }
-
-  void _showTermsAndPrivacyDialog() {
-    showDialog(
-      context: context,
-      builder: (context) => AlertDialog(
-        title: const Text('תנאי שימוש ומדיניות פרטיות'),
-        content: SizedBox(
-          width: double.maxFinite,
-          height: 400,
-          child: SingleChildScrollView(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  'תנאי שימוש',
-                  style: Theme.of(context).textTheme.headlineSmall?.copyWith(
-                        fontWeight: FontWeight.bold,
-                        color: Theme.of(context).colorScheme.primary,
-                      ),
-                ),
-                const SizedBox(height: 10),
-                const Text(
-                  'ברוכים הבאים לאפליקציית "שכונתי". השימוש באפליקציה כפוף לתנאים הבאים. אנא קרא אותם בעיון:',
-                  style: TextStyle(fontSize: 14),
-                ),
-                const SizedBox(height: 10),
-                _buildBulletPoint(
-                    'גיל מינימלי: השימוש באפליקציה מיועד למשתמשים מעל גיל 18 בלבד. החברה שומרת על זכותה לבקש הוכחת גיל בכל שלב.'),
-                _buildBulletPoint(
-                    'עזרה הדדית: האפליקציה מיועדת לעזרה הדדית בין שכנים - חיבור בין מבקשי עזרה לנותני עזרה בקהילה המקומית.'),
-                _buildBulletPoint(
-                    'אחריות המשתמש: המשתמשים אחראים באופן בלעדי לתוכן שהם מפרסמים ולכל אינטראקציה עם משתמשים אחרים. שכונתי היא מתווכת בלבד.'),
-                _buildBulletPoint(
-                    'בטיחות: במקרה של חשד לסכנה או התנהגות לא הולמת, יש לדווח מיד לתמיכה או לרשויות הרלוונטיות.'),
-                _buildBulletPoint(
-                    'פרטיות: אנו מחויבים להגן על פרטיותך. המידע האישי שלך ישמש בהתאם למדיניות הפרטיות שלנו.'),
-                _buildBulletPoint(
-                    'הגבלת אחריות: האפליקציה ניתנת "כמות שהיא" (AS IS). אנו לא מתחייבים לזמינות רציפה, ללא תקלות או ללא שגיאות.'),
-                _buildBulletPoint(
-                    'ביטול בקשות: ניתן לבטל בקשה עד 30 דקות ממועד הפרסום. ביטול מאוחר יותר מותנה בהסכמת נותן השירות.'),
-                _buildBulletPoint(
-                    'שינויים בתנאים: אנו שומרים לעצמנו את הזכות לשנות את תנאי השימוש מעת לעת. שינויים ייכנסו לתוקף עם פרסומם באפליקציה.'),
-                const SizedBox(height: 20),
-                Text(
-                  'מדיניות פרטיות',
-                  style: Theme.of(context).textTheme.headlineSmall?.copyWith(
-                        fontWeight: FontWeight.bold,
-                        color: Theme.of(context).colorScheme.primary,
-                      ),
-                ),
-                const SizedBox(height: 10),
-                const Text(
-                  'מדיניות פרטיות זו מתארת כיצד אנו אוספים, משתמשים ומגנים על המידע האישי שלך באפליקציית "שכונתי":',
-                  style: TextStyle(fontSize: 14),
-                ),
-                const SizedBox(height: 10),
-                _buildBulletPoint(
-                    'איסוף מידע: אנו אוספים מידע שאתה מספק לנו ישירות (כגון שם, אימייל, מיקום גיאוגרפי) ומידע שנאסף אוטומטית (כגון נתוני שימוש).'),
-                _buildBulletPoint(
-                    'שימוש במידע: אנו משתמשים במידע כדי לספק, לתחזק ולשפר את השירותים שלנו, להתאים אישית את חווית המשתמש, ולתקשר איתך.'),
-                _buildBulletPoint(
-                    'שיתוף מידע: אנו לא נשתף את המידע האישי שלך עם צדדים שלישיים ללא הסכמתך, למעט במקרים הנדרשים על פי חוק או לצורך מתן השירותים.'),
-                _buildBulletPoint(
-                    'אבטחת מידע: אנו נוקטים באמצעי אבטחה סבירים כדי להגן על המידע שלך מפני גישה בלתי מורשית, שימוש או חשיפה. מיקום גיאוגרפי נשמר באופן מוצפן.'),
-                _buildBulletPoint(
-                    'בקרת פרטיות: יש לך בקרה מלאה על מי רואה את המידע שלך. תוכל להגדיר רמות פרטיות שונות עבור בקשות שונות.'),
-                _buildBulletPoint(
-                    'עוגיות וטכנולוגיות דומות: אנו עשויים להשתמש בעוגיות ובטכנולוגיות דומות כדי לשפר את חווית המשתמש ולנתח את השימוש באפליקציה.'),
-                _buildBulletPoint(
-                    'זכויותיך: יש לך זכות לגשת, לתקן או למחוק את המידע האישי שלך. למימוש זכויות אלו, אנא צור קשר איתנו.'),
-                _buildBulletPoint(
-                    'שינויים במדיניות: אנו רשאים לעדכן מדיניות פרטיות זו מעת לעת. שינויים ייכנסו לתוקף עם פרסומם באפליקציה.'),
-                const SizedBox(height: 20),
-                Container(
-                  padding: const EdgeInsets.all(12),
-                  decoration: BoxDecoration(
-                    color: Colors.red[50],
-                    borderRadius: BorderRadius.circular(8),
-                    border: Border.all(color: Colors.red[200]!),
-                  ),
-                  child: Row(
-                    children: [
-                      Icon(Icons.warning, color: Colors.red[600], size: 20),
-                      const SizedBox(width: 8),
-                      Expanded(
-                        child: Text(
-                          'חשוב: עליך לאשר את תנאי השימוש ומדיניות הפרטיות כדי להמשיך להשתמש באפליקציה.',
-                          style: TextStyle(
-                              fontSize: 12, 
-                              fontWeight: FontWeight.bold,
-                              color: Colors.red[800]),
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ],
+          const SizedBox(height: 16),
+          // Copyright
+          Text(
+            l10n.copyright,
+            textAlign: TextAlign.center,
+            style: TextStyle(
+              fontSize: 11,
+              color: Colors.white.withValues(alpha: 0.6),
             ),
           ),
-        ),
-        actions: [
-          TextButton(
-            onPressed: () => Navigator.pop(context),
-            child: const Text('הבנתי'),
-          ),
         ],
       ),
     );
   }
 
-  Widget _buildBulletPoint(String text) {
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 5.0),
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          const Text('• ', style: TextStyle(fontSize: 14)),
-          Expanded(
-            child: Text(text, style: const TextStyle(fontSize: 14)),
-          ),
-        ],
+  Widget _buildLegalLink(String text, VoidCallback onTap) {
+    return GestureDetector(
+      onTap: onTap,
+      child: Text(
+        text,
+        style: TextStyle(
+          fontSize: 12,
+          color: Theme.of(context).colorScheme.primary,
+          decoration: TextDecoration.underline,
+          fontWeight: FontWeight.w500,
+        ),
       ),
     );
   }
@@ -670,8 +647,10 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
     showDialog(
       context: context,
       builder: (context) => StatefulBuilder(
-        builder: (context, setState) => AlertDialog(
-          title: Text(isSignUp ? 'הרשמה' : 'התחברות'),
+        builder: (context, setState) {
+          final l10n = AppLocalizations.of(context);
+          return AlertDialog(
+            title: Text(isSignUp ? l10n.register : l10n.login),
           content: SizedBox(
             width: 300,
             child: SingleChildScrollView(
@@ -680,10 +659,10 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
                 children: [
                 TextField(
                   controller: emailController,
-                  decoration: const InputDecoration(
-                    labelText: 'אימייל',
-                    prefixIcon: Icon(Icons.email),
-                    border: OutlineInputBorder(),
+                      decoration: InputDecoration(
+                        labelText: l10n.email,
+                        prefixIcon: const Icon(Icons.email),
+                        border: const OutlineInputBorder(),
                   ),
                   keyboardType: TextInputType.emailAddress,
                 ),
@@ -691,7 +670,7 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
                 TextField(
                   controller: passwordController,
                   decoration: InputDecoration(
-                    labelText: 'סיסמה',
+                        labelText: l10n.password,
                     prefixIcon: const Icon(Icons.lock),
                     suffixIcon: IconButton(
                       icon: Icon(obscurePassword ? Icons.visibility : Icons.visibility_off),
@@ -708,7 +687,7 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
                       value: isSignUp,
                       onChanged: (value) => setState(() => isSignUp = value ?? false),
                     ),
-                    const Text('הרשמה חדשה'),
+                        Text(l10n.newRegistration),
                   ],
                 ),
                 if (!isSignUp) ...[
@@ -720,9 +699,9 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
                         Navigator.pop(context);
                         _showForgotPasswordDialog();
                       },
-                      child: const Text(
-                        'שכחתי סיסמה',
-                        style: TextStyle(
+                          child: Text(
+                            l10n.forgotPassword,
+                            style: const TextStyle(
                           color: Colors.blue,
                           decoration: TextDecoration.underline,
                         ),
@@ -737,7 +716,30 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
           actions: [
             TextButton(
               onPressed: () => Navigator.pop(context),
-              child: const Text('ביטול'),
+                child: Text(l10n.cancel),
+            ),
+            // ✅ לחצן "התחבר ללא אימות" - מופיע גם בהתחברות וגם בהרשמה
+            TextButton(
+              onPressed: () async {
+                // סגירת הדיאלוג לפני ההתחברות
+                if (mounted) {
+                  Navigator.pop(context);
+                }
+                if (isSignUp) {
+                  // הרשמה ללא אימות
+                  await _handleEmailLoginWithoutVerification(
+                    emailController.text,
+                    passwordController.text,
+                  );
+                } else {
+                  // התחברות ללא אימות - ננסה להתחבר בלי לבדוק אימות אימייל
+                  await _handleEmailLoginWithoutVerificationForExistingUser(
+                    emailController.text,
+                    passwordController.text,
+                  );
+                }
+              },
+              child: Text(l10n.loginWithoutVerification),
             ),
             ElevatedButton(
               onPressed: () async {
@@ -751,10 +753,11 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
                   isSignUp,
                 );
               },
-              child: Text(isSignUp ? 'הרשמה' : 'התחברות'),
+                child: Text(isSignUp ? l10n.register : l10n.login),
             ),
           ],
-        ),
+          );
+        },
       ),
     );
   }
@@ -764,23 +767,25 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
 
     showDialog(
       context: context,
-      builder: (context) => AlertDialog(
-        title: const Text('שכחתי סיסמה'),
+      builder: (context) {
+        final l10n = AppLocalizations.of(context);
+        return AlertDialog(
+          title: Text(l10n.forgotPassword),
         content: SingleChildScrollView(
           child: Column(
             mainAxisSize: MainAxisSize.min,
             children: [
-            const Text(
-              'הזן את כתובת האימייל שלך ונשלח לך קישור לאיפוס הסיסמה:',
-              style: TextStyle(fontSize: 14),
+                Text(
+                  l10n.pleaseEnterEmail,
+                  style: const TextStyle(fontSize: 14),
             ),
             const SizedBox(height: 16),
             TextField(
               controller: emailController,
-              decoration: const InputDecoration(
-                labelText: 'אימייל',
-                prefixIcon: Icon(Icons.email),
-                border: OutlineInputBorder(),
+                  decoration: InputDecoration(
+                    labelText: l10n.email,
+                    prefixIcon: const Icon(Icons.email),
+                    border: const OutlineInputBorder(),
                 hintText: 'example@email.com',
               ),
               keyboardType: TextInputType.emailAddress,
@@ -789,20 +794,20 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
             Container(
               padding: const EdgeInsets.all(12),
               decoration: BoxDecoration(
-                color: Colors.orange[50],
+                color: Theme.of(context).colorScheme.tertiaryContainer,
                 borderRadius: BorderRadius.circular(8),
-                border: Border.all(color: Colors.orange[200]!),
+                border: Border.all(color: Theme.of(context).colorScheme.tertiary),
               ),
               child: Row(
                 children: [
-                  Icon(Icons.warning, color: Colors.orange[600], size: 20),
+                  Icon(Icons.warning, color: Theme.of(context).colorScheme.tertiary, size: 20),
                   const SizedBox(width: 8),
                   Expanded(
                     child: Text(
-                      'וודא שהאימייל שייך לך! אם תזין אימייל של מישהו אחר, הוא יקבל את קישור איפוס הסיסמה.',
+                      l10n.verifyEmailBelongsToYou,
                       style: TextStyle(
                         fontSize: 12,
-                        color: Colors.orange[700],
+                        color: Theme.of(context).colorScheme.onTertiaryContainer,
                       ),
                     ),
                   ),
@@ -815,14 +820,14 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
         actions: [
           TextButton(
             onPressed: () => Navigator.pop(context),
-            child: const Text('ביטול'),
+              child: Text(l10n.cancel),
           ),
           ElevatedButton(
             onPressed: () async {
               if (emailController.text.isEmpty) {
                 ScaffoldMessenger.of(context).showSnackBar(
-                  const SnackBar(
-                    content: Text('אנא הזן כתובת אימייל'),
+                    SnackBar(
+                      content: Text(l10n.pleaseEnterEmail),
                     backgroundColor: Colors.red,
                   ),
                 );
@@ -832,23 +837,26 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
               Navigator.pop(context);
               await _handleForgotPassword(emailController.text);
             },
-            child: const Text('שלח קישור'),
+              child: Text(l10n.sendLink),
           ),
         ],
-      ),
+        );
+      },
     );
   }
 
   Future<void> _handleForgotPassword(String email) async {
+    if (!mounted) return;
     try {
       setState(() => _isLoading = true);
       
       await FirebaseAuth.instance.sendPasswordResetEmail(email: email);
       
       if (mounted) {
+        final l10n = AppLocalizations.of(context);
         ScaffoldMessenger.of(context).showSnackBar(
           SnackBar(
-            content: Text('קישור לאיפוס סיסמה נשלח ל-$email'),
+            content: Text(l10n.passwordResetLinkSentTo(email)),
             backgroundColor: Colors.green,
             duration: const Duration(seconds: 5),
           ),
@@ -880,6 +888,200 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
     }
   }
 
+  // ✅ התחברות למשתמש קיים ללא אימות אימייל
+  Future<void> _handleEmailLoginWithoutVerificationForExistingUser(String email, String password) async {
+    if (email.isEmpty || password.isEmpty) {
+      ScaffoldMessenger.of(context).showSnackBar(
+        const SnackBar(
+          content: Text('אנא מלא את כל השדות'),
+          backgroundColor: Colors.red,
+        ),
+      );
+      return;
+    }
+
+    if (!mounted) return;
+    setState(() => _isLoading = true);
+    try {
+      // ניסיון להתחבר למשתמש קיים
+      final cred = await FirebaseAuth.instance.signInWithEmailAndPassword(
+        email: email,
+        password: password,
+      );
+      
+      debugPrint('✅ User signed in: ${cred.user!.uid}');
+      
+      // ✅ לא בודקים אימות אימייל - מכניסים את המשתמש ישירות
+      // טעינת הפרופיל
+      final userDoc = await FirebaseFirestore.instance
+          .collection('users')
+          .doc(cred.user!.uid)
+          .get();
+      
+      if (userDoc.exists) {
+        debugPrint('✅ User profile found - logged in without email verification');
+        await playSuccessSound();
+        
+        if (mounted) {
+          widget.onLoginSuccess?.call();
+        }
+      } else {
+        debugPrint('⚠️ User profile not found');
+        await FirebaseAuth.instance.signOut();
+        if (mounted) {
+          ScaffoldMessenger.of(context).showSnackBar(
+            const SnackBar(
+              content: Text('פרופיל המשתמש לא נמצא'),
+              backgroundColor: Colors.red,
+            ),
+          );
+        }
+      }
+    } on FirebaseAuthException catch (e) {
+      debugPrint('❌ Error in login without verification: $e');
+      if (mounted) {
+        String errorMessage = 'שגיאה בהתחברות';
+        if (e.code == 'user-not-found') {
+          errorMessage = 'משתמש לא נמצא. אנא הירשם תחילה.';
+        } else if (e.code == 'wrong-password') {
+          errorMessage = 'סיסמה שגויה';
+        } else if (e.code == 'invalid-email') {
+          errorMessage = 'כתובת אימייל לא תקינה';
+        }
+        ScaffoldMessenger.of(context).showSnackBar(
+          SnackBar(
+            content: Text(errorMessage),
+            backgroundColor: Colors.red,
+          ),
+        );
+      }
+    } catch (e) {
+      debugPrint('❌ Error in login without verification: $e');
+      if (mounted) {
+        ScaffoldMessenger.of(context).showSnackBar(
+          SnackBar(
+            content: Text('שגיאה: $e'),
+            backgroundColor: Colors.red,
+          ),
+        );
+      }
+    } finally {
+      if (mounted) {
+        setState(() => _isLoading = false);
+      }
+    }
+  }
+
+  // ✅ התחברות/הרשמה ללא אימות אימייל - פשוט לרשום בפיירבייס ולהכניס
+  Future<void> _handleEmailLoginWithoutVerification(String email, String password) async {
+    if (email.isEmpty || password.isEmpty) {
+      ScaffoldMessenger.of(context).showSnackBar(
+        const SnackBar(
+          content: Text('אנא מלא את כל השדות'),
+          backgroundColor: Colors.red,
+        ),
+      );
+      return;
+    }
+
+    if (!mounted) return;
+    setState(() => _isLoading = true);
+    try {
+      // בדיקה אם המשתמש כבר אישר את תנאי השימוש
+      final hasAcceptedTerms = await TermsService.hasUserAcceptedTerms();
+      
+      if (!hasAcceptedTerms) {
+        // הצגת מסך תנאי שימוש ומדיניות פרטיות
+        if (mounted) {
+          final result = await Navigator.push<bool>(
+            context,
+            MaterialPageRoute(
+              builder: (context) => TermsAndPrivacyScreen(
+                onAccept: () async {
+                  await TermsService.acceptTerms();
+                  if (!context.mounted) return;
+                  Navigator.pop(context, true);
+                },
+                onDecline: () {
+                  Navigator.pop(context, false);
+                },
+              ),
+            ),
+          );
+          
+          if (result != true) {
+            if (mounted) {
+              setState(() => _isLoading = false);
+            }
+            return;
+          }
+        }
+      }
+      
+      // יצירת משתמש חדש ב-Firebase Auth
+      final cred = await FirebaseAuth.instance.createUserWithEmailAndPassword(
+        email: email,
+        password: password,
+      );
+      
+      debugPrint('✅ User created in Firebase Auth: ${cred.user!.uid}');
+      
+      // ✅ לא שולחים אימייל אימות - המשתמש נרשם ללא אימות
+      // ✅ לא מתנתקים - המשתמש נכנס ישירות לאפליקציה
+      
+      // יצירת פרופיל משתמש אורח ב-Firestore
+      final now = DateTime.now();
+      final guestTrialEndDate = now.add(const Duration(days: 60));
+      
+      final userData = {
+        'uid': cred.user!.uid,
+        'displayName': email.split('@')[0],
+        'email': email,
+        'userType': 'guest',
+        'createdAt': Timestamp.fromDate(now),
+        'isSubscriptionActive': true, // ✅ פעיל גם בלי אימות אימייל
+        'subscriptionStatus': 'active',
+        'emailVerified': false, // לא מאומת אבל יכול להשתמש באפליקציה
+        'accountStatus': 'active',
+        'guestTrialStartDate': Timestamp.fromDate(now),
+        'guestTrialEndDate': Timestamp.fromDate(guestTrialEndDate),
+        'maxRequestsPerMonth': 10,
+        'maxRadius': 3.0,
+        'canCreatePaidRequests': true, // ✅ יכול גם בלי אימות אימייל
+        'businessCategories': [],
+        'hasAcceptedTerms': true,
+      };
+      
+      await FirebaseFirestore.instance
+          .collection('users')
+          .doc(cred.user!.uid)
+          .set(userData);
+      
+      debugPrint('✅ User registered without email verification - logged in immediately');
+      
+      await playSuccessSound();
+      
+      // ✅ המשתמש נשאר מחובר ונכנס ישירות לאפליקציה
+      if (mounted) {
+        widget.onLoginSuccess?.call();
+      }
+    } catch (e) {
+      debugPrint('❌ Error in login without verification: $e');
+      if (mounted) {
+        ScaffoldMessenger.of(context).showSnackBar(
+          SnackBar(
+            content: Text('שגיאה: $e'),
+            backgroundColor: Colors.red,
+          ),
+        );
+      }
+    } finally {
+      if (mounted) {
+        setState(() => _isLoading = false);
+      }
+    }
+  }
+
   Future<void> _handleEmailLogin(String email, String password, bool isSignUp) async {
     if (email.isEmpty || password.isEmpty) {
       ScaffoldMessenger.of(context).showSnackBar(
@@ -891,6 +1093,7 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
       return;
     }
 
+    if (!mounted) return;
     setState(() => _isLoading = true);
     try {
       if (isSignUp) {
@@ -907,6 +1110,8 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
                   onAccept: () async {
                     // שמירת אישור התנאים
                     await TermsService.acceptTerms();
+                    // Guard context usage after async gap - check context.mounted for builder context
+                    if (!context.mounted) return;
                     Navigator.pop(context, true);
                   },
                   onDecline: () {
@@ -919,19 +1124,74 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
             
             // אם המשתמש לא אישר את התנאים, לא נמשיך
             if (result != true) {
+              if (mounted) {
+                setState(() => _isLoading = false);
+              }
               return;
             }
           }
         }
         
+        // Firebase Auth יכשל אוטומטית אם האימייל כבר קיים
         final cred = await FirebaseAuth.instance.createUserWithEmailAndPassword(
           email: email,
           password: password,
         );
         
+        // שמירת האימייל והסיסמה לשליחה מחדש
+        _pendingVerificationEmail = email;
+        _pendingVerificationPassword = password;
+        
+        // בדיקה אם האימייל כבר מאומת (למקרה ש-Firebase Auth מאמת אוטומטית)
+        await cred.user!.reload();
+        final isEmailVerified = cred.user!.emailVerified;
+        
+        if (isEmailVerified) {
+          // האימייל כבר מאומת - לא צריך לשלוח אימייל אימות
+          debugPrint('⚠️ Email is already verified - skipping verification email');
+        } else {
+          // שליחת אימייל אימות מותאם אישית דרך Cloud Function
+          try {
+            final emailSent = await CloudFunctionService.sendCustomVerificationEmail(
+              email: email,
+              userId: cred.user!.uid,
+              password: password, // שולח את הסיסמה לאימייל
+            );
+            
+            if (emailSent) {
+              debugPrint('✅ Custom verification email sent to: $email');
+            } else {
+              // אם Cloud Function נכשל, נשתמש ב-Firebase Auth כגיבוי
+              debugPrint('⚠️ Cloud Function failed, using Firebase Auth as fallback');
+              // זיהוי platform להגדרת actionCodeSettings
+              final isMobile = !kIsWeb && (Platform.isAndroid || Platform.isIOS);
+              final actionCodeSettings = ActionCodeSettings(
+                url: 'https://nearme-970f3.web.app/?email=${Uri.encodeComponent(email)}',
+                handleCodeInApp: isMobile, // true אם mobile, false אם web
+                androidPackageName: isMobile && Platform.isAndroid ? 'com.example.flutter1' : null,
+                iOSBundleId: isMobile && Platform.isIOS ? 'com.example.flutter1' : null,
+              );
+              await cred.user!.sendEmailVerification(actionCodeSettings);
+            }
+          } catch (e) {
+            // אם יש שגיאה, נשתמש ב-Firebase Auth כגיבוי
+            debugPrint('⚠️ Error sending custom email, using Firebase Auth: $e');
+            
+            // זיהוי platform להגדרת actionCodeSettings
+            final isMobile = !kIsWeb && (Platform.isAndroid || Platform.isIOS);
+            final actionCodeSettings = ActionCodeSettings(
+              url: 'https://nearme-970f3.web.app/?email=${Uri.encodeComponent(email)}',
+              handleCodeInApp: isMobile, // true אם mobile, false אם web
+              androidPackageName: isMobile && Platform.isAndroid ? 'com.example.flutter1' : null,
+              iOSBundleId: isMobile && Platform.isIOS ? 'com.example.flutter1' : null,
+            );
+            await cred.user!.sendEmailVerification(actionCodeSettings);
+          }
+        }
+        
         // יצירת פרופיל משתמש אורח ב-Firestore
         final now = DateTime.now();
-        final guestTrialEndDate = now.add(const Duration(days: 30));
+        final guestTrialEndDate = now.add(const Duration(days: 60)); // 60 יום תקופת ניסיון
         
         final userData = {
           'uid': cred.user!.uid,
@@ -939,15 +1199,15 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
           'email': email,
           'userType': 'guest', // כל המשתמשים החדשים נרשמים כאורחים
           'createdAt': Timestamp.fromDate(now),
-          'isSubscriptionActive': true, // תקופת אורח פעילה
-          'subscriptionStatus': 'active',
-          'emailVerified': true, // זמנית - ללא אימות
-          'accountStatus': 'active', // סטטוס חשבון פעיל
+          'isSubscriptionActive': isEmailVerified, // פעיל רק אם האימייל מאומת
+          'subscriptionStatus': isEmailVerified ? 'active' : 'pending_verification', // ממתין לאימות אימייל
+          'emailVerified': isEmailVerified, // שמירת הסטטוס האמיתי
+          'accountStatus': isEmailVerified ? 'active' : 'pending_verification', // ממתין לאימות אימייל
           'guestTrialStartDate': Timestamp.fromDate(now),
           'guestTrialEndDate': Timestamp.fromDate(guestTrialEndDate),
           'maxRequestsPerMonth': 10, // גבוה יותר לאורחים
           'maxRadius': 3.0, // 3 ק"מ לאורחים
-          'canCreatePaidRequests': true, // אורחים יכולים ליצור בקשות בתשלום
+          'canCreatePaidRequests': isEmailVerified, // יכול רק אם האימייל מאומת
           'businessCategories': [], // יבחרו במסך הבא
           'hasAcceptedTerms': true,
         };
@@ -957,30 +1217,371 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
             .doc(cred.user!.uid)
             .set(userData);
         
+        // אם האימייל לא מאומת - התנתקות מהמשתמש עד שיאמת את האימייל
+        if (!isEmailVerified) {
+          await FirebaseAuth.instance.signOut();
+          debugPrint('🔒 User signed out - waiting for email verification');
+        } else {
+          debugPrint('✅ Email already verified - user can continue');
+        }
+        
         await playSuccessSound();
         
-        // מעבר למסך בחירת קטגוריות
-        if (mounted) {
-          Navigator.pushReplacement(
-            context,
-            MaterialPageRoute(
-                builder: (context) => GuestCategorySelectionScreen(
-                  displayName: userData['displayName'] as String,
-                  email: userData['email'] as String,
+        // אם האימייל לא מאומת - הצגת הודעה למשתמש שהוא צריך לאמת את האימייל
+        if (!isEmailVerified && mounted) {
+          await showDialog(
+            context: context,
+            barrierDismissible: false,
+            builder: (context) => AlertDialog(
+              title: const Text('אימות אימייל נדרש 📧'),
+              content: Column(
+                mainAxisSize: MainAxisSize.min,
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  const Text(
+                    'שלחנו לך אימייל אימות לכתובת:',
+                    style: TextStyle(fontSize: 16),
+                  ),
+                  const SizedBox(height: 8),
+                  Text(
+                    email,
+                    style: TextStyle(
+                      fontSize: 16,
+                      fontWeight: FontWeight.bold,
+                      color: Theme.of(context).colorScheme.primary,
+                    ),
+                  ),
+                  const SizedBox(height: 16),
+                  const Text(
+                    'אנא פתח את האימייל ולחץ על הקישור לאימות.\n'
+                    'לאחר האימות תוכל להתחבר לאפליקציה.',
+                    style: TextStyle(fontSize: 14),
+                  ),
+                  const SizedBox(height: 16),
+                  TextButton(
+                    onPressed: () async {
+                      // שליחה מחדש של אימייל אימות
+                      try {
+                        if (_pendingVerificationEmail != null && _pendingVerificationPassword != null) {
+                          // התחברות זמנית לשליחת אימייל
+                          final tempCred = await FirebaseAuth.instance.signInWithEmailAndPassword(
+                            email: _pendingVerificationEmail!,
+                            password: _pendingVerificationPassword!,
+                          );
+                          // שליחת אימייל אימות מותאם אישית דרך Cloud Function
+                          try {
+                            final emailSent = await CloudFunctionService.sendCustomVerificationEmail(
+                              email: _pendingVerificationEmail!,
+                              userId: tempCred.user!.uid,
+                              password: _pendingVerificationPassword!, // שולח את הסיסמה לאימייל
+                            );
+                            
+                            if (!emailSent) {
+                              // אם Cloud Function נכשל, נשתמש ב-Firebase Auth כגיבוי
+                              final isMobile = !kIsWeb && (Platform.isAndroid || Platform.isIOS);
+                              final actionCodeSettings = ActionCodeSettings(
+                                url: 'https://nearme-970f3.web.app/?email=${Uri.encodeComponent(_pendingVerificationEmail!)}',
+                                handleCodeInApp: isMobile,
+                                androidPackageName: isMobile && Platform.isAndroid ? 'com.example.flutter1' : null,
+                                iOSBundleId: isMobile && Platform.isIOS ? 'com.example.flutter1' : null,
+                              );
+                              await tempCred.user!.sendEmailVerification(actionCodeSettings);
+                            }
+                          } catch (e) {
+                            // אם יש שגיאה, נשתמש ב-Firebase Auth כגיבוי
+                            final isMobile = !kIsWeb && (Platform.isAndroid || Platform.isIOS);
+                            final actionCodeSettings = ActionCodeSettings(
+                              url: 'https://nearme-970f3.web.app/?email=${Uri.encodeComponent(_pendingVerificationEmail!)}',
+                              handleCodeInApp: isMobile,
+                              androidPackageName: isMobile && Platform.isAndroid ? 'com.example.flutter1' : null,
+                              iOSBundleId: isMobile && Platform.isIOS ? 'com.example.flutter1' : null,
+                            );
+                            await tempCred.user!.sendEmailVerification(actionCodeSettings);
+                          }
+                          await FirebaseAuth.instance.signOut();
+                          
+                          if (context.mounted) {
+                            ScaffoldMessenger.of(context).showSnackBar(
+                              const SnackBar(
+                                content: Text('אימייל אימות נשלח מחדש!'),
+                                backgroundColor: Colors.green,
+                              ),
+                            );
+                          }
+                        }
+                      } catch (e) {
+                        if (context.mounted) {
+                          ScaffoldMessenger.of(context).showSnackBar(
+                            SnackBar(
+                              content: Text('שגיאה בשליחת אימייל: $e'),
+                              backgroundColor: Colors.red,
+                            ),
+                          );
+                        }
+                      }
+                    },
+                    child: const Text('שלח אימייל אימות מחדש'),
+                  ),
+                ],
+              ),
+              actions: [
+                TextButton(
+                  onPressed: () => Navigator.of(context).pop(),
+                  child: const Text('הבנתי'),
                 ),
+              ],
             ),
           );
+          
+          return; // לא ממשיכים להתחברות עד אימות אימייל
+        } else if (isEmailVerified && mounted) {
+          // האימייל מאומת - המשך ישירות למסך הבית
+          widget.onLoginSuccess?.call();
           return;
         }
       } else {
-        await FirebaseAuth.instance.signInWithEmailAndPassword(
+        // התחברות - בדיקת אימות אימייל
+        final cred = await FirebaseAuth.instance.signInWithEmailAndPassword(
           email: email,
           password: password,
         );
         
-        // המשך ישירות ללא בדיקת אימות אימייל (זמנית)
+        // בדיקה אם האימייל מאומת
+        await cred.user!.reload(); // רענון נתוני המשתמש
+        final currentUser = cred.user!;
+        
+        // ✅ בדיקה אם המשתמש נרשם ללא אימות (emailVerified: false ב-Firestore)
+        final userDoc = await FirebaseFirestore.instance
+            .collection('users')
+            .doc(currentUser.uid)
+            .get();
+        
+        final userData = userDoc.data();
+        final emailVerifiedInFirestore = userData?['emailVerified'] as bool?;
+        
+        // אם המשתמש נרשם ללא אימות (emailVerified: false), נכניס אותו ישירות
+        final shouldSkipVerification = emailVerifiedInFirestore == false;
+        
+        if (!shouldSkipVerification && !currentUser.emailVerified) {
+          // האימייל לא מאומת - הצגת הודעה
+          await FirebaseAuth.instance.signOut(); // התנתקות
+          
+          // שמירת האימייל והסיסמה לשליחה מחדש
+          final loginEmail = email;
+          final loginPassword = password;
+          
+          if (mounted) {
+            await showDialog(
+              context: context,
+              barrierDismissible: false,
+              builder: (context) => AlertDialog(
+                title: Row(
+                  children: [
+                    const Icon(Icons.email, color: Colors.blue, size: 28),
+                    const SizedBox(width: 8),
+                    const Text('אימות אימייל נדרש 📧'),
+                  ],
+                ),
+                content: SingleChildScrollView(
+                  child: Column(
+                    mainAxisSize: MainAxisSize.min,
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      Container(
+                        padding: const EdgeInsets.all(12),
+                        decoration: BoxDecoration(
+                          color: Theme.of(context).colorScheme.primaryContainer.withOpacity(0.3),
+                          borderRadius: BorderRadius.circular(8),
+                          border: Border.all(
+                            color: Theme.of(context).colorScheme.primary.withOpacity(0.3),
+                          ),
+                        ),
+                        child: Column(
+                          crossAxisAlignment: CrossAxisAlignment.start,
+                          children: [
+                            const Text(
+                              'עליך לאמת את האימייל שלך לפני שתוכל להתחבר.',
+                              style: TextStyle(
+                                fontSize: 16,
+                                fontWeight: FontWeight.bold,
+                              ),
+                            ),
+                            const SizedBox(height: 8),
+                            const Text(
+                              'אנא פתח את האימייל שנשלח לך ולחץ על הקישור לאימות.',
+                              style: TextStyle(fontSize: 14),
+                            ),
+                          ],
+                        ),
+                      ),
+                      const SizedBox(height: 16),
+                      Row(
+                        children: [
+                          const Icon(Icons.email_outlined, size: 18, color: Colors.grey),
+                          const SizedBox(width: 8),
+                          Expanded(
+                            child: Text(
+                              'כתובת האימייל: $loginEmail',
+                              style: TextStyle(
+                                fontSize: 14,
+                                fontWeight: FontWeight.bold,
+                                color: Theme.of(context).colorScheme.primary,
+                              ),
+                            ),
+                          ),
+                        ],
+                      ),
+                      const SizedBox(height: 20),
+                      Container(
+                        padding: const EdgeInsets.all(12),
+                        decoration: BoxDecoration(
+                          color: Theme.of(context).colorScheme.secondaryContainer.withOpacity(0.3),
+                          borderRadius: BorderRadius.circular(8),
+                        ),
+                        child: Row(
+                          children: [
+                            const Icon(Icons.info_outline, size: 20, color: Colors.blue),
+                            const SizedBox(width: 8),
+                            Expanded(
+                              child: Text(
+                                'לא קיבלת את האימייל? לחץ על הכפתור למטה כדי לשלוח אותו שוב.',
+                                style: TextStyle(
+                                  fontSize: 13,
+                                  color: Theme.of(context).colorScheme.onSurfaceVariant,
+                                ),
+                              ),
+                            ),
+                          ],
+                        ),
+                      ),
+                      const SizedBox(height: 16),
+                      SizedBox(
+                        width: double.infinity,
+                        child: ElevatedButton.icon(
+                          onPressed: () async {
+                            Navigator.of(context).pop(); // סגירת הדיאלוג
+                            
+                            // שליחה מחדש של אימייל אימות
+                            try {
+                              setState(() => _isLoading = true);
+                              
+                              // התחברות זמנית לשליחת אימייל
+                              final tempCred = await FirebaseAuth.instance.signInWithEmailAndPassword(
+                                email: loginEmail,
+                                password: loginPassword,
+                              );
+                              
+                              // שליחת אימייל אימות מותאם אישית דרך Cloud Function
+                              // זיהוי פלטפורמה
+                              final isMobile = !kIsWeb && (Platform.isAndroid || Platform.isIOS);
+                              
+                              try {
+                                final emailSent = await CloudFunctionService.sendCustomVerificationEmail(
+                                  email: loginEmail,
+                                  userId: tempCred.user!.uid,
+                                  password: loginPassword, // שולח את הסיסמה לאימייל
+                                );
+                                
+                                if (!emailSent) {
+                                  // אם Cloud Function נכשל, נשתמש ב-Firebase Auth כגיבוי
+                                  final actionCodeSettings = ActionCodeSettings(
+                                    url: 'https://nearme-970f3.web.app/email-verified?email=${Uri.encodeComponent(loginEmail)}',
+                                    handleCodeInApp: isMobile,
+                                    androidPackageName: isMobile && Platform.isAndroid ? 'com.example.flutter1' : null,
+                                    iOSBundleId: isMobile && Platform.isIOS ? 'com.example.flutter1' : null,
+                                  );
+                                  await tempCred.user!.sendEmailVerification(actionCodeSettings);
+                                }
+                              } catch (e) {
+                                debugPrint('Error sending custom email: $e');
+                                // אם יש שגיאה, נשתמש ב-Firebase Auth כגיבוי
+                                final actionCodeSettings = ActionCodeSettings(
+                                  url: 'https://nearme-970f3.web.app/email-verified?email=${Uri.encodeComponent(loginEmail)}',
+                                  handleCodeInApp: isMobile,
+                                  androidPackageName: isMobile && Platform.isAndroid ? 'com.example.flutter1' : null,
+                                  iOSBundleId: isMobile && Platform.isIOS ? 'com.example.flutter1' : null,
+                                );
+                                await tempCred.user!.sendEmailVerification(actionCodeSettings);
+                              }
+                              
+                              await FirebaseAuth.instance.signOut();
+                              
+                              if (mounted) {
+                                setState(() => _isLoading = false);
+                                ScaffoldMessenger.of(context).showSnackBar(
+                                  const SnackBar(
+                                    content: Text('✅ אימייל אימות נשלח מחדש! בדוק את תיבת הדואר הנכנס שלך.'),
+                                    backgroundColor: Colors.green,
+                                    duration: Duration(seconds: 4),
+                                  ),
+                                );
+                              }
+                            } catch (e) {
+                              debugPrint('Error resending verification email: $e');
+                              if (mounted) {
+                                setState(() => _isLoading = false);
+                                ScaffoldMessenger.of(context).showSnackBar(
+                                  SnackBar(
+                                    content: Text('❌ שגיאה בשליחת אימייל: $e'),
+                                    backgroundColor: Colors.red,
+                                    duration: const Duration(seconds: 4),
+                                  ),
+                                );
+                              }
+                            }
+                          },
+                          icon: const Icon(Icons.send, size: 20),
+                          label: const Text('שלח אימייל אימות מחדש'),
+                          style: ElevatedButton.styleFrom(
+                            padding: const EdgeInsets.symmetric(vertical: 12),
+                            backgroundColor: Theme.of(context).colorScheme.primary,
+                            foregroundColor: Colors.white,
+                          ),
+                        ),
+                      ),
+                    ],
+                  ),
+                ),
+                actions: [
+                  TextButton(
+                    onPressed: () => Navigator.of(context).pop(),
+                    child: const Text('הבנתי'),
+                  ),
+                ],
+              ),
+            );
+          }
+          setState(() => _isLoading = false);
+          return; // לא ממשיכים להתחברות
+        }
+        
+        // ✅ אם המשתמש נרשם ללא אימות, נכניס אותו ישירות
+        if (shouldSkipVerification && mounted) {
+          debugPrint('✅ User registered without verification - logging in directly');
+          await playSuccessSound();
+          widget.onLoginSuccess?.call();
+          setState(() => _isLoading = false);
+          return;
+        }
+        
+        // האימייל מאומת - עדכון Firestore
+        await FirebaseFirestore.instance
+            .collection('users')
+            .doc(currentUser.uid)
+            .update({
+          'emailVerified': true,
+          'isSubscriptionActive': true,
+          'subscriptionStatus': 'active',
+          'accountStatus': 'active',
+          'canCreatePaidRequests': true,
+        });
+        
+        // שמירת FCM token למשתמש
+        await PushNotificationService.updateUserToken();
         
         await playSuccessSound();
+        // Guard context usage after async gap
+        if (!mounted) return;
         ScaffoldMessenger.of(context).showSnackBar(
           const SnackBar(
             content: Text('התחברות הושלמה בהצלחה!'),
@@ -1022,10 +1623,91 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
       }
     } catch (e) {
       await playErrorSound();
+      // Guard context usage after async gap
+      if (!mounted) return;
+      
+      debugPrint('❌ Login error: $e');
+      debugPrint('❌ Error type: ${e.runtimeType}');
+      if (e is FirebaseAuthException) {
+        debugPrint('❌ Firebase Auth error code: ${e.code}');
+        debugPrint('❌ Firebase Auth error message: ${e.message}');
+      }
+      
+      final l10n = AppLocalizations.of(context);
+      
+      // טיפול בשגיאות Firebase Auth
+      String errorMessage = l10n.loginError;
+      
+      if (e is FirebaseAuthException) {
+        // שימוש בקוד המדויק של Firebase
+        switch (e.code) {
+          case 'user-not-found':
+          case 'USER_NOT_FOUND':
+            errorMessage = l10n.emailNotRegistered;
+            break;
+          case 'wrong-password':
+          case 'WRONG_PASSWORD':
+            errorMessage = l10n.wrongPassword;
+            break;
+          case 'invalid-credential':
+          case 'INVALID_CREDENTIAL':
+            // Firebase לא מבדיל בין אימייל לא רשום לסיסמה שגויה מטעמי אבטחה
+            // במקרה של permission-denied ב-Firestore, לא ננסה לבדוק את האימייל
+            // ונציג הודעה כללית
+            errorMessage = l10n.emailOrPasswordWrong;
+            break;
+          case 'email-already-in-use':
+          case 'EMAIL_ALREADY_IN_USE':
+            errorMessage = l10n.userAlreadyRegisteredPleaseLogin;
+            break;
+          default:
+            // בדיקה נוספת למקרה שהקוד לא מזוהה
+            final errorString = e.toString().toLowerCase();
+            if (errorString.contains('user-not-found')) {
+              errorMessage = l10n.emailNotRegistered;
+            } else if (errorString.contains('wrong-password') || 
+                       errorString.contains('invalid-credential')) {
+              errorMessage = l10n.wrongPassword;
+            } else if (errorString.contains('email-already-in-use')) {
+              errorMessage = l10n.userAlreadyRegisteredPleaseLogin;
+            }
+        }
+      } else {
+        // בדיקה לגביית שגיאות לא מ-FirebaseAuthException
+        // בדיקה אם זו שגיאת Firestore permission-denied
+        if (e.toString().contains('permission-denied') || 
+            e.toString().contains('PERMISSION_DENIED') ||
+            e.toString().contains('cloud_firestore/permission-denied')) {
+          errorMessage = l10n.loginError; // הודעה כללית במקרה של permission denied
+        } else {
+          final errorString = e.toString().toLowerCase();
+          if (errorString.contains('user-not-found') || 
+              errorString.contains('user_not_found')) {
+            errorMessage = l10n.emailNotRegistered;
+          } else if (errorString.contains('wrong-password') || 
+                     errorString.contains('wrong_password') ||
+                     errorString.contains('invalid-credential') ||
+                     errorString.contains('invalid_credential')) {
+            errorMessage = l10n.wrongPassword;
+          } else if (errorString.contains('email-already-in-use') ||
+                     errorString.contains('email_already_in_use')) {
+            errorMessage = l10n.userAlreadyRegisteredPleaseLogin;
+          }
+        }
+      }
+      
+      // Guard context usage after async gap
+      if (!mounted) return;
+      
+      // אם זו הודעה על משתמש שכבר רשום, נציג אותה בכחול (לא אדום)
+      final isUserAlreadyRegistered = errorMessage == l10n.userAlreadyRegisteredPleaseLogin || 
+                                      errorMessage == l10n.userAlreadyRegistered;
+      
       ScaffoldMessenger.of(context).showSnackBar(
         SnackBar(
-          content: Text('שגיאה: $e'),
-          backgroundColor: Colors.red,
+          content: Text(errorMessage),
+          backgroundColor: isUserAlreadyRegistered ? Colors.blue : Colors.red,
+          duration: isUserAlreadyRegistered ? const Duration(seconds: 4) : const Duration(seconds: 3),
         ),
       );
     } finally {
@@ -1037,10 +1719,15 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
 
 
   Future<void> _handleGoogleLogin() async {
+    if (!mounted) return;
     setState(() => _isLoading = true);
     try {
       final user = await GoogleAuthService.signInWithGoogle();
       if (user != null) {
+        // Firebase Auth כבר מטפל בבדיקת אימייל קיים
+        // אם יש משתמש עם אותו אימייל, Firebase Auth יזרוק שגיאה
+        // נבדוק רק אם המשתמש כבר קיים ב-Firestore לפי UID (אחרי שהמשתמש מאומת)
+        
         await playSuccessSound();
         
         // בדיקה אם המשתמש כבר קיים במערכת
@@ -1052,7 +1739,7 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
         if (!userDoc.exists) {
           // משתמש חדש - יצירת פרופיל אורח
           final now = DateTime.now();
-          final guestTrialEndDate = now.add(const Duration(days: 30));
+          final guestTrialEndDate = now.add(const Duration(days: 60)); // 60 יום תקופת ניסיון
           
           final userData = {
             'uid': user.uid,
@@ -1078,17 +1765,10 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
               .doc(user.uid)
               .set(userData);
           
-          // מעבר למסך בחירת קטגוריות
+          // מעבר ישיר למסך הבית (ללא בחירת קטגוריות)
           if (mounted) {
-            Navigator.pushReplacement(
-              context,
-              MaterialPageRoute(
-                builder: (context) => GuestCategorySelectionScreen(
-                  displayName: userData['displayName'] as String,
-                  email: userData['email'] as String,
-                ),
-              ),
-            );
+            // קריאה לפונקציית ההצלחה
+            widget.onLoginSuccess?.call();
             return;
           }
         } else {
@@ -1105,6 +1785,8 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
                     onAccept: () async {
                       // שמירת אישור התנאים
                       await TermsService.acceptTerms();
+                      // Guard context usage after async gap - check context.mounted for builder context
+                      if (!context.mounted) return;
                       Navigator.pop(context, true);
                     },
                     onDecline: () {
@@ -1118,6 +1800,9 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
               
               // אם המשתמש לא אישר את התנאים, לא נמשיך
               if (result != true) {
+              if (mounted) {
+                setState(() => _isLoading = false);
+              }
                 return;
               }
             }
@@ -1151,9 +1836,10 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
       } else {
         // הצג הודעת שגיאה ידידותית
         if (mounted) {
+          final l10n = AppLocalizations.of(context);
           ScaffoldMessenger.of(context).showSnackBar(
-            const SnackBar(
-              content: Text('התחברות נכשלה - נסה שוב'),
+            SnackBar(
+              content: Text(l10n.loginError),
               backgroundColor: Colors.red,
             ),
           );
@@ -1161,10 +1847,14 @@ class _YokiStyleAuthScreenState extends State<YokiStyleAuthScreen>
       }
     } catch (e) {
       await playErrorSound();
+      debugPrint('❌ Google login error: $e');
+      debugPrint('❌ Error type: ${e.runtimeType}');
+      
       if (mounted) {
+        final l10n = AppLocalizations.of(context);
         ScaffoldMessenger.of(context).showSnackBar(
           SnackBar(
-            content: Text('שגיאה בהתחברות: $e'),
+            content: Text(l10n.loginError),
             backgroundColor: Colors.red,
           ),
         );
