@@ -2,7 +2,9 @@ import 'package:flutter/material.dart';
 import 'yoki_splash_screen.dart';
 
 class SplashScreen extends StatefulWidget {
-  const SplashScreen({super.key});
+  final Function(Locale)? onLanguageSelected;
+  
+  const SplashScreen({super.key, this.onLanguageSelected});
 
   @override
   State<SplashScreen> createState() => _SplashScreenState();
@@ -11,6 +13,8 @@ class SplashScreen extends StatefulWidget {
 class _SplashScreenState extends State<SplashScreen> {
   @override
   Widget build(BuildContext context) {
-    return const YokiSplashScreen();
+    return YokiSplashScreen(
+      onLanguageSelected: widget.onLanguageSelected,
+    );
   }
 }
