@@ -15,6 +15,8 @@ class ChatService {
     required String helperId,
   }) async {
     try {
+      final currentUserId = _auth.currentUser?.uid;
+      
       // בדיקה אם כבר קיים צ'אט בין השניים הספציפיים
       final existingChat = await _firestore
           .collection('chats')
@@ -27,6 +29,39 @@ class ChatService {
         final chatData = doc.data();
         final participants = List<String>.from(chatData['participants'] ?? []);
         if (participants.contains(creatorId) && participants.contains(helperId)) {
+          // בדיקה אם הצ'אט נמחק על ידי המשתמש הנוכחי
+          final deletedBy = List<String>.from(chatData['deletedBy'] ?? []);
+          final isClosed = chatData['isClosed'] as bool? ?? false;
+          
+          if (currentUserId != null && deletedBy.contains(currentUserId)) {
+            // אם הצ'אט נמחק על ידי המשתמש הנוכחי, נפתח אותו מחדש במקום ליצור צ'אט חדש
+            debugPrint('Found existing chat ${doc.id} that was deleted by current user $currentUserId, reopening it...');
+            
+            // פתיחת הצ'אט מחדש - הסרת המשתמש מ-deletedBy ופתיחת הצ'אט
+            await _firestore.collection('chats').doc(doc.id).update({
+              'deletedBy': FieldValue.arrayRemove([currentUserId]),
+              'isClosed': false, // פתיחת הצ'אט מחדש
+              'updatedAt': Timestamp.fromDate(DateTime.now()),
+            });
+            
+            debugPrint('✅ Reopened chat ${doc.id} for user $currentUserId');
+            return doc.id;
+          }
+          
+          // אם הצ'אט סגור אבל לא נמחק על ידי המשתמש הנוכחי, נפתח אותו מחדש
+          if (isClosed && currentUserId != null && !deletedBy.contains(currentUserId)) {
+            debugPrint('Found closed chat ${doc.id}, reopening it...');
+            
+            // פתיחת הצ'אט מחדש
+            await _firestore.collection('chats').doc(doc.id).update({
+              'isClosed': false, // פתיחת הצ'אט מחדש
+              'updatedAt': Timestamp.fromDate(DateTime.now()),
+            });
+            
+            debugPrint('✅ Reopened closed chat ${doc.id}');
+            return doc.id;
+          }
+          
           debugPrint('Found existing chat between $creatorId and $helperId: ${doc.id}');
           return doc.id;
         }
@@ -54,10 +89,13 @@ class ChatService {
     }
   }
 
-  /// שליחת הודעה בצ'אט
+  /// שליחת הודעה בצ'אט (תמיכה בהודעות טקסט, קול ותמונה)
   static Future<bool> sendMessage({
     required String chatId,
     required String text,
+    String? type, // 'text', 'voice', or 'image'
+    String? data, // Base64 string or URL
+    int? duration, // Duration in seconds for voice messages
   }) async {
     try {
       final user = _auth.currentUser;
@@ -72,6 +110,25 @@ class ChatService {
       if (!chatDoc.exists) return false;
       
       final chatData = chatDoc.data()!;
+      
+      // בדיקה אם הצ'אט נמחק על ידי המשתמש הנוכחי
+      final deletedBy = List<String>.from(chatData['deletedBy'] ?? []);
+      if (deletedBy.contains(user.uid)) {
+        debugPrint('⚠️ Chat $chatId was deleted by current user ${user.uid}, cannot send message');
+        return false;
+      }
+      
+      // בדיקה אם הצ'אט סגור - אם כן, נפתח אותו מחדש
+      final isClosed = chatData['isClosed'] as bool? ?? false;
+      if (isClosed) {
+        debugPrint('🔄 Chat $chatId is closed, reopening it...');
+        await _firestore.collection('chats').doc(chatId).update({
+          'isClosed': false, // פתיחת הצ'אט מחדש
+          'updatedAt': Timestamp.fromDate(DateTime.now()),
+        });
+        debugPrint('✅ Reopened closed chat $chatId');
+      }
+      
       final participants = List<String>.from(chatData['participants'] ?? []);
       final requestId = chatData['requestId'] as String? ?? '';
       
@@ -80,12 +137,71 @@ class ChatService {
       
       if (otherParticipants.isEmpty) return false;
 
+      // בדיקה אם זו הודעה ראשונה ממבקש השירות (יוצר הבקשה)
+      // זה מבטיח שהבקשה תופיע בתחילת הרשימה במסך "פניות שלי" של נותן השירות
+      bool shouldUpdateInterestTime = false;
+      String? helperIdForInterestUpdate;
+      if (requestId.isNotEmpty) {
+        try {
+          // קבלת פרטי הבקשה כדי לבדוק מי יוצר הבקשה
+          final requestDoc = await _firestore
+              .collection('requests')
+              .doc(requestId)
+              .get();
+          
+          if (requestDoc.exists) {
+            final requestData = requestDoc.data()!;
+            final creatorId = requestData['createdBy'] as String?;
+            
+            // אם השולח הוא יוצר הבקשה (מבקש השירות), נבדוק אם זו הודעה ראשונה ממנו
+            if (creatorId == user.uid && otherParticipants.isNotEmpty) {
+              // בדיקה אם יש הודעות קודמות ממבקש השירות בצ'אט
+              final messagesFromCreatorSnapshot = await _firestore
+                  .collection('chats')
+                  .doc(chatId)
+                  .collection('messages')
+                  .where('from', isEqualTo: user.uid)
+                  .limit(1)
+                  .get();
+              
+              // אם אין הודעות קודמות ממבקש השירות, זו הודעה ראשונה ממנו
+              if (messagesFromCreatorSnapshot.docs.isEmpty) {
+                shouldUpdateInterestTime = true;
+                helperIdForInterestUpdate = otherParticipants.first; // נותן השירות
+                debugPrint('✅ First message from creator $user.uid to helper $helperIdForInterestUpdate in request $requestId');
+              }
+            }
+          }
+        } catch (e) {
+          debugPrint('⚠️ Failed to check if first message from creator: $e');
+          // נמשיך לשלוח את ההודעה גם אם הבדיקה נכשלה
+        }
+      }
+
+      // Parse message type
+      MessageType messageType = MessageType.text;
+      if (type != null) {
+        switch (type) {
+          case 'voice':
+            messageType = MessageType.voice;
+            break;
+          case 'image':
+            messageType = MessageType.image;
+            break;
+          default:
+            messageType = MessageType.text;
+        }
+      }
+
       final message = Message(
         messageId: '', // יוגדר על ידי Firestore
         from: user.uid,
         text: text,
         sentAt: DateTime.now(),
         isSystemMessage: false,
+        type: messageType,
+        data: data,
+        duration: duration,
       );
 
       // שמירת ההודעה
@@ -103,6 +219,101 @@ class ChatService {
         'lastMessage': text,
         'updatedAt': Timestamp.fromDate(DateTime.now()),
       });
+
+      // עדכון זמן ההתעניינות ואם מבקש השירות שולח הודעה ראשונה, הוספת נותן השירות ל-helpers
+      // זה מבטיח שהבקשה תופיע בתחילת הרשימה במסך "פניות שלי" של נותן השירות
+      // רק אם הבקשה היא "בתשלום" ונותן השירות הוא אורח/עסקי מנוי/מנהל
+      if (shouldUpdateInterestTime && helperIdForInterestUpdate != null && requestId.isNotEmpty) {
+        try {
+          // הוספת נותן השירות ל-`helpers` array של הבקשה כדי שהבקשה תופיע אצלו במסך "פניות שלי"
+          try {
+            final requestRef = _firestore.collection('requests').doc(requestId);
+            final requestDoc = await requestRef.get();
+            
+            if (requestDoc.exists) {
+              final requestData = requestDoc.data()!;
+              final requestType = requestData['type'] as String?;
+              
+              // בדיקה אם הבקשה היא "בתשלום"
+              if (requestType != 'paid') {
+                debugPrint('ℹ️ Request $requestId is not paid, skipping helper addition');
+              } else {
+                // בדיקה אם נותן השירות הוא אורח/עסקי מנוי/מנהל
+                final helperDoc = await _firestore
+                    .collection('users')
+                    .doc(helperIdForInterestUpdate)
+                    .get();
+                
+                if (!helperDoc.exists) {
+                  debugPrint('⚠️ Helper $helperIdForInterestUpdate not found in users collection');
+                } else {
+                  final helperData = helperDoc.data()!;
+                  final helperUserType = helperData['userType'] as String?;
+                  final helperIsAdmin = helperData['isAdmin'] as bool? ?? false;
+                  final helperEmail = helperData['email'] as String?;
+                  
+                  // בדיקה אם נותן השירות הוא אורח/עסקי מנוי (לא מנהל)
+                  final isGuest = helperUserType == 'guest';
+                  final isBusinessSubscription = helperUserType == 'business' && 
+                      (helperData['isSubscriptionActive'] as bool? ?? false);
+                  final isAdmin = helperIsAdmin || 
+                      helperEmail == 'admin@gmail.com' || 
+                      helperEmail == 'haitham.ay82@gmail.com';
+                  
+                  // מנהלים לא מתווספים ל-helpers array - הם יכולים לראות את כל הבקשות אבל לא מופיעים ב"פניות שלי"
+                  if (isAdmin) {
+                    debugPrint('ℹ️ Helper $helperIdForInterestUpdate is admin - skipping helper addition (admins can see all requests but do not appear in "My Requests")');
+                  } else if (!isGuest && !isBusinessSubscription) {
+                    debugPrint('ℹ️ Helper $helperIdForInterestUpdate is not guest/business subscription, skipping helper addition');
+                  } else {
+                    final helpers = List<String>.from(requestData['helpers'] ?? []);
+                    
+                    // אם נותן השירות עדיין לא ב-`helpers` array, נוסיף אותו
+                    if (!helpers.contains(helperIdForInterestUpdate)) {
+                      final currentStatus = requestData['status'] as String?;
+                      
+                      // עדכון helpers
+                      final updateData = <String, dynamic>{
+                        'helpers': FieldValue.arrayUnion([helperIdForInterestUpdate]),
+                        'helpersCount': FieldValue.increment(1),
+                      };
+                      
+                      // אם יש עוזרים והסטטוס הוא "פתוח", עדכן ל-"בטיפול"
+                      if (helpers.isEmpty && currentStatus == 'open') {
+                        updateData['status'] = 'inProgress';
+                        debugPrint('✅ Added helper: Updating status from "open" to "inProgress"');
+                      }
+                      
+                      await requestRef.update(updateData);
+                      debugPrint('✅ Added helper $helperIdForInterestUpdate to request $requestId helpers array (first message from creator)');
+                    } else {
+                      debugPrint('ℹ️ Helper $helperIdForInterestUpdate already in request $requestId helpers array');
+                    }
+                  }
+                }
+              }
+            }
+          } catch (e) {
+            debugPrint('⚠️ Failed to add helper to request when sending first message: $e');
+            // נמשיך גם אם יש שגיאה בהוספת helper
+          }
+          
+          // עדכון זמן ההתעניינות ב-user_interests collection
+          await _firestore
+              .collection('user_interests')
+              .doc('${helperIdForInterestUpdate}_$requestId')
+              .set({
+            'userId': helperIdForInterestUpdate,
+            'requestId': requestId,
+            'interestTime': FieldValue.serverTimestamp(),
+          }, SetOptions(merge: true));
+          
+          debugPrint('✅ Updated interest time for helper $helperIdForInterestUpdate in request $requestId (first message from creator)');
+        } catch (e) {
+          debugPrint('⚠️ Failed to update interest time when sending first message: $e');
+          // לא נעצור את התהליך בגלל שגיאה בעדכון זמן ההתעניינות
+        }
+      }
 
       // שליחת התראה לכל המשתתפים האחרים
       try {
@@ -174,10 +385,13 @@ class ChatService {
             .toList());
   }
 
-  /// שליחת הודעה עם התראות
+  /// שליחת הודעה עם התראות (תמיכה בהודעות טקסט, קול ותמונה)
   static Future<bool> sendMessageWithNotification({
     required String chatId,
     required String text,
+    String? type, // 'text', 'voice', or 'image'
+    String? data, // Base64 string or URL
+    int? duration, // Duration in seconds for voice messages
   }) async {
     try {
       final user = _auth.currentUser;
@@ -192,12 +406,127 @@ class ChatService {
       if (!chatDoc.exists) return false;
 
       final chatData = chatDoc.data()!;
+      
+      // בדיקה אם הצ'אט נמחק על ידי המשתמש הנוכחי
+      final deletedBy = List<String>.from(chatData['deletedBy'] ?? []);
+      if (deletedBy.contains(user.uid)) {
+        debugPrint('⚠️ Chat $chatId was deleted by current user ${user.uid}, cannot send message');
+        return false;
+      }
+      
+      // בדיקה אם הצ'אט סגור - אם כן, נפתח אותו מחדש
+      final isClosed = chatData['isClosed'] as bool? ?? false;
+      if (isClosed) {
+        debugPrint('🔄 Chat $chatId is closed, reopening it...');
+        await _firestore.collection('chats').doc(chatId).update({
+          'isClosed': false, // פתיחת הצ'אט מחדש
+          'updatedAt': Timestamp.fromDate(DateTime.now()),
+        });
+        debugPrint('✅ Reopened closed chat $chatId');
+      }
+      
       final participants = List<String>.from(chatData['participants'] ?? []);
       
       // מציאת כל המקבלים (כל המשתתפים חוץ מהשולח)
       final recipients = participants.where((id) => id != user.uid).toList();
 
       if (recipients.isEmpty) return false;
+
+      final requestId = chatData['requestId'] as String? ?? '';
+      
+      // בדיקה אם זו הודעה ראשונה ממבקש השירות (יוצר הבקשה)
+      // זה מבטיח שהבקשה תופיע בתחילת הרשימה במסך "פניות שלי" של נותן השירות
+      bool shouldUpdateInterestTime = false;
+      String? helperIdForInterestUpdate;
+      if (requestId.isNotEmpty) {
+        try {
+          // קבלת פרטי הבקשה כדי לבדוק מי יוצר הבקשה
+          final requestDoc = await _firestore
+              .collection('requests')
+              .doc(requestId)
+              .get();
+          
+          if (requestDoc.exists) {
+            final requestData = requestDoc.data()!;
+            final creatorId = requestData['createdBy'] as String?;
+            
+            // אם השולח הוא יוצר הבקשה (מבקש השירות), נבדוק אם זו הודעה ראשונה ממנו
+            if (creatorId == user.uid && recipients.isNotEmpty) {
+              // בדיקה אם יש הודעות קודמות ממבקש השירות בצ'אט
+              final messagesFromCreatorSnapshot = await _firestore
+                  .collection('chats')
+                  .doc(chatId)
+                  .collection('messages')
+                  .where('from', isEqualTo: user.uid)
+                  .limit(1)
+                  .get();
+              
+              // אם אין הודעות קודמות ממבקש השירות, זו הודעה ראשונה ממנו
+              if (messagesFromCreatorSnapshot.docs.isEmpty) {
+                shouldUpdateInterestTime = true;
+                helperIdForInterestUpdate = recipients.first; // נותן השירות
+                debugPrint('✅ First message from creator $user.uid to helper $helperIdForInterestUpdate in request $requestId');
+              }
+            }
+          }
+        } catch (e) {
+          debugPrint('⚠️ Failed to check if first message from creator: $e');
+          // נמשיך לשלוח את ההודעה גם אם הבדיקה נכשלה
+        }
+      }
+
+      // Parse message type
+      MessageType messageType = MessageType.text;
+      if (type != null) {
+        switch (type) {
+          case 'voice':
+            messageType = MessageType.voice;
+            break;
+          case 'image':
+            messageType = MessageType.image;
+            break;
+          default:
+            messageType = MessageType.text;
+        }
+      }
+
+      // בדיקת הגבלת 50 הודעות - מחיקת הודעה הישנה ביותר אם יש 50 הודעות
+      // נשתמש בשאילתה פשוטה יותר שלא דורשת אינדקס מורכב
+      final allMessagesSnapshot = await _firestore
+          .collection('chats')
+          .doc(chatId)
+          .collection('messages')
+          .where('isSystemMessage', isEqualTo: false)
+          .get();
+      
+      // אם יש 50 הודעות או יותר, מחק את הישנה ביותר
+      if (allMessagesSnapshot.docs.length >= 50) {
+        // מציאת ההודעה הישנה ביותר (ללא orderBy כדי לא לדרוש אינדקס)
+        Message? oldestMessage;
+        DateTime? oldestDate;
+        
+        for (var doc in allMessagesSnapshot.docs) {
+          final data = doc.data();
+          final sentAt = data['sentAt'] as Timestamp?;
+          if (sentAt != null) {
+            final sentAtDate = sentAt.toDate();
+            if (oldestDate == null || sentAtDate.isBefore(oldestDate)) {
+              oldestDate = sentAtDate;
+              oldestMessage = Message.fromFirestore(doc);
+            }
+          }
+        }
+        
+        if (oldestMessage != null) {
+          await _firestore
+              .collection('chats')
+              .doc(chatId)
+              .collection('messages')
+              .doc(oldestMessage.messageId)
+              .delete();
+          debugPrint('✅ Deleted oldest message ${oldestMessage.messageId} to maintain 50 message limit');
+        }
+      }
 
       // שליחת ההודעה
       final message = Message(
@@ -206,6 +535,9 @@ class ChatService {
         text: text,
         sentAt: DateTime.now(),
         isSystemMessage: false,
+        type: messageType,
+        data: data,
+        duration: duration,
       );
       
       await _firestore
@@ -224,6 +556,101 @@ class ChatService {
         'lastMessageFrom': user.uid,
         'unreadFor': recipients, // מי צריך לקרוא - כל המקבלים
       });
+
+      // עדכון זמן ההתעניינות ואם מבקש השירות שולח הודעה ראשונה, הוספת נותן השירות ל-helpers
+      // זה מבטיח שהבקשה תופיע בתחילת הרשימה במסך "פניות שלי" של נותן השירות
+      // רק אם הבקשה היא "בתשלום" ונותן השירות הוא אורח/עסקי מנוי/מנהל
+      if (shouldUpdateInterestTime && helperIdForInterestUpdate != null && requestId.isNotEmpty) {
+        try {
+          // הוספת נותן השירות ל-`helpers` array של הבקשה כדי שהבקשה תופיע אצלו במסך "פניות שלי"
+          try {
+            final requestRef = _firestore.collection('requests').doc(requestId);
+            final requestDoc = await requestRef.get();
+            
+            if (requestDoc.exists) {
+              final requestData = requestDoc.data()!;
+              final requestType = requestData['type'] as String?;
+              
+              // בדיקה אם הבקשה היא "בתשלום"
+              if (requestType != 'paid') {
+                debugPrint('ℹ️ Request $requestId is not paid, skipping helper addition');
+              } else {
+                // בדיקה אם נותן השירות הוא אורח/עסקי מנוי/מנהל
+                final helperDoc = await _firestore
+                    .collection('users')
+                    .doc(helperIdForInterestUpdate)
+                    .get();
+                
+                if (!helperDoc.exists) {
+                  debugPrint('⚠️ Helper $helperIdForInterestUpdate not found in users collection');
+                } else {
+                  final helperData = helperDoc.data()!;
+                  final helperUserType = helperData['userType'] as String?;
+                  final helperIsAdmin = helperData['isAdmin'] as bool? ?? false;
+                  final helperEmail = helperData['email'] as String?;
+                  
+                  // בדיקה אם נותן השירות הוא אורח/עסקי מנוי (לא מנהל)
+                  final isGuest = helperUserType == 'guest';
+                  final isBusinessSubscription = helperUserType == 'business' && 
+                      (helperData['isSubscriptionActive'] as bool? ?? false);
+                  final isAdmin = helperIsAdmin || 
+                      helperEmail == 'admin@gmail.com' || 
+                      helperEmail == 'haitham.ay82@gmail.com';
+                  
+                  // מנהלים לא מתווספים ל-helpers array - הם יכולים לראות את כל הבקשות אבל לא מופיעים ב"פניות שלי"
+                  if (isAdmin) {
+                    debugPrint('ℹ️ Helper $helperIdForInterestUpdate is admin - skipping helper addition (admins can see all requests but do not appear in "My Requests")');
+                  } else if (!isGuest && !isBusinessSubscription) {
+                    debugPrint('ℹ️ Helper $helperIdForInterestUpdate is not guest/business subscription, skipping helper addition');
+                  } else {
+                    final helpers = List<String>.from(requestData['helpers'] ?? []);
+                    
+                    // אם נותן השירות עדיין לא ב-`helpers` array, נוסיף אותו
+                    if (!helpers.contains(helperIdForInterestUpdate)) {
+                      final currentStatus = requestData['status'] as String?;
+                      
+                      // עדכון helpers
+                      final updateData = <String, dynamic>{
+                        'helpers': FieldValue.arrayUnion([helperIdForInterestUpdate]),
+                        'helpersCount': FieldValue.increment(1),
+                      };
+                      
+                      // אם יש עוזרים והסטטוס הוא "פתוח", עדכן ל-"בטיפול"
+                      if (helpers.isEmpty && currentStatus == 'open') {
+                        updateData['status'] = 'inProgress';
+                        debugPrint('✅ Added helper: Updating status from "open" to "inProgress"');
+                      }
+                      
+                      await requestRef.update(updateData);
+                      debugPrint('✅ Added helper $helperIdForInterestUpdate to request $requestId helpers array (first message from creator)');
+                    } else {
+                      debugPrint('ℹ️ Helper $helperIdForInterestUpdate already in request $requestId helpers array');
+                    }
+                  }
+                }
+              }
+            }
+          } catch (e) {
+            debugPrint('⚠️ Failed to add helper to request when sending first message: $e');
+            // נמשיך גם אם יש שגיאה בהוספת helper
+          }
+          
+          // עדכון זמן ההתעניינות ב-user_interests collection
+          await _firestore
+              .collection('user_interests')
+              .doc('${helperIdForInterestUpdate}_$requestId')
+              .set({
+            'userId': helperIdForInterestUpdate,
+            'requestId': requestId,
+            'interestTime': FieldValue.serverTimestamp(),
+          }, SetOptions(merge: true));
+          
+          debugPrint('✅ Updated interest time for helper $helperIdForInterestUpdate in request $requestId (first message from creator)');
+        } catch (e) {
+          debugPrint('⚠️ Failed to update interest time when sending first message: $e');
+          // לא נעצור את התהליך בגלל שגיאה בעדכון זמן ההתעניינות
+        }
+      }
 
       // שליחת התראה לכל המקבלים (לא לשולח)
       for (final recipientId in recipients) {
@@ -314,20 +741,26 @@ class ChatService {
         }
         
         // בדיקה אם המצב עדכני (פחות מ-30 שניות)
+        // אם המצב לא עדכני (יותר מ-30 שניות), זה אומר שהאפליקציה כנראה סגורה - נשלח התראה
         if (lastUpdated != null) {
           final timeDiff = DateTime.now().difference(lastUpdated.toDate()).inSeconds;
-          if (timeDiff < 30) { // המצב עדכני
+          if (timeDiff < 30) { // המצב עדכני - המשתמש באפליקציה
             if (isInChat) {
               debugPrint('User $recipientId is in a different chat, will send notification');
             } else {
               debugPrint('User $recipientId is not in any chat, will send notification');
             }
           } else {
-            debugPrint('User state is outdated ($timeDiff seconds), will send notification');
+            // המצב לא עדכני - האפליקציה כנראה סגורה - נשלח התראה
+            debugPrint('User state is outdated ($timeDiff seconds) - app is likely closed, will send notification');
           }
+        } else {
+          // אין זמן עדכון - נשלח התראה
+          debugPrint('No lastUpdated timestamp, will send notification');
         }
       } else {
-        debugPrint('No user state found for $recipientId, will send notification');
+        // אין מצב משתמש - האפליקציה כנראה סגורה - נשלח התראה
+        debugPrint('No user state found for $recipientId - app is likely closed, will send notification');
       }
       
       debugPrint('Recipient $recipientId is not in chat $chatId, notification will be sent');
@@ -346,13 +779,38 @@ class ChatService {
     required String chatId,
   }) async {
     try {
-      // שליחת push notification דרך Cloud Functions
+      // קבלת פרטי הבקשה לכותרת
+      String requestTitle = 'בקשה';
+      try {
+        final chatDoc = await _firestore.collection('chats').doc(chatId).get();
+        if (chatDoc.exists) {
+          final chatData = chatDoc.data()!;
+          final requestId = chatData['requestId'] as String?;
+          if (requestId != null) {
+            final requestDoc = await _firestore.collection('requests').doc(requestId).get();
+            if (requestDoc.exists) {
+              final requestData = requestDoc.data()!;
+              requestTitle = requestData['title'] as String? ?? 'בקשה';
+            }
+          }
+        }
+      } catch (e) {
+        debugPrint('Error getting request title: $e');
+      }
+      
+      // שליחת push notification דרך push_notifications collection (יש לה Cloud Function)
       await _firestore
-          .collection('chat_notifications')
+          .collection('push_notifications')
           .add({
-        'recipientId': recipientId,
-        'senderName': senderName,
-        'chatId': chatId,
+        'userId': recipientId,
+        'title': 'הודעה חדשה בצ\'אט 💬',
+        'body': '$senderName: הודעה חדשה',
+        'payload': 'chat_message',
+        'data': {
+          'chatId': chatId,
+          'senderName': senderName,
+          'requestTitle': requestTitle,
+        },
         'createdAt': FieldValue.serverTimestamp(),
       });
 
@@ -448,7 +906,7 @@ class ChatService {
     }
   }
 
-  /// מחיקת הודעה (סימון כמחוקה)
+  /// מחיקת הודעה (סימון כמחוקה - גם להודעות קוליות)
   static Future<bool> deleteMessage({
     required String chatId,
     required String messageId,
@@ -470,7 +928,7 @@ class ChatService {
       final messageData = messageDoc.data()!;
       if (messageData['from'] != user.uid) return false;
 
-      // סימון ההודעה כמחוקה
+      // כל ההודעות (טקסט וקוליות) - סימון כמחוקה (soft delete)
       await _firestore
           .collection('chats')
           .doc(chatId)
@@ -482,7 +940,9 @@ class ChatService {
         'deletedAt': FieldValue.serverTimestamp(),
       });
 
-      debugPrint('Message $messageId deleted successfully');
+      final messageType = messageData['type'] as String?;
+      debugPrint('Message $messageId (type: $messageType) marked as deleted');
+
       return true;
     } catch (e) {
       debugPrint('Error deleting message: $e');
