@@ -42,7 +42,9 @@ class ShareService {
   static Future<void> shareGeneral(Request request) async {
     try {
       final message = _buildShareMessage(request);
-      await Share.share(message, subject: 'בקשה מעניינת - NearMe');
+      // TODO: Replace with SharePlus.instance.share() when ShareParams API is stable
+      // ignore: deprecated_member_use
+      await Share.share(message);
     } catch (e) {
       debugPrint('Error sharing: $e');
     }
@@ -53,20 +55,54 @@ class ShareService {
     final appUrl = 'https://nearme-970f3.web.app';
     final deepLink = '$appUrl/request/${request.requestId}';
     
+    // בניית פרטי הבקשה
+    final categoryName = _getCategoryName(request.category);
+    final urgencyText = _getUrgencyText(request);
+    final typeText = request.type == RequestType.paid ? '💰 בתשלום' : '🆓 חינם';
+    final deadlineText = request.deadline != null 
+        ? '⏰ תאריך יעד: ${_formatDate(request.deadline!)}'
+        : '';
+    
     return '''
-🎯 בקשה מעניינת ב-NearMe!
+🎯 בקשה מעניינת ב-"שכונתי"!
 
 📝 ${request.title}
 📍 ${request.location?.name ?? 'מיקום לא צוין'}
-📅 ${_formatDate(request.createdAt)}
+🏷️ קטגוריה: $categoryName
+$typeText $urgencyText
+📅 פורסם: ${_formatDate(request.createdAt)}
+$deadlineText
 
+📄 תיאור:
 ${request.description}
 
-🔗 הורד את האפליקציה: $appUrl
-📱 או לחץ כאן: $deepLink
+💡 רוצה לעזור? הורד את האפליקציה "שכונתי" וצור קשר ישיר!
 
-#NearMe #בקשות #עזרה
+📱 הורד עכשיו:
+$appUrl
+
+🔗 או לחץ כאן לפתיחת הבקשה:
+$deepLink
+
+🤝 בואו נבנה קהילה חזקה יותר יחד!
+#שכונתי #עזרה_הדדית #בקשות #קהילה #ישראל
 ''';
+  }
+  
+  /// קבלת שם הקטגוריה בעברית
+  static String _getCategoryName(RequestCategory category) {
+    return category.categoryDisplayName;
+  }
+  
+  /// קבלת טקסט דחיפות
+  static String _getUrgencyText(Request request) {
+    if (request.isUrgent || request.urgencyLevel == UrgencyLevel.emergency) {
+      return '🚨 דחוף מאוד!';
+    } else if (request.urgencyLevel == UrgencyLevel.urgent24h) {
+      return '⏰ דחוף - תוך 24 שעות';
+    } else {
+      return '';
+    }
   }
 
   /// עיצוב תאריך
