@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import '../l10n/app_localizations.dart';
 
 class RememberMeDialog extends StatefulWidget {
   final String loginMethod;
@@ -6,11 +7,11 @@ class RememberMeDialog extends StatefulWidget {
   final VoidCallback onDontRemember;
 
   const RememberMeDialog({
-    Key? key,
+    super.key,
     required this.loginMethod,
     required this.onRemember,
     required this.onDontRemember,
-  }) : super(key: key);
+  });
 
   @override
   State<RememberMeDialog> createState() => _RememberMeDialogState();
@@ -19,25 +20,27 @@ class RememberMeDialog extends StatefulWidget {
 class _RememberMeDialogState extends State<RememberMeDialog> {
   bool _rememberMe = false;
 
-  String get _methodDisplayName {
+  String _methodDisplayName(BuildContext context) {
+    final l10n = AppLocalizations.of(context);
     switch (widget.loginMethod) {
       case 'email':
-        return 'אימייל וסיסמה';
+        return l10n.emailAndPassword;
       case 'google':
-        return 'גוגל';
+        return 'גוגל'; // Google - שם מותג, לא מתורגם
       case 'facebook':
-        return 'פייסבוק';
+        return 'פייסבוק'; // Facebook - שם מותג, לא מתורגם
       case 'instagram':
-        return 'אינסטגרם';
+        return 'אינסטגרם'; // Instagram - שם מותג, לא מתורגם
       case 'tiktok':
-        return 'טיקטוק';
+        return 'טיקטוק'; // TikTok - שם מותג, לא מתורגם
       default:
-        return 'החשבון שלך';
+        return l10n.yourAccount; // נצטרך להוסיף תרגום זה
     }
   }
 
   @override
   Widget build(BuildContext context) {
+    final l10n = AppLocalizations.of(context);
     return AlertDialog(
       shape: RoundedRectangleBorder(
         borderRadius: BorderRadius.circular(16),
@@ -46,15 +49,20 @@ class _RememberMeDialogState extends State<RememberMeDialog> {
         children: [
           Icon(
             Icons.fingerprint,
-            color: Theme.of(context).primaryColor,
+            color: Theme.of(context).brightness == Brightness.dark 
+                ? Colors.white 
+                : Theme.of(context).primaryColor,
             size: 28,
           ),
           const SizedBox(width: 12),
-          const Text(
-            'זכור אותי',
+          Text(
+            l10n.rememberMe,
             style: TextStyle(
               fontSize: 20,
               fontWeight: FontWeight.bold,
+              color: Theme.of(context).brightness == Brightness.dark 
+                  ? Colors.white 
+                  : null,
             ),
           ),
         ],
@@ -64,18 +72,22 @@ class _RememberMeDialogState extends State<RememberMeDialog> {
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
           Text(
-            'האם תרצה לשמור את פרטי הכניסה שלך?',
+            l10n.saveCredentialsQuestion,
             style: TextStyle(
               fontSize: 16,
-              color: Colors.grey[700],
+              color: Theme.of(context).brightness == Brightness.dark 
+                  ? Colors.white 
+                  : Colors.grey[700],
             ),
           ),
           const SizedBox(height: 8),
           Text(
-            'שיטת כניסה: $_methodDisplayName',
+            l10n.loginMethod(_methodDisplayName(context)),
             style: TextStyle(
               fontSize: 14,
-              color: Colors.grey[600],
+              color: Theme.of(context).brightness == Brightness.dark 
+                  ? Colors.white 
+                  : Colors.grey[600],
               fontWeight: FontWeight.w500,
             ),
           ),
@@ -83,24 +95,34 @@ class _RememberMeDialogState extends State<RememberMeDialog> {
           Container(
             padding: const EdgeInsets.all(12),
             decoration: BoxDecoration(
-              color: Colors.blue[50],
+              color: Theme.of(context).brightness == Brightness.dark 
+                  ? Colors.blue.shade900.withValues(alpha: 0.3) 
+                  : Colors.blue[50],
               borderRadius: BorderRadius.circular(8),
-              border: Border.all(color: Colors.blue[200]!),
+              border: Border.all(
+                color: Theme.of(context).brightness == Brightness.dark 
+                    ? Colors.blue[700]! 
+                    : Colors.blue[200]!,
+              ),
             ),
             child: Row(
               children: [
                 Icon(
                   Icons.info_outline,
-                  color: Colors.blue[700],
+                  color: Theme.of(context).brightness == Brightness.dark 
+                      ? Colors.white 
+                      : Colors.blue[700],
                   size: 20,
                 ),
                 const SizedBox(width: 8),
                 Expanded(
                   child: Text(
-                    'אם תבחר כן, תוכל להיכנס אוטומטית בפעם הבאה',
+                    l10n.saveCredentialsInfo,
                     style: TextStyle(
                       fontSize: 12,
-                      color: Colors.blue[700],
+                      color: Theme.of(context).brightness == Brightness.dark 
+                          ? Colors.white 
+                          : Colors.blue[700],
                     ),
                   ),
                 ),
@@ -115,13 +137,23 @@ class _RememberMeDialogState extends State<RememberMeDialog> {
                 _rememberMe = value ?? false;
               });
             },
-            title: const Text(
-              'שמור את פרטי הכניסה שלי',
-              style: TextStyle(fontSize: 16),
+            title: Text(
+              l10n.saveCredentialsText,
+              style: TextStyle(
+                fontSize: 16,
+                color: Theme.of(context).brightness == Brightness.dark 
+                    ? Colors.white 
+                    : null,
+              ),
             ),
-            subtitle: const Text(
-              'אני רוצה להיכנס אוטומטית בפעם הבאה',
-              style: TextStyle(fontSize: 12),
+            subtitle: Text(
+              l10n.autoLoginText,
+              style: TextStyle(
+                fontSize: 12,
+                color: Theme.of(context).brightness == Brightness.dark 
+                    ? Colors.white70 
+                    : null,
+              ),
             ),
             controlAffinity: ListTileControlAffinity.leading,
             contentPadding: EdgeInsets.zero,
@@ -130,23 +162,27 @@ class _RememberMeDialogState extends State<RememberMeDialog> {
         ],
       ),
       actions: [
+        // כפתור "לא תודה" - תמיד מוחק את כל הפרטים השמורים, גם אם הצ'קבוקס מסומן
         TextButton(
           onPressed: () {
             Navigator.of(context).pop();
-            widget.onDontRemember();
+            widget.onDontRemember(); // זה תמיד ימחק את כל הפרטים השמורים
           },
           child: Text(
-            'לא, תודה',
+            l10n.noThanks,
             style: TextStyle(
-              color: Colors.grey[600],
+              color: Theme.of(context).brightness == Brightness.dark 
+                  ? Colors.white 
+                  : Colors.grey[600],
               fontSize: 16,
             ),
           ),
         ),
+        // כפתור "שמור" - פעיל רק אם הצ'קבוקס מסומן, ושומר את הפרטים
         ElevatedButton(
           onPressed: _rememberMe ? () {
             Navigator.of(context).pop();
-            widget.onRemember();
+            widget.onRemember(); // זה ישמור את הפרטים רק אם הצ'קבוקס מסומן
           } : null,
           style: ElevatedButton.styleFrom(
             backgroundColor: _rememberMe 
@@ -158,11 +194,12 @@ class _RememberMeDialogState extends State<RememberMeDialog> {
             ),
             padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 12),
           ),
-          child: const Text(
-            'שמור',
+          child: Text(
+            l10n.save,
             style: TextStyle(
               fontSize: 16,
               fontWeight: FontWeight.bold,
+              color: Colors.white,
             ),
           ),
         ),
